% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AccessiblePeaks}
\alias{AccessiblePeaks}
\title{Accessible peaks}
\usage{
AccessiblePeaks(
  object,
  assay = NULL,
  idents = NULL,
  cells = NULL,
  min.cells = 10
)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Name of assay to use}

\item{idents}{A set of identity classes to find accessible peaks for}

\item{cells}{A vector of cells to find accessible peaks for}

\item{min.cells}{Minimum number of cells with the peak accessible (>0 counts)
for the peak to be called accessible}
}
\value{
Returns a vector of peak names
}
\description{
Find accessible peaks in a set of cells
}
\concept{utilities}
