% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{AddFragments}
\alias{AddFragments}
\title{Add a single Fragment object to a ChromatinAssay}
\usage{
AddFragments(object, fragments)
}
\arguments{
\item{object}{A \code{\link{ChromatinAssay}} object}

\item{fragments}{A \code{\link{Fragment}} object}
}
\description{
Add a single Fragment object to a ChromatinAssay
}
