% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{AddMotifMatrix}
\alias{AddMotifMatrix}
\alias{AddMotifMatrix.default}
\alias{AddMotifMatrix.Assay}
\alias{AddMotifMatrix.Seurat}
\title{AddMotifMatrix}
\usage{
AddMotifMatrix(object, ...)

\method{AddMotifMatrix}{default}(object, motifs, assay = NULL,
  verbose = TRUE)

\method{AddMotifMatrix}{Assay}(object, motifs, assay = NULL,
  verbose = TRUE)

\method{AddMotifMatrix}{Seurat}(object, motifs, assay = NULL,
  verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{motifs}{A motif x region matrix. Regions in the motif matrix (colnames)
must match the regions in the assay region x cell matrix (rownames ).}

\item{assay}{Name of the assay to use}

\item{verbose}{Display messages}
}
\value{
Returns a Seurat object
}
\description{
Add a motif x region matrix to a Seurat assay
}
