% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{as.ChromatinAssay5}
\alias{as.ChromatinAssay5}
\alias{as.ChromatinAssay5.Assay5}
\title{Convert objects to a \link[=ChromatinAssay5-class]{ChromatinAssay5} object}
\usage{
as.ChromatinAssay5(x, ...)

\method{as.ChromatinAssay5}{Assay5}(
  x,
  annotation = NULL,
  fragments = NULL,
  bias = NULL,
  motifs = NULL,
  links = NULL,
  region.aggregation = NULL,
  ...
)
}
\arguments{
\item{x}{An object to convert to class \link[=ChromatinAssay5-class]{ChromatinAssay5}}

\item{...}{Arguments passed to other methods}

\item{annotation}{Genomic annotation. It must have the following columns:
\itemize{
\item{tx_id or transcript_id: Transcript ID}
\item{gene_name: Gene name}
\item{gene_id: Gene ID}
\item{gene_biotype: Gene biotype (e.g. "protein_coding", "lincRNA")}
\item{type: Annotation type (e.g. "exon", "gap")}
}}

\item{fragments}{A list of \link[=Fragment2-class]{Fragment2} objects}

\item{bias}{Tn5 integration bias matrix}

\item{motifs}{A \link[=Motif-class]{Motif} object}

\item{links}{Genomic links TODO}

\item{region.aggregation}{A named list of \link[=RegionAggregation-class]{RegionAggregation}}
}
\description{
Convert objects to a \link[=ChromatinAssay5-class]{ChromatinAssay5} object
}
\concept{assay}
