% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{as.GRangesAssay.ChromatinAssay}
\alias{as.GRangesAssay.ChromatinAssay}
\title{Update ChromatinAssay object to GRangesAssay}
\usage{
\method{as.GRangesAssay}{ChromatinAssay}(x, ...)
}
\description{
Update ChromatinAssay object to GRangesAssay
}
\concept{assay}
