% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AskForChoices}
\alias{AskForChoices}
\title{Ask for a choice. Use to select one genome from multi-species data}
\usage{
AskForChoices(choices)
}
\arguments{
\item{choices}{: multi genomes selection}
}
\value{
selected genome
}
\description{
Ask for a choice. Use to select one genome from multi-species data
}
