% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atac_small_old}
\alias{atac_small_old}
\title{A small example scATAC-seq dataset (old)}
\format{
A Seurat object with the following assays
\describe{
\item{peaks}{A peak x cell dataset}
\item{bins}{A 5 kb genome bin x cell dataset}
\item{RNA}{A gene x cell dataset}
}
}
\source{
\url{https://www.10xgenomics.com/datasets/10k-human-pbmcs-atac-v2-chromium-controller-2-standard}
}
\usage{
atac_small_old
}
\description{
A subsetted version of 10x Genomics 10k human (hg19) PBMC scATAC-seq dataset.
This object was included in Signac v1 as the small test data object.
}
\concept{data}
\keyword{datasets}
