% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AverageCounts}
\alias{AverageCounts}
\title{Average Counts}
\usage{
AverageCounts(object, assay = NULL, group.by = NULL, verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Name of assay to use. Default is the active assay}

\item{group.by}{Grouping variable to use. Default is the active identities}

\item{verbose}{Display messages}
}
\value{
Returns a dataframe
}
\description{
Compute the mean counts per group of cells for a given assay
}
\examples{
AverageCounts(atac_small)
}
\concept{utilities}
