% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blacklist_ce10}
\alias{blacklist_ce10}
\title{Genomic blacklist regions for C. elegans ce10s}
\format{A GRanges object}
\source{
https://sites.google.com/site/anshulkundaje/projects/blacklists
}
\usage{
blacklist_ce10
}
\description{
Genomic blacklist regions for C. elegans ce10s
}
\keyword{datasets}
