% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blacklist_dm3}
\alias{blacklist_dm3}
\title{Genomic blacklist regions for Drosophila dm3}
\format{A GRanges object}
\source{
https://sites.google.com/site/anshulkundaje/projects/blacklists
}
\usage{
blacklist_dm3
}
\description{
Genomic blacklist regions for Drosophila dm3
}
\keyword{datasets}
