% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blacklist_hg19}
\alias{blacklist_hg19}
\title{Genomic blacklist regions for Human hg19}
\format{A GRanges object}
\source{
https://www.encodeproject.org/annotations/ENCSR636HFF/
}
\usage{
blacklist_hg19
}
\description{
Genomic blacklist regions for Human hg19
}
\keyword{datasets}
