% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blacklist_hg38}
\alias{blacklist_hg38}
\title{Genomic blacklist regions for Human GRCh38}
\format{A GRanges object}
\source{
https://sites.google.com/site/anshulkundaje/projects/blacklists
}
\usage{
blacklist_hg38
}
\description{
Genomic blacklist regions for Human GRCh38
}
\keyword{datasets}
