% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blacklist_hg38_unified}
\alias{blacklist_hg38_unified}
\title{Unified genomic blacklist regions for Human GRCh38}
\format{
A GRanges object
}
\source{
\url{https://www.encodeproject.org/files/ENCFF356LFX/}

\url{https://doi.org/10.1038/s41598-019-45839-z}
}
\usage{
blacklist_hg38_unified
}
\description{
Manually curated genomic blacklist regions for the hg38 genome by Anshul
Kundaje and Anna Shcherbina. See
\url{https://www.encodeproject.org/files/ENCFF356LFX/} for a description of
how this blacklist was curated.
}
\author{
Anshul Kundaje

Anna Shcherbina
}
\concept{data}
\keyword{datasets}
