% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalculateCoverages}
\alias{CalculateCoverages}
\title{CalculateCoverages}
\usage{
CalculateCoverages(reads, cells.per.group, reads.per.group,
  scale.factor = 10^4, window = 100, verbose = TRUE)
}
\arguments{
\item{reads}{Dataframe containing reads for a region, generated by \code{\link{GetReadsInRegion}}}

\item{cells.per.group}{Number of cells in each group. Must be a named vector. Used for normalization of track heights.}

\item{reads.per.group}{Average number of reads per group.}

\item{scale.factor}{Scaling factor for track height. Default is 10^4.}

\item{window}{Smoothing window to use}

\item{verbose}{Display messages}
}
\description{
Calculate normalized read coverage per base per cell group
}
