% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mito.R
\name{CallVariants}
\alias{CallVariants}
\title{Call mitochondrial variants}
\usage{
CallVariants(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{An Assay object}

\item{verbose}{Display messages}
}
\value{
Returns a dataframe
}
\description{
Identify mitochondrial variants present in single cells.
}
