% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CellAggregation-class}
\alias{CellAggregation-class}
\alias{CellAggregation}
\title{CellAggregation class}
\description{
The CellAggregation class enables storage of counts centered on a group of
genomic regions, aggregated across cells for each region. The main data
object stored is a region-by-position matrix.
}
\section{Slots}{

\describe{
\item{\code{matrix}}{List of region-by-position matrices}

\item{\code{cells}}{List of character vectors containing the cell names contained in
each matrix}

\item{\code{upstream}}{Integer denoting number of bases upstream of the centered
position that are stored in the matrix}

\item{\code{downstream}}{Integer denoting number of bases downstream of the centered
position that are stored in the matrix}
}}

\seealso{
\code{\link{RegionAggregation}}
}
\concept{heatmap}
