% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CellsPerGroup}
\alias{CellsPerGroup}
\title{Cells per group}
\usage{
CellsPerGroup(object, group.by = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{group.by}{A grouping variable. Default is the active identities}
}
\value{
Returns a vector
}
\description{
Count the number of cells in each group
}
\examples{
CellsPerGroup(atac_small)
}
\concept{utilities}
