% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CheckFileExist}
\alias{CheckFileExist}
\title{Check whether a file is exists, then stop with a message}
\usage{
CheckFileExist(path)
}
\arguments{
\item{path}{path to  a file on your system}
}
\description{
Check whether a file is exists, then stop with a message
}
\examples{
path2yourfile <- "./example.txt"
CheckFileExist(path2yourfile)

}
