% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CheckRegion}
\alias{CheckRegion}
\title{CheckRegion}
\usage{
CheckRegion(region)
}
\arguments{
\item{region}{A genomic region, or a vector of regions}
}
\description{
Make sure that the given region is ok. Raises error if not correct, otherwise returns region.
}
