% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ChromatinAssay-class}
\alias{ChromatinAssay-class}
\title{The ChromatinAssay class (old)}
\description{
The ChromatinAssay object is an extended \code{\link[SeuratObject:Assay]{SeuratObject::Assay()}}
for the storage and analysis of single-cell chromatin data.
}
\details{
This is an old object class used in Signac v1
}
\section{Slots}{

\describe{
\item{\code{ranges}}{A \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} object describing the
genomic location of features in the object}

\item{\code{motifs}}{A \link[=Motif-class]{Motif} object}

\item{\code{fragments}}{A list of \link[=Fragment2-class]{Fragment2} objects.}

\item{\code{seqinfo}}{A \code{\link[Seqinfo:Seqinfo]{Seqinfo::Seqinfo()}} object containing basic
information about the genome sequence used.}

\item{\code{annotation}}{A  \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} object containing
genomic annotations. This should be a GRanges object with the following
columns:
\itemize{
\item{tx_id: Transcript ID}
\item{gene_name: Gene name}
\item{gene_id: Gene ID}
\item{gene_biotype: Gene biotype (e.g. "protein_coding", "lincRNA")}
\item{type: Annotation type (e.g. "exon", "gap")}
}}

\item{\code{bias}}{A vector containing Tn5 integration bias information
(frequency of Tn5 integration at different kmers)}

\item{\code{positionEnrichment}}{A named list of matrices containing positional
enrichment scores for Tn5 integration (for example, enrichment at the TSS)}

\item{\code{links}}{A \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} object describing linked
genomic positions, such as co-accessible sites or enhancer-gene regulatory
relationships. This should be a \code{GRanges} object, where the start and
end coordinates are the two linked genomic positions, and must contain a
"score" metadata column.}
}}

\keyword{internal}
