% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ChromatinAssay5-class}
\alias{ChromatinAssay5-class}
\alias{ChromatinAssay5}
\title{The ChromatinAssay5 class}
\description{
The ChromatinAssay5 object is an extended \code{\link[SeuratObject]{Assay5}}
for the storage and analysis of single-cell chromatin data. This class does
not require that the stored features are genomic ranges.
}
\section{Slots}{

\describe{
\item{\code{fragments}}{A list of \code{\link{Fragment}} objects.}

\item{\code{annotation}}{A  \code{\link[GenomicRanges]{GRanges}} object containing
genomic annotations. This should be a GRanges object with the following 
columns:
\itemize{
  \item{tx_id: Transcript ID}
  \item{gene_name: Gene name}
  \item{gene_id: Gene ID}
  \item{gene_biotype: Gene biotype (e.g. "protein_coding", "lincRNA")}
  \item{type: Annotation type (e.g. "exon", "gap")}
}}

\item{\code{bias}}{A vector containing Tn5 integration bias information
(frequency of Tn5 integration at different kmers)}

\item{\code{positionEnrichment}}{A named list of matrices containing positional
enrichment scores for Tn5 integration (for example, enrichment at the TSS)}
}}

\concept{assay}
