% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ClusterMotifs}
\alias{ClusterMotifs}
\title{ClusterMotifs}
\usage{
ClusterMotifs(object, ...)
}
\arguments{
\item{object}{A Seurat object}
}
\value{
Returns a Seurat object
}
\description{
Cluster motifs by co-occurrence in genomic regions.
Computes the Jaccard similarity between motifs based on the proportion of
regions that they co-occur in. This is normalized for the overall frequency of the motifs.
Motifs are then clustered, and cluster identities stored in the Motif object meta data.
}
