% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Colourise}
\alias{Colourise}
\title{This function was imported from testthat package
https://github.com/r-lib/testthat}
\usage{
Colourise(text, fg = "black", bg = NULL)
}
\arguments{
\item{text}{character vector}

\item{fg}{foreground colour, defaults to white}

\item{bg}{background colour, defaults to transparent}
}
\description{
Colourise text for display in the terminal.
}
\details{
If R is not currently running in a system that supports terminal colours
the text will be returned unchanged.

Allowed colours are: black, blue, brown, cyan, dark gray, green, light
blue, light cyan, light gray, light green, light purple, light red,
purple, red, white, yellow
}
\dontrun{
print(Colourise("Red", "red"))
cat(Colourise("Red", "red"), "\\n")
cat(Colourise("White on red", "white", "red"), "\\n")
}
