% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CountsInRegion}
\alias{CountsInRegion}
\title{Counts in region}
\usage{
CountsInRegion(object, assay, regions, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Name of a chromatin assay in the object to use}

\item{regions}{A GRanges object}

\item{...}{Additional arguments passed to \code{\link[IRanges:findOverlaps-methods]{IRanges::findOverlaps()}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\value{
Returns a numeric vector
}
\description{
Count reads per cell overlapping a given set of regions
}
\examples{
\donttest{
CountsInRegion(
  object = atac_small,
  assay = 'bins',
  regions = blacklist_hg19
)
}
}
\concept{utilities}
