% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iranges-methods.R
\name{coverage,GRangesAssay-method}
\alias{coverage,GRangesAssay-method}
\alias{coverage}
\alias{coverage,Seurat-method}
\title{Coverage of a \link[=GRangesAssay-class]{GRangesAssay} object}
\usage{
\S4method{coverage}{GRangesAssay}(
  x,
  shift = 0L,
  width = NULL,
  weight = 1L,
  method = c("auto", "sort", "hash")
)

\S4method{coverage}{Seurat}(
  x,
  shift = 0L,
  width = NULL,
  weight = 1L,
  method = c("auto", "sort", "hash")
)
}
\arguments{
\item{x}{A \link[=GRangesAssay-class]{GRangesAssay} object}

\item{shift}{How much each range should be shifted before coverage is
computed. See \code{\link[IRanges:coverage]{IRanges::coverage()}} in the \link[IRanges:IRanges]{IRanges} package.}

\item{width}{Specifies the length of the returned coverage vectors.
See \code{\link[IRanges:coverage]{IRanges::coverage()}} in the \link[IRanges:IRanges]{IRanges} package.}

\item{weight}{Assigns weight to each range in \code{x}.
See \code{\link[IRanges:coverage]{IRanges::coverage()}} in the \link[IRanges:IRanges]{IRanges} package.}

\item{method}{See \code{\link[IRanges:coverage]{IRanges::coverage()}} in the \link[IRanges:IRanges]{IRanges}
package}
}
\description{
This is the \code{coverage} method for \link[=GRangesAssay-class]{GRangesAssay} objects.
}
\section{Functions}{
\itemize{
\item \code{coverage(GRangesAssay)}: method for GRangesAssay objects

\item \code{coverage(Seurat)}: method for Seurat objects

}}
\seealso{
\itemize{
\item \code{\link[IRanges:coverage]{IRanges::coverage()}}
\item \code{\link[GenomicRanges:coverage]{GenomicRanges::coverage()}}
\item \link[=GRangesAssay-class]{GRangesAssay}
}
}
\concept{coverage}
