% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateGRangesAssay}
\alias{CreateGRangesAssay}
\alias{CreateChromatinAssay5}
\title{ChromatinAssay5 and GRangesAssay object constructors}
\usage{
CreateGRangesAssay(counts, data, ranges = NULL, ...)

CreateChromatinAssay5(
  counts,
  data,
  fragments = NULL,
  annotation = NULL,
  motifs = NULL,
  links = NULL,
  bias = NULL,
  region.aggregation = NULL,
  validate.fragments = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{counts}{A two-dimensional matrix}

\item{data}{Optional pre-normalized count matrix}

\item{ranges}{A set of \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} corresponding to
the rows of the input matrix. If \code{NULL}, genomic ranges will be extracted
from the rownames of the input data matrix, and those rownames must be
formatted as \code{chromosome:start-end}.}

\item{...}{Additional arguments passed to \code{\link[=CreateChromatinAssay5]{CreateChromatinAssay5()}}
and \code{\link[SeuratObject:CreateAssay5Object]{SeuratObject::CreateAssay5Object()}}.}

\item{fragments}{Path to a tabix-indexed fragments file for the data
contained in the input matrix. If multiple fragment files are required,
you can add additional \code{\link[=Fragment]{Fragment()}} object to the assay after it is
created using the \code{\link[=CreateFragmentObject]{CreateFragmentObject()}} and
\code{\link[=Fragments]{Fragments()}} functions. Alternatively, a list of
\code{\link[=Fragment]{Fragment()}} objects can be provided.}

\item{annotation}{A set of \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} containing
annotations for the genome used. It must have the following columns:
\itemize{
\item{tx_id or transcript_id: Transcript ID}
\item{gene_name: Gene name}
\item{gene_id: Gene ID}
\item{gene_biotype: Gene biotype (e.g. "protein_coding", "lincRNA")}
\item{type: Annotation type (e.g. "exon", "gap")}
}}

\item{motifs}{A Motif object}

\item{links}{A named list of \code{\link[InteractionSet:GInteractions]{InteractionSet::GInteractions()}}
objects}

\item{bias}{A Tn5 integration bias matrix}

\item{region.aggregation}{A named list of \link[=RegionAggregation-class]{RegionAggregation}
objects.}

\item{validate.fragments}{Check that cells in the assay are present in the
fragment file.}

\item{verbose}{Display messages}
}
\value{
Returns a Seurat object
}
\description{
The \code{\link[=ChromatinAssay5-class]{ChromatinAssay5-class()}} and \code{\link[=GRangesAssay-class]{GRangesAssay-class()}}
classes extend \code{\link[SeuratObject:Assay5]{SeuratObject::Assay5()}} objects for single-cell
chromatin data, with or without genomic ranges for each feature in the assay.
Use \code{CreateGRangesAssay} to construct a \code{GRangesAssay} object or
\code{CreateChromatinAssay5} to construct a \code{ChromatinAssay5} object.
}
\seealso{
\code{\link[=CreateChromatinAssay5]{CreateChromatinAssay5()}}
}
\concept{assay}
