% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{CreateMotifActivityMatrix}
\alias{CreateMotifActivityMatrix}
\title{CreateMotifActivityMatrix}
\usage{
CreateMotifActivityMatrix(object, assay = NULL, span = 0.5,
  verbose = TRUE, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Which assay to use. This must contain a Motif object.}

\item{span}{Degree of loess smoothing, passed to \code{\link[stats]{loess}}. Default is 0.5.}

\item{verbose}{Display messages}

\item{...}{Additional arguments passed to \code{\link[stats]{loess}}}
}
\value{
Returns a matrix
}
\description{
Create a matrix of the normalized motif accessibility per cell.
Fits a loess model per motif between the total cell accessibility and the sequencing depth for the cell,
and stores the model residual as the normalized motif activity in the cell.
}
