% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateRegionAggregationObject}
\alias{CreateRegionAggregationObject}
\title{Create a RegionAggregation object}
\usage{
CreateRegionAggregationObject(
  mat,
  regions,
  upstream,
  downstream,
  name,
  cells = NULL,
  expected = NULL,
  verbose = TRUE
)
}
\arguments{
\item{mat}{A cell-by-position matrix}

\item{regions}{A \code{\link[GenomicRanges:granges]{GenomicRanges::granges()}} object containing the
regions aggregated across.}

\item{upstream}{Integer denoting number of bases upstream of the
centered position that are stored in the matrix}

\item{downstream}{Integer denoting number of bases downstream of the
centered position that are stored in the matrix}

\item{name}{A name for the set of regions aggregated}

\item{cells}{A vector of cells where each element is the cell barcode
included in the region aggregation matrix. The order of cells in this
vector should correspond to the order of cells in the matrix. If \code{NULL}, cell
names will be extracted from the rownames of the matrix.}

\item{expected}{A vector containing expected number of Tn5 insertions per
position. If \code{NULL}, the expected value will be set uniformly to 1 for each
position.}

\item{verbose}{Display messages.}
}
\value{
Returns a \link[=RegionAggregation-class]{RegionAggregation} object
}
\description{
Create a \code{\link[=RegionAggregation-class]{RegionAggregation-class()}} object to store the insertions over
group of regions
}
\concept{footprinting}
