% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CreateRegionPileupMatrix}
\alias{CreateRegionPileupMatrix}
\title{Create cut site pileup matrix}
\usage{
CreateRegionPileupMatrix(object, regions, upstream = 1000,
  downstream = 1000, assay = NULL, cells = NULL, verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{regions}{A GRanges object}

\item{assay}{Name of the assay to use}

\item{cells}{Which cells to include. If NULL, use all cells}

\item{verbose}{Display messages}
}
\description{
For a set of aligned genomic ranges, find the total number of
integration sites per cell per base.
}
