% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateSignacObject}
\alias{CreateSignacObject}
\title{Create Seurat object with a chromatin assay}
\usage{
CreateSignacObject(counts, assay = "ATAC", project = "SignacProject",
  ranges = NULL, fragments = NULL, annotation = NULL,
  genome = NULL, motifs = NULL, sep = c("-", "-"), ...)
}
\arguments{
\item{counts}{Unnormalized data (raw counts). Should be a sparse matrix.}

\item{assay}{Name of the assay corresponding to the initial input data. Default is "ATAC".}

\item{project}{Sets the project name for the object}

\item{ranges}{A \code{\link[GenomicRanges]{GRanges}} object containing the genomic
position of each row of the counts matrix}

\item{fragments}{A character vector containing path/s to tabix-indexed fragment file/s
for cells in the object}

\item{annotation}{A \code{\link[GenomicRanges]{GRanges}} object containing
genomic annotations for the genome used}

\item{genome}{Name of genome used}

\item{motifs}{A \code{\link{Motif}} object}

\item{sep}{Charaters used to separate the chromosome, start, and end coordinates
in the row names of the data matrix}

\item{...}{Parameters passed to \code{\link[Seurat]{CreateSeuratObject}}}
}
\description{
This is a wrapper for \code{\link[Seurat]{CreateSeuratObject}} to create
an object containing a \code{\link{ChromatinAssay}} rather than a
standard Seurat \code{\link[Seurat]{Assay}}.
}
