% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CutMatrix}
\alias{CutMatrix}
\title{Generate matrix of integration sites}
\usage{
CutMatrix(object, region, assay = NULL, cells = NULL, verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{region}{A GRanges object containing the region of interest}

\item{assay}{A name of assay to use. Must be a \code{\link{ChromatinAssay}}
containing a list of \code{\link{Fragment}} objects.}

\item{cells}{Which cells to include in the matrix. If NULL (default), use all
cells in the object}

\item{verbose}{Display messages}
}
\value{
Returns a sparse matrix
}
\description{
Generates a cell-by-position matrix of Tn5 integration sites.
}
\examples{
fpath <- system.file("extdata", "fragments.tsv.gz", package="Signac")
atac_small <- SetFragments(atac_small, file = fpath)
CutMatrix(
 object = atac_small,
 region = StringToGRanges("chr1-10245-762629")
)
}
