% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DensityScatter}
\alias{DensityScatter}
\title{Scatterplot colored by point density}
\usage{
DensityScatter(object, x, y, log_x = FALSE, log_y = FALSE, quantiles = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{x}{Name of metadata variable to plot on x axis}

\item{y}{Name of metadata variable to plot on y axis}

\item{log_x}{log10 transform x values}

\item{log_y}{log10 transform y values}

\item{quantiles}{Vector of quantiles to display
for x and y data distribution. Must be integer values
between 0 and 100.
TRUE can be passed as a shorthand way to set
\code{c(5, 10, 90, 95)}. If FALSE or NULL, no quantile
information is displayed}
}
\value{
Returns a ggplot object
}
\description{
Create a scatterplot using variables in the object metadata
and color cells by the density of points in the x-y space.
}
\concept{visualization}
