% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{DownsampleFeatures}
\alias{DownsampleFeatures}
\title{Downsample Features}
\usage{
DownsampleFeatures(object, assay = NULL, n = 20000, verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Name of assay to use. Default is the active assay.}

\item{n}{Number of features to retain (default 20000).}

\item{verbose}{Display messages}
}
\value{
Returns a \code{\link[SeuratObject:Seurat]{SeuratObject::Seurat()}} object with
\code{\link[SeuratObject:VariableFeatures]{SeuratObject::VariableFeatures()}} set to the randomly sampled features.
}
\description{
Randomly downsample features and assign to VariableFeatures for the object.
This will select n features at random.
}
\examples{
DownsampleFeatures(atac_small, n = 10)
}
\concept{preprocessing}
