% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{EnrichmentScore}
\alias{EnrichmentScore}
\title{EnrichmentScore}
\usage{
EnrichmentScore(object, assay, verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Name of assay to use}

\item{verbose}{Display messages}
}
\description{
Compute the enrichment of reads centered on a given feature (for example,
transcriptional start sites). This will compute the ratio of reads 1000 bp upstream
and 1000 bp downstream to the centered position.
}
\details{
See \url{https://www.encodeproject.org/data-standards/terms/#enrichment}
}
