% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwig.R
\name{ExportGroupBW}
\alias{ExportGroupBW}
\title{Export bigwig files for groups of cells}
\usage{
ExportGroupBW(
  object,
  assay = NULL,
  group.by = NULL,
  idents = NULL,
  normMethod = "RC",
  tileSize = 100,
  minCells = 5,
  cutoff = NULL,
  chromosome = NULL,
  seqlengths = NULL,
  outdir = getwd(),
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Name of assay to use}

\item{group.by}{The metadata variable used to group the cells}

\item{idents}{Identities to include (defined by group.by parameter)}

\item{normMethod}{Normalization method for the bigwig files. Default 'RC'.
'RC' will divide the number of fragments in a tile by the total number of
fragments in the group. A scaling factor of 10^4 will be applied.
'ncells' will divide the number of fragments in a tile by the number of cells
in the group. 'none' will apply no normalization method.
A vector of values for each cell can also be passed as a metadata column
name. A scaling factor of 10^4 will be applied.}

\item{tileSize}{The size of the tiles in the bigwig file}

\item{minCells}{The minimum number of cells in a group to be exported}

\item{cutoff}{The maximum number of fragments in a given genomic tile}

\item{chromosome}{A vector of chromosomes to export. If \code{NULL}, use all
chromosomes present in \code{seqlengths}.}

\item{seqlengths}{Chromosome lengths used to define the genomic tiles. Can be
a named numeric vector of chromosome lengths, or any object with a
\code{seqlengths} method such as a \code{BSgenome} or
\code{\link[Seqinfo:Seqinfo]{Seqinfo}} object. If \code{NULL}, the chromosome
lengths stored in the object are used; note that these are frequently unset,
in which case \code{seqlengths} must be supplied.}

\item{outdir}{Directory to write output files (split bed files and bigwigs)}

\item{verbose}{Display messages}
}
\value{
Returns a list of paths to the bigwig files that were created
}
\description{
Export coverage tracks as bigwig files, one per group of cells. Fragments
are split by group, the genome is tiled, and the number of fragments in each
tile is counted and (optionally) normalized before writing the bigwig files.
}
\examples{
\dontrun{
ExportGroupBW(object, assay = "peaks")
}
}
\concept{bigwig}
