% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastConvertS4ToSparseMT}
\alias{FastConvertS4ToSparseMT}
\title{FastConvertS4ToSparseMT}
\usage{
FastConvertS4ToSparseMT(mat)
}
\arguments{
\item{mat}{A S4 matrix}
}
\description{
Convert from S4 to SpMat
}
