% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastConvertToDiagonalSparseMat}
\alias{FastConvertToDiagonalSparseMat}
\title{FastConvertToDiagonalSparseMat}
\usage{
FastConvertToDiagonalSparseMat(mat)
}
\arguments{
\item{mat}{A sparse matrix}
}
\description{
Get diag sparse matrix from a sparse matrix
}
