% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastConvertToTripletMat}
\alias{FastConvertToTripletMat}
\title{FastConvertToTripletMat}
\usage{
FastConvertToTripletMat(s)
}
\arguments{
\item{s}{A SEXP type}
}
\description{
Get list triplet of SEXP
}
