% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastCreateFromTriplet}
\alias{FastCreateFromTriplet}
\title{FastCreateFromTriplet}
\usage{
FastCreateFromTriplet(vec1, vec2, vec_val)
}
\arguments{
\item{vec1}{An integer vector}

\item{vec2}{An integer vector}

\item{vec_val}{An double vector}
}
\description{
Create sparse matrix from triplet
}
