% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastCreateSparseMat}
\alias{FastCreateSparseMat}
\title{FastCreateSparseMat}
\usage{
FastCreateSparseMat(nrow, ncol)
}
\arguments{
\item{nrow}{An integer number}

\item{ncol}{An integer number}
}
\description{
Generate a sparse matrix with the elements along the main diagonal set to one and off-diagonal elements set to zero
}
