% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastDiffVector}
\alias{FastDiffVector}
\title{FastDiffVector}
\usage{
FastDiffVector(a, b)
}
\arguments{
\item{a}{An integer vector}

\item{b}{An integer vector}
}
\description{
This function is used to diff 2 vector
}
