% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetColOfSparseMat}
\alias{FastGetColOfSparseMat}
\title{FastGetColOfSparseMat}
\usage{
FastGetColOfSparseMat(mat, j)
}
\arguments{
\item{mat}{A dense matrix}

\item{i}{A integer number}
}
\description{
Get col of sparse matrix
}
