% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetColsOfSparseMat}
\alias{FastGetColsOfSparseMat}
\title{FastGetColsOfSparseMat}
\usage{
FastGetColsOfSparseMat(mat, start, end)
}
\arguments{
\item{mat}{A sparse matrix}

\item{start}{A integer number}

\item{end}{A integer number}
}
\description{
Get cols of sparse matrix
}
