% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetMedianSparseMatByAllCols}
\alias{FastGetMedianSparseMatByAllCols}
\title{FastGetMedianSparseMatByAllCols}
\usage{
FastGetMedianSparseMatByAllCols(mat)
}
\arguments{
\item{mat}{A sparse matrix}
}
\description{
Median all cols in a sparse matrix
}
