% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetMedianSparseMatByAllRows}
\alias{FastGetMedianSparseMatByAllRows}
\title{FastGetMedianSparseMatByAllRows}
\usage{
FastGetMedianSparseMatByAllRows(mat)
}
\arguments{
\item{mat}{A sparse matrix}
}
\description{
Median all rows in a sparse matrix
}
