% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetRowOfSparseMat}
\alias{FastGetRowOfSparseMat}
\title{FastGetRowOfSparseMat}
\usage{
FastGetRowOfSparseMat(mat, i)
}
\arguments{
\item{mat}{A dense matrix}

\item{i}{A integer number}
}
\description{
Get row of sparse matrix
}
