% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetRowsOfMat}
\alias{FastGetRowsOfMat}
\title{FastGetRowsOfMat}
\usage{
FastGetRowsOfMat(mat, vec)
}
\arguments{
\item{mat}{A matrix}

\item{vec}{A row vector}
}
\description{
This function is used to get some rows of matrix
}
