% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetRowsOfSparseMat}
\alias{FastGetRowsOfSparseMat}
\title{FastGetRowsOfSparseMat}
\usage{
FastGetRowsOfSparseMat(mat, start, end)
}
\arguments{
\item{mat}{A sparse matrix}

\item{start}{A integer number}

\item{end}{A integer number}
}
\description{
Get rows of sparse matrix
}
