% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetSubMat}
\alias{FastGetSubMat}
\title{FastGetSubMat}
\usage{
FastGetSubMat(mat, rvec, cvec)
}
\arguments{
\item{mat}{A matrix}

\item{vec}{A col vector}
}
\description{
This function is used to get subview matrix
}
