% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetSubSparseMat}
\alias{FastGetSubSparseMat}
\title{FastGetSubSparseMat}
\usage{
FastGetSubSparseMat(mat, rrvec, ccvec, need_perform_row, need_perform_col)
}
\arguments{
\item{mat}{A sparse matrix}

\item{rrvec}{A row vector}

\item{ccvec}{A col vector}

\item{need_perform_row}{A bool}

\item{need_perform_col}{A bool}
}
\description{
Subview sparse matrix
}
