% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetSubSparseMatByCols}
\alias{FastGetSubSparseMatByCols}
\title{FastGetSubSparseMatByCols}
\usage{
FastGetSubSparseMatByCols(mat, cvec)
}
\arguments{
\item{mat}{A sparse matrix}

\item{cvec}{A col vector}
}
\description{
Get subview of a sparse matrix by cols
}
