% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetSubSparseMatByRows}
\alias{FastGetSubSparseMatByRows}
\title{FastGetSubSparseMatByRows}
\usage{
FastGetSubSparseMatByRows(mat, rvec)
}
\arguments{
\item{mat}{A sparse matrix}

\item{rvec}{A row vector}
}
\description{
Get subview of a sparse matrix by rows
}
