% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetSumSparseMatByAllCols}
\alias{FastGetSumSparseMatByAllCols}
\title{FastGetSumSparseMatByAllCols}
\usage{
FastGetSumSparseMatByAllCols(mat)
}
\arguments{
\item{mat}{A sparse matrix}
}
\description{
Sum all cols in a sparse matrix
}
