% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetSumSparseMatByAllRows}
\alias{FastGetSumSparseMatByAllRows}
\title{FastGetSumSparseMatByAllRows}
\usage{
FastGetSumSparseMatByAllRows(mat)
}
\arguments{
\item{mat}{A sparse matrix}
}
\description{
Sum all rows in a sparse matrix
}
