% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetSumSparseMatByCols}
\alias{FastGetSumSparseMatByCols}
\title{FastGetSumSparseMatByCols}
\usage{
FastGetSumSparseMatByCols(mat, cvec)
}
\arguments{
\item{mat}{A sparse matrix}

\item{cvec}{A col vector}
}
\description{
Sum of cols in a sparse matrix
}
