% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetSumSparseMatByRows}
\alias{FastGetSumSparseMatByRows}
\title{FastGetSumSparseMatByRows}
\usage{
FastGetSumSparseMatByRows(mat, rvec)
}
\arguments{
\item{mat}{A sparse matrix}

\item{rvec}{A col vector}
}
\description{
Sum of rows in a sparse matrix
}
