% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastMatMult}
\alias{FastMatMult}
\title{FastMatMult}
\usage{
FastMatMult(mat1, mat2)
}
\arguments{
\item{mat1}{A matrix}

\item{mat2}{A matrix}
}
\description{
This function is used to add two matrix
}
