% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastSparseMatMult}
\alias{FastSparseMatMult}
\title{FastSparseMatMult}
\usage{
FastSparseMatMult(mat1, mat2)
}
\arguments{
\item{mat1}{A sparse matrix}

\item{mat2}{A sparse matrix}
}
\description{
Multiply two sparse matrix
}
