% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastSparseMatMultDD}
\alias{FastSparseMatMultDD}
\title{FastSparseMatMultDD}
\usage{
FastSparseMatMultDD(mat1, mat2)
}
\arguments{
\item{mat1}{A dense matrix}

\item{mat2}{A dense matrix}
}
\description{
Multiply two dense matrix
}
