% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastSparseMatMultSD}
\alias{FastSparseMatMultSD}
\title{FastSparseMatMultSD}
\usage{
FastSparseMatMultSD(mat1, mat2)
}
\arguments{
\item{mat1}{A sparse matrix}

\item{mat2}{A dense matrix}
}
\description{
Multiply a sparse matrix with a dense matrix
}
