% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastSparseMatMultWithNum}
\alias{FastSparseMatMultWithNum}
\title{FastSparseMatMultWithNum}
\usage{
FastSparseMatMultWithNum(mat, k)
}
\arguments{
\item{mat}{A sparse matrix}

\item{k}{A integer number}
}
\description{
Multiply sparse matrix with number
}
