% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastSparseMatTranspose}
\alias{FastSparseMatTranspose}
\title{FastSparseMatTranspose}
\usage{
FastSparseMatTranspose(mat)
}
\arguments{
\item{mat}{A sparse matrix}
}
\description{
Transpose sparse matrix
}
