% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastStatsOfSparseMat}
\alias{FastStatsOfSparseMat}
\title{FastStatsOfSparseMat}
\usage{
FastStatsOfSparseMat(mat)
}
\arguments{
\item{mat}{An sparse matrix}
}
\description{
Get statistic of sparse matrix
}
