% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{FilterCells}
\alias{FilterCells}
\title{Filter cells from fragment file}
\usage{
FilterCells(fragments, cells, outfile = NULL, verbose = TRUE)
}
\arguments{
\item{fragments}{Path to a fragment file}

\item{cells}{A vector of cells to keep}

\item{outfile}{Name for output file}

\item{verbose}{Display messages}
}
\description{
Remove all fragments that are not from an allowed set of cell barcodes from
the fragment file. This will create a new file on disk that only contains
fragments from cells specified in the \code{cells} argument.
}
\examples{
fpath <- system.file("extdata", "fragments.tsv.gz", package="Signac")
FilterCells(
  fragments = fpath,
  cells = head(colnames(atac_small)),
  outfile = "/dev/null"
)
}
\concept{fragments}
