% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{FilterDataBasic}
\alias{FilterDataBasic}
\title{Filter epxression matrix}
\usage{
FilterDataBasic(object, slot = "raw.data", min.cells = NULL,
  min.genes = 200, not.expressed = 0, verbose = TRUE)
}
\arguments{
\item{object}{Signac object}

\item{slot}{Data for filtering. Default is raw.data}

\item{min.cells}{The minimum number of cells for a gene. Can be a numeric variable.
If NULL, it will be the at least 10 or the 2% of the total cells.}

\item{min.genes}{The minimum number of genes for a cell. Default is 200.}

\item{not.expressed}{Genes that have expression value equal or less than
this number is considered not expressed. Default is 0.}

\item{verbose}{Talkative or not}
}
\description{
Filter epxression matrix
}
