% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iranges-methods.R
\name{findOverlaps-methods}
\alias{findOverlaps-methods}
\alias{findOverlaps}
\alias{findOverlaps,Vector,GRangesAssay-method}
\alias{findOverlaps,GRangesAssay,Vector-method}
\alias{findOverlaps,GRangesAssay,GRangesAssay-method}
\alias{findOverlaps,Vector,Seurat-method}
\alias{findOverlaps,Seurat,Vector-method}
\alias{findOverlaps,Seurat,Seurat-method}
\alias{countOverlaps,Vector,GRangesAssay-method}
\alias{countOverlaps}
\alias{countOverlaps,GRangesAssay,Vector-method}
\alias{countOverlaps,GRangesAssay,GRangesAssay-method}
\alias{countOverlaps,Seurat,Vector-method}
\alias{countOverlaps,Vector,Seurat-method}
\alias{countOverlaps,Seurat,Seurat-method}
\title{Find overlapping ranges for GRangesAssay objects}
\usage{
\S4method{findOverlaps}{Vector,GRangesAssay}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  select = c("all", "first", "last", "arbitrary"),
  ignore.strand = FALSE
)

\S4method{findOverlaps}{GRangesAssay,Vector}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  select = c("all", "first", "last", "arbitrary"),
  ignore.strand = FALSE
)

\S4method{findOverlaps}{GRangesAssay,GRangesAssay}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  select = c("all", "first", "last", "arbitrary"),
  ignore.strand = FALSE
)

\S4method{findOverlaps}{Vector,Seurat}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  select = c("all", "first", "last", "arbitrary"),
  ignore.strand = FALSE
)

\S4method{findOverlaps}{Seurat,Vector}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  select = c("all", "first", "last", "arbitrary"),
  ignore.strand = FALSE
)

\S4method{findOverlaps}{Seurat,Seurat}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  select = c("all", "first", "last", "arbitrary"),
  ignore.strand = FALSE
)

\S4method{countOverlaps}{Vector,GRangesAssay}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  ignore.strand = FALSE
)

\S4method{countOverlaps}{GRangesAssay,Vector}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  ignore.strand = FALSE
)

\S4method{countOverlaps}{GRangesAssay,GRangesAssay}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  ignore.strand = FALSE
)

\S4method{countOverlaps}{Seurat,Vector}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  ignore.strand = FALSE
)

\S4method{countOverlaps}{Vector,Seurat}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  ignore.strand = FALSE
)

\S4method{countOverlaps}{Seurat,Seurat}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  ignore.strand = FALSE
)
}
\arguments{
\item{query, subject}{A \link[=GRangesAssay-class]{GRangesAssay} object}

\item{maxgap, minoverlap, type, select, ignore.strand}{See
\verb{?[findOverlaps][GenomicRanges::findOverlaps]} in the \link[GenomicRanges:GenomicRanges]{GenomicRanges} and
\link[IRanges:IRanges]{IRanges} packages.}
}
\value{
See \code{\link[GenomicRanges:findOverlaps]{GenomicRanges::findOverlaps()}}
}
\description{
The \verb{findOverlaps, countOverlaps} methods are available for
\link[=GRangesAssay-class]{GRangesAssay} objects. This allows finding overlaps between
genomic ranges and the ranges stored in the \link[=GRangesAssay-class]{GRangesAssay}
}
\details{
If a \link[=GRangesAssay-class]{GRangesAssay} assay is set as the default assay in a
\code{\link[SeuratObject:Seurat]{SeuratObject::Seurat()}} object, you can also call \code{findOverlaps}
directly on the Seurat object.
}
\section{Functions}{
\itemize{
\item \code{findOverlaps(query = GRangesAssay, subject = Vector)}: method for GRangesAssay, Vector

\item \code{findOverlaps(query = GRangesAssay, subject = GRangesAssay)}: method for GRangesAssay, GRangesAssay

\item \code{findOverlaps(query = Vector, subject = Seurat)}: method for Vector, Seurat

\item \code{findOverlaps(query = Seurat, subject = Vector)}: method for Seurat, Vector

\item \code{findOverlaps(query = Seurat, subject = Seurat)}: method for Seurat, Seurat

\item \code{countOverlaps(query = Vector, subject = GRangesAssay)}: method for Vector, GRangesAssay

\item \code{countOverlaps(query = GRangesAssay, subject = Vector)}: method for GRangesAssay, Vector

\item \code{countOverlaps(query = GRangesAssay, subject = GRangesAssay)}: method for GRangesAssay, GRangesAssay

\item \code{countOverlaps(query = Seurat, subject = Vector)}: method for Seurat, Vector

\item \code{countOverlaps(query = Vector, subject = Seurat)}: method for Vector, Seurat

\item \code{countOverlaps(query = Seurat, subject = Seurat)}: method for Seurat, Seurat

}}
\seealso{
\itemize{
\item \link[IRanges:findOverlaps-methods]{IRanges::findOverlaps-methods}
\item \link[GenomicRanges:findOverlaps-methods]{GenomicRanges::findOverlaps-methods}
\item \link[=GRangesAssay-class]{GRangesAssay}
}
}
\concept{overlaps}
