% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{FindTopFeatures}
\alias{FindTopFeatures}
\alias{FindTopFeatures.default}
\alias{FindTopFeatures.Assay5}
\alias{FindTopFeatures.StdAssay}
\alias{FindTopFeatures.Seurat}
\title{Find most frequently observed features}
\usage{
FindTopFeatures(object, ...)

\method{FindTopFeatures}{default}(object, assay = NULL, min.cutoff = "q5", verbose = TRUE, ...)

\method{FindTopFeatures}{Assay5}(
  object,
  assay = NULL,
  layer = "counts",
  min.cutoff = "q5",
  key = "topfeatures",
  verbose = TRUE,
  ...
)

\method{FindTopFeatures}{StdAssay}(
  object,
  assay = NULL,
  layer = "counts",
  min.cutoff = "q5",
  key = "topfeatures",
  verbose = TRUE,
  ...
)

\method{FindTopFeatures}{Seurat}(
  object,
  assay = NULL,
  layer = "counts",
  min.cutoff = "q5",
  key = "topfeatures",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \link[SeuratObject:Seurat]{SeuratObject::Seurat} object}

\item{...}{Arguments passed to other methods}

\item{assay}{Name of assay to use}

\item{min.cutoff}{Cutoff for feature to be included in the VariableFeatures
for the object. This can be a percentile specified as 'q' followed by the
minimum percentile, for example 'q5' to set the top 95\% most common features
as the VariableFeatures for the object. Alternatively, this can be an integer
specifying the minimum number of counts for the feature
to be included in the set of VariableFeatures. For example, setting to 10
will include features with >10 total counts in the set of VariableFeatures.
If \code{NULL}, include all features in VariableFeatures.}

\item{verbose}{Display messages}

\item{layer}{Name of layer to use}

\item{key}{Key to use when storing the highly variable feature information in
the assay.}
}
\value{
Returns a \code{\link[SeuratObject:Seurat]{SeuratObject::Seurat()}} object
}
\description{
Find top features for a given assay based on total number of counts for the
feature. Can specify a minimum cell count, or a lower percentile
bound to determine the set of variable features. Running this function will
store the total counts and percentile rank for each feature in the feature
metadata for the assay. To only compute the feature metadata, without
changing the variable features for the assay, set \code{min.cutoff=NA}.
}
\examples{
FindTopFeatures(object = atac_small[["peaks"]]["data"])
FindTopFeatures(object = atac_small[["peaks"]])
FindTopFeatures(object = atac_small[["peaks"]])
FindTopFeatures(atac_small)
}
\concept{preprocessing}
