% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/fragments.R
\docType{class}
\name{Fragment2-class}
\alias{Fragment2-class}
\alias{Fragment}
\alias{seqlevels,Fragment2-method}
\alias{seqlevels<-,Fragment2-method}
\alias{seqlevelsStyle,Fragment2-method}
\alias{seqlevelsStyle<-,Fragment2-method}
\alias{renameSeqlevels,Fragment2-method}
\title{The Fragment class}
\usage{
\S4method{seqlevels}{Fragment2}(x)

\S4method{seqlevels}{Fragment2}(x) <- value

\S4method{seqlevelsStyle}{Fragment2}(x)

\S4method{seqlevelsStyle}{Fragment2}(x) <- value

\S4method{renameSeqlevels}{Fragment2}(x, value)
}
\arguments{
\item{x}{A \link[=Fragment2-class]{Fragment2} object}

\item{value}{A named character vector or name of a seqlevelsStyle}
}
\description{
The Fragment class is designed to hold information needed for working with
fragment files.
}
\section{Functions}{
\itemize{
\item \code{seqlevels(Fragment2)}: Get sequence levels

\item \code{seqlevels(Fragment2) <- value}: Set sequence levels

\item \code{seqlevelsStyle(Fragment2)}: Get sequence levels style

\item \code{seqlevelsStyle(Fragment2) <- value}: Set sequence levels style

\item \code{renameSeqlevels(Fragment2)}: Rename sequence levels

}}
\section{Slots}{

\describe{
\item{\code{file.path}}{Path to the fragment file on disk.}

\item{\code{file.index}}{Path to the fragment file index on disk.}

\item{\code{hash}}{A vector of two md5sums: first element is the md5sum of the
fragment file, the second element is the md5sum of the index.}

\item{\code{cells}}{A named vector of cells where each element is the cell barcode
as it appears in the fragment file, and the name of each element is the
corresponding cell barcode as stored in the ChromatinAssay5 object.}

\item{\code{seqlevels}}{A named vector of sequence levels (eg, chromosome name) where
each element is the sequence name as it appears in the fragment file, and
then name of each element is the corresponding sequence name as stored in the
ChromatinAssay5 object.}
}}

\concept{fragments}
