% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{GenomeBinMatrix}
\alias{GenomeBinMatrix}
\title{Genome bin matrix}
\usage{
GenomeBinMatrix(
  fragments,
  genome,
  cells = NULL,
  binsize = 5000,
  process_n = 2000,
  bpcells = FALSE,
  bpcells.dir = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fragments}{Path to tabix-indexed fragments file or a list of
\code{\link[=Fragment]{Fragment()}} objects}

\item{genome}{A vector of chromosome sizes for the genome. This is used to
construct the genome bin coordinates. The can be obtained by calling
\code{seqlengths} on a
\code{\link[BSgenome:BSgenome-class]{BSgenome::BSgenome-class()}} object.}

\item{cells}{Vector of cells to include. If NULL, include all cells found
in the fragments file}

\item{binsize}{Size of the genome bins to use}

\item{process_n}{Number of regions to load into memory at a time, per thread.
Processing more regions at once can be faster but uses more memory.}

\item{bpcells}{If \code{TRUE}, return a \code{BPCells::IterableMatrix} backed by an
on-disk BPCells directory instead of an in-memory sparse matrix. See
\code{\link[=FeatureMatrix]{FeatureMatrix()}}.}

\item{bpcells.dir}{Optional path to persist the BPCells directory beyond
the R session. See \code{\link[=FeatureMatrix]{FeatureMatrix()}}.}

\item{verbose}{Display messages.}

\item{...}{Arguments passed to \code{\link[=FeatureMatrix]{FeatureMatrix()}}.}
}
\value{
Returns a sparse matrix, or a \code{BPCells::IterableMatrix} when
\code{bpcells = TRUE}.
}
\description{
Construct a bin x cell matrix from a fragments file.
}
\details{
This function bins the genome and calls \code{\link[=FeatureMatrix]{FeatureMatrix()}} to
construct a bin x cell matrix.
}
\examples{
\donttest{
genome <- 780007
names(genome) <- "chr1"
fpath <- system.file("extdata", "fragments.tsv.gz", package = "Signac")
fragments <- CreateFragmentObject(fpath, cells = colnames(atac_small))
GenomeBinMatrix(
  fragments = fragments,
  genome = genome,
  binsize = 1000
)
}
}
\concept{quantification}
