% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venice.R
\name{GetCellsIndices}
\alias{GetCellsIndices}
\title{This function helps to extract the cell indices given two indents
We adapted some utilities functions in Seurat v3 package (Butler et al., Nature Biotechnology 2018) to handle the annotation of an
object
This function are edited from the FindMarkers function of Seurat package v3
https://github.com/satijalab/seurat/blob/master/R/differential_expression.R}
\usage{
GetCellsIndices(data.use, object, ident.1, ident.2)
}
\arguments{
\item{data.use}{The sparse matrix object that contains the counts data}

\item{object}{The Seurat v3 object}

\item{ident.1}{Identity class to define markers for; pass an object of class
\code{phylo} or 'clustertree' to find markers for a node in a cluster tree;
passing 'clustertree' requires \code{\link{BuildClusterTree}} to have been run}

\item{ident.2}{A second identity class for comparison; if \code{NULL},
use all other cells for comparison; if an object of class \code{phylo} or
'clustertree' is passed to \code{ident.1}, must pass a node to find markers for}
}
\description{
This function helps to extract the cell indices given two indents
We adapted some utilities functions in Seurat v3 package (Butler et al., Nature Biotechnology 2018) to handle the annotation of an
object
This function are edited from the FindMarkers function of Seurat package v3
https://github.com/satijalab/seurat/blob/master/R/differential_expression.R
}
