% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprinting.R
\name{GetFootprintData}
\alias{GetFootprintData}
\title{Extract footprint data for a set of transcription factors}
\usage{
GetFootprintData(
  object,
  features,
  assay = NULL,
  group.by = NULL,
  idents = NULL
)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{A vector of features to plot}

\item{assay}{Name of assay to use}

\item{group.by}{A grouping variable}

\item{idents}{Set of identities to include in the plot}
}
\description{
Extract footprint data for a set of transcription factors
}
