% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetFragments}
\alias{GetFragments}
\title{GetFragments}
\usage{
GetFragments(object, assay = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Name of the assay use to store the fragments file path}
}
\value{
Returns the path to a fragments file stored in the Assay if present
}
\description{
Retrieve path to fragments file from assay object, and checks that the file
exists and is indexed before returning the file path.
}
\examples{
fpath <- system.file("extdata", "fragments.tsv.gz", package="Signac")
atac_small <- SetFragments(object = atac_small, file = fpath)
GetFragments(object = atac_small)
}
