% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetGroups}
\alias{GetGroups}
\title{Get vector of cell names and associated identity}
\usage{
GetGroups(object, group.by, idents)
}
\arguments{
\item{object}{A Seurat object}

\item{group.by}{Identity class to group cells by}

\item{idents}{which identities to include}
}
\value{
Returns a named vector
}
\description{
Get vector of cell names and associated identity
}
