% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/links.R
\name{GetLinkedGenes}
\alias{GetLinkedGenes}
\alias{GetLinkedGenes.Seurat}
\alias{GetLinkedGenes.Assay5}
\alias{GetLinkedGenes.ChromatinAssay5}
\title{Get genes linked to peaks}
\usage{
GetLinkedGenes(object, ...)

\method{GetLinkedGenes}{Seurat}(object, features, key, assay = NULL, min.abs.score = 0, ...)

\method{GetLinkedGenes}{Assay5}(object, ...)

\method{GetLinkedGenes}{ChromatinAssay5}(object, features, key, min.abs.score = 0, ...)
}
\arguments{
\item{object}{A \link[SeuratObject:Seurat]{SeuratObject::Seurat} object}

\item{...}{Arguments passed to other methods}

\item{features}{A list of peaks to find linked genes for}

\item{key}{Key for link information to extract from the assay}

\item{assay}{Name of assay to use. If NULL, use the default assay}

\item{min.abs.score}{Minimum absolute value of the link score for a link to
be returned}
}
\value{
Returns a character vector of peaks
}
\description{
Retrieve genomic link information for a given set of regions.
}
\seealso{
GetLinkedPeaks
}
\concept{links}
