% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/links.R
\name{GetLinkedPeaks}
\alias{GetLinkedPeaks}
\alias{GetLinkedPeaks.Seurat}
\alias{GetLinkedPeaks.Assay5}
\alias{GetLinkedPeaks.ChromatinAssay5}
\title{Get peaks linked to genes}
\usage{
GetLinkedPeaks(object, ...)

\method{GetLinkedPeaks}{Seurat}(object, features, key, assay = NULL, min.abs.score = 0, ...)

\method{GetLinkedPeaks}{Assay5}(object, ...)

\method{GetLinkedPeaks}{ChromatinAssay5}(object, features, key, min.abs.score = 0, ...)
}
\arguments{
\item{object}{A \link[SeuratObject:Seurat]{SeuratObject::Seurat} object}

\item{...}{Arguments passed to other methods}

\item{features}{A list of genes to find linked peaks for}

\item{key}{Key for link information to extract from the assay}

\item{assay}{Name of assay to use. If NULL, use the default assay}

\item{min.abs.score}{Minimum absolute value of the link score for a link to
be returned}
}
\value{
Returns a character vector of genes
}
\description{
Retrieve genomic link information for a given set of genes.
}
\seealso{
GetLinkedGenes
}
\concept{links}
