% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetListAttributes}
\alias{GetListAttributes}
\title{GetListAttributes}
\usage{
GetListAttributes(filePath, groupName, datasetName)
}
\arguments{
\item{filePath}{A HDF5 path}

\item{groupName}{A string (HDF5 dataset)}

\item{datasetName}{A dataset name}
}
\description{
Get list attribute of a dataset
}
