% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetListObjectNames}
\alias{GetListObjectNames}
\title{GetListObjectNames}
\usage{
GetListObjectNames(filePath, groupName)
}
\arguments{
\item{filePath}{A HDF5 path}

\item{groupName}{A string (HDF5 dataset)}
}
\description{
Get list object of a group
}
