% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetListRootObjectNames}
\alias{GetListRootObjectNames}
\title{GetListRootObjectNames}
\usage{
GetListRootObjectNames(filePath)
}
\arguments{
\item{filePath}{An HDF5 path}
}
\description{
Get list groups
}
