% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetReadsInRegion}
\alias{GetReadsInRegion}
\title{GetReadsInRegion}
\usage{
GetReadsInRegion(
  cellmap,
  region,
  tabix.file,
  cells = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{cellmap}{A mapping of cell names in the fragment file to cell names in
the Seurat object. Should be a named vector where each element is a cell name
that appears in the fragment file and the name of each element is the
name of the cell in the Seurat object.}

\item{region}{A genomic region, specified as a string in the format
'chr:start-end'. Can be a vector of regions.}

\item{tabix.file}{A TabixFile object.}

\item{cells}{Cells to include. Default is all cells present in the object.}

\item{verbose}{Display messages}

\item{...}{Additional arguments passed to \code{\link{StringToGRanges}}}
}
\value{
Returns a data frame
}
\description{
Extract reads for each cell within a given genomic region or set of regions
}
\examples{
fpath <- system.file("extdata", "fragments.tsv.gz", package="Signac")
atac_small <- SetFragments(object = atac_small, file = fpath)
region <- StringToGRanges(regions = "chr1-10245-762629")
GetReadsInRegion(object = atac_small, region = region)
}
