% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges-methods.R
\name{granges-methods}
\alias{granges-methods}
\alias{granges}
\alias{granges,GRangesAssay-method}
\alias{granges,Seurat-method}
\title{Access genomic ranges for GRangesAssay objects}
\usage{
\S4method{granges}{GRangesAssay}(x, use.names = TRUE, use.mcols = FALSE, ...)

\S4method{granges}{Seurat}(x, use.names = TRUE, use.mcols = FALSE, ...)
}
\arguments{
\item{x}{A \link[=GRangesAssay-class]{GRangesAssay} object}

\item{use.names}{Whether the names on the genomic ranges should be
propagated to the returned object.}

\item{use.mcols}{Not supported for \link[=GRangesAssay-class]{GRangesAssay} objects}

\item{...}{Additional arguments}
}
\value{
Returns a \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} object
}
\description{
Methods for accessing \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} object
information stored in a \link[=GRangesAssay-class]{GRangesAssay} object.
}
\section{Functions}{
\itemize{
\item \code{granges(Seurat)}: method for Seurat objects

}}
\examples{
granges(atac_small)
}
\seealso{
\itemize{
\item \code{\link[GenomicRanges:granges]{GenomicRanges::granges()}}
\item \link[GenomicRanges:granges]{GenomicRanges::GRanges()}
}
}
\concept{granges}
