% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GRangesToString}
\alias{GRangesToString}
\title{GRangesToString}
\usage{
GRangesToString(grange, sep = c("-", "-"))
}
\arguments{
\item{grange}{A GRanges object}

\item{sep}{Separator for the genomic coordinate string}
}
\description{
Convert GRanges object to a vector of strings
}
