% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{GWASTrack}
\alias{GWASTrack}
\title{Create GWAS locus zoom track}
\usage{
GWASTrack(
  gwas,
  region,
  ld.file = NULL,
  ld.lead.snp = NULL,
  credset.file = NULL,
  credset.threshold = 0.01,
  p.threshold = 5e-08,
  ymax = NULL,
  point.size = 1,
  point.color = "steelblue",
  show.axis = TRUE
)
}
\arguments{
\item{gwas}{Path to GWAS summary statistics file, or a dataframe
containing the gwas data in the GWAS-SSF format}

\item{region}{Genomic region (\link[GenomicRanges:GRanges]{GenomicRanges::GRanges} or a string that can
be converted to \code{GRanges} like "chr10:112900000-113100000")}

\item{ld.file}{Path to LD file. Optional.}

\item{ld.lead.snp}{Lead SNP for LD (required if ld.file provided)}

\item{credset.file}{Path to credible set file. Optional.}

\item{credset.threshold}{PIP threshold for credible sets (default: 0.01)}

\item{p.threshold}{Genome-wide significance threshold (default: 5e-8)}

\item{ymax}{Maximum y-axis value (default: auto)}

\item{point.size}{Point size (default: 1)}

\item{point.color}{Point color when no LD (default: "steelblue")}

\item{show.axis}{Show x-axis (default: TRUE)}
}
\value{
ggplot2 object
}
\description{
This function will plot the p-values associated with variants in a given
region of the genome (genome position: x-axis; -log10(p): y-axis).
}
\details{
If an LD file is provided using the \code{ld.file} parameter, or a column
named \code{r2} is present in the input GWAS data, the variants will be
colored according to their LD value.

If a fine mapping file is provided using the \code{credset.file} parameter,
or a column named \code{in_credset} is present in the input GWAS data,
variants in the credible set will be denoted by shape. If LD information is
not provided, credible variants will also be denoted by color.
}
\concept{visualization}
