% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{Harmonize}
\alias{Harmonize}
\title{Integrate multiple datasets using Harmony}
\usage{
Harmonize(object, group.by, assay = NULL, reduction = "lsi",
  verbose = TRUE, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{group.by}{Grouping variable.}

\item{assay}{Name of the assay to use. If NULL, use the default assay.}

\item{reduction}{Name of dimension reduction to harmonize}

\item{verbose}{Display messages}

\item{...}{Additional arguments passed to \code{\link[harmony]{HarmonyMatrix}}}
}
\value{
Returns a Seurat object
}
\description{
Runs Harmony on a set of cell embeddings for the integration of multiple datasets.
Harmony was developed by Ilya Korsunsky et al., if you use Harmony please consider citing:
\url{https://doi.org/10.1101/461954}
}
