% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HarmonyMarker}
\alias{HarmonyMarker}
\title{HarmonyMarker}
\usage{
HarmonyMarker(S4_mtx, mtx, cluster)
}
\arguments{
\item{S4_mtx}{A sparse matrix}

\item{cluster}{A numeric vector}
}
\description{
Find gene marker for a cluster in sparse matrix
}
