% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HarmonyMarkerH5}
\alias{HarmonyMarkerH5}
\title{HarmonyMarkerH5}
\usage{
HarmonyMarkerH5(hdf5Path, cluster)
}
\arguments{
\item{hdf5Path}{A string path}

\item{cluster}{A numeric vector}
}
\description{
Find gene marker for a cluster in H5 file
}
