% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{head.Fragment2}
\alias{head.Fragment2}
\title{Return the first rows of a fragment file}
\usage{
\method{head}{Fragment2}(x, n = 6L, ...)
}
\arguments{
\item{x}{a \code{Fragment2} object}

\item{n}{an integer specifying the number of rows to return from the fragment
file}

\item{...}{additional arguments passed to \code{\link[utils:read.table]{utils::read.table()}}}
}
\value{
The first \code{n} rows of a fragment file as a \code{data.frame}
with the following columns: chrom, start, end, barcode, readCount.
}
\description{
Returns the first \code{n} rows of a fragment file. This allows the content
of a fragment file to be inspected.
}
\concept{fragments}
