% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{header}
\alias{header}
\title{Return the fragment file header}
\usage{
header(x)
}
\arguments{
\item{x}{a \code{Fragment2} object}
}
\value{
Returns a character vector with the fragment file header lines
}
\description{
Returns any comment lines present at the start of the fragment file,
before any fragment entries.
}
\concept{fragments}
