% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mito.R
\name{IdentifyVariants}
\alias{IdentifyVariants}
\title{Identify mitochondrial variants}
\usage{
IdentifyVariants(
  object,
  refallele,
  stabilize_variance = TRUE,
  low_coverage_threshold = 10,
  verbose = TRUE
)
}
\arguments{
\item{object}{An Assay object}

\item{refallele}{A dataframe containing reference alleles for the
mitochondrial genome.}

\item{verbose}{Display messages}

\item{stabilize.variance}{Stabilize variance}

\item{low.coverage.threshold}{Low coverage threshold}
}
\value{
Returns a dataframe
}
\description{
Identify mitochondrial variants present in single cells.
}
\examples{
\dontrun{
data.dir <- "path/to/data/directory"
mgatk <- ReadMGATK(dir = data.dir)
variant.df <- CallVariants(
  object = mgatk$counts,
  refallele = mgatk$refallele
)
}
}
