% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iranges-methods.R
\name{inter-range-methods}
\alias{inter-range-methods}
\alias{range}
\alias{range,GRangesAssay-method}
\alias{range,Seurat-method}
\alias{reduce,GRangesAssay-method}
\alias{reduce}
\alias{reduce,Seurat-method}
\alias{gaps,GRangesAssay-method}
\alias{gaps}
\alias{gaps,Seurat-method}
\alias{disjoin,GRangesAssay-method}
\alias{disjoin}
\alias{disjoin,Seurat-method}
\alias{isDisjoint,GRangesAssay-method}
\alias{isDisjoint}
\alias{isDisjoint,Seurat-method}
\alias{disjointBins,GRangesAssay-method}
\alias{disjointBins}
\alias{disjointBins,Seurat-method}
\title{Inter-range transformations for \link[=GRangesAssay-class]{GRangesAssay} objects}
\usage{
\S4method{range}{GRangesAssay}(x, ..., with.revmap = FALSE, na.rm = FALSE)

\S4method{range}{Seurat}(x, ..., with.revmap = FALSE, na.rm = FALSE)

\S4method{reduce}{GRangesAssay}(x, drop.empty.ranges = FALSE, ...)

\S4method{reduce}{Seurat}(x, drop.empty.ranges = FALSE, ...)

\S4method{gaps}{GRangesAssay}(x, start = NA, end = NA)

\S4method{gaps}{Seurat}(x, start = NA, end = NA)

\S4method{disjoin}{GRangesAssay}(x, ...)

\S4method{disjoin}{Seurat}(x, ...)

\S4method{isDisjoint}{GRangesAssay}(x, ...)

\S4method{isDisjoint}{Seurat}(x, ...)

\S4method{disjointBins}{GRangesAssay}(x, ...)

\S4method{disjointBins}{Seurat}(x, ...)
}
\arguments{
\item{x}{A \link[=GRangesAssay-class]{GRangesAssay} object}

\item{...}{Additional arguments}

\item{with.revmap}{See \code{\link[IRanges:inter-range-methods]{IRanges::inter-range-methods()}} in the
\pkg{IRanges} packages}

\item{na.rm}{Ignored}

\item{drop.empty.ranges}{See
\verb{?[inter-range-methods][IRanges::inter-range-methods]}}

\item{start, end}{See \verb{?[inter-range-methods][IRanges::inter-range-methods]}}
}
\description{
The \verb{range, reduce, gaps, disjoin, isDisjoint, disjointBins} methods
are available for \link[=GRangesAssay-class]{GRangesAssay} objects.
}
\section{Functions}{
\itemize{
\item \code{range(Seurat)}: method for \link[SeuratObject:Seurat-class]{SeuratObject::Seurat}
objects

\item \code{reduce(GRangesAssay)}: method for \link[=GRangesAssay-class]{GRangesAssay} objects

\item \code{reduce(Seurat)}: method for \link[SeuratObject:Seurat-class]{SeuratObject::Seurat}
objects

\item \code{gaps(GRangesAssay)}: method for \link[=GRangesAssay-class]{GRangesAssay} objects

\item \code{gaps(Seurat)}: method for \link[SeuratObject:Seurat-class]{SeuratObject::Seurat}
objects

\item \code{disjoin(GRangesAssay)}: method for \link[=GRangesAssay-class]{GRangesAssay} objects

\item \code{disjoin(Seurat)}: method for Seurat objects

\item \code{isDisjoint(GRangesAssay)}: method for \link[=GRangesAssay-class]{GRangesAssay} objects

\item \code{isDisjoint(Seurat)}: method for \link[SeuratObject:Seurat-class]{SeuratObject::Seurat}
objects

\item \code{disjointBins(GRangesAssay)}: method for \link[=GRangesAssay-class]{GRangesAssay} objects

\item \code{disjointBins(Seurat)}: method for \link[SeuratObject:Seurat-class]{SeuratObject::Seurat}
objects

}}
\seealso{
\itemize{
\item \link[IRanges:inter-range-methods]{IRanges::inter-range-methods}
\item \link[GenomicRanges:inter-range-methods]{GenomicRanges::inter-range-methods}
\item \link[=GRangesAssay-class]{GRangesAssay}
}
}
\concept{inter_range}
