% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{LoadGWAS}
\alias{LoadGWAS}
\title{Load GWAS-SSF file}
\usage{
LoadGWAS(gwas.file)
}
\arguments{
\item{gwas.file}{Path to GWAS/QTL summary statistics file (TSV)}
}
\value{
data.frame with following columns present: chromosome, base_pair_location, p_value
}
\description{
Load GWAS-SSF file
}
\details{
Input format based on GWAS-SSF (NHGRI-EBI GWAS Catalog standard).

Required columns: \code{chromosome}, \code{base_pair_location}, \code{p_value}

Optional columns: \code{variant_id}, \code{beta}, \code{standard_error}

Column matching is case-insensitive. Also accepts QTL data in the same format.
}
\references{
\doi{10.1101/2022.07.15.500230}
}
