% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{LoadLDData}
\alias{LoadLDData}
\title{Load LD data from LDlink or PLINK}
\usage{
LoadLDData(ld.file)
}
\arguments{
\item{ld.file}{Path to LD file (LDlink or PLINK pairwise format)}
}
\description{
Load LD data from LDlink or PLINK
}
\details{
LD must be pairwise r² relative to one lead SNP.
  Matched to GWAS by position (CHR + POS).
}
