% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LookupGeneCoords}
\alias{LookupGeneCoords}
\title{Get gene coordinates}
\usage{
LookupGeneCoords(object, gene, assay = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{gene}{Name of a gene to extract}

\item{assay}{Name of assay to use}
}
\description{
Extract the coordinates of the longest transcript for a gene stored in the
annotations within an object.
}
\examples{
LookupGeneCoords(atac_small, gene = "MIR1302-10")
}
\concept{utilities}
