% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{MergeFragments}
\alias{MergeFragments}
\title{Merge fragment bed-like files for multple 10X scATAC runs. Requires awk, sort, bgzip, and tabix.}
\usage{
MergeFragments(
  sample_meta,
  combined_fragments_file,
  do_run = F,
  force = F,
  temp_folder = "temp"
)
}
\arguments{
\item{sample_meta}{Dataframe with columns "path" to the cellranger-atac results and "sample" for the sample names. 
Sample names must behave well on the command line. This is not well tested, but for example you should avoid '$>|()\'.}

\item{combined_fragments_file}{Final result. Must end in '.gz'.}

\item{do_run}{If TRUE, runs the commands via R's system function. If FALSE, prints them for you to run manually.}

\item{force}{If TRUE, issue warnings instead of errors. Not recommended.}

\item{temp_folder}{Where to put very large intermediate files.}
}
\description{
Merge fragment bed-like files for multple 10X scATAC runs. Requires awk, sort, bgzip, and tabix.
}
