% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{MergeRegionAggregation}
\alias{MergeRegionAggregation}
\title{Condense a list of RegionAggregation objects}
\usage{
MergeRegionAggregation(x = NULL)
}
\arguments{
\item{x}{A list of \code{RegionAggregation} objects}
}
\value{
A list of \code{RegionAggregation} objects, where compatible objects
have been merged.
}
\description{
Takes a list of \code{RegionAggregation} objects and merges compatible
objects that share the same feature na,e. Compatibility is defined by
identical upstream/downstream extension, regions, and expected insertion
values. Objects with different feature names are never merged.
}
\details{
The function performs strict invariant checks and will error if:
\itemize{
\item RegionAggregation objects with the same feature name have different
region widths.
\item RegionAggregation objects with the same feature name contain
overlapping cell barcodes.
}

Compatibility between two RegionAggregation objects is determined by
\code{IsCompatibleRegionAggregation()}
}
\concept{RegionAggregation}
