% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{MotifDimPlot}
\alias{MotifDimPlot}
\title{MotifDimPlot}
\usage{
MotifDimPlot(object, assay = NULL, group.by = NULL, reduction = "tSNE")
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Which assay to use. Default is the active assay.}

\item{group.by}{A set of identities to group by (present in the Motif object metadata).}

\item{reduction}{Which dimension reduction to use. Default is tSNE.}
}
\value{
Returns a \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot motifs in reduced dimesions.
}
