% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{MotifHeatmap}
\alias{MotifHeatmap}
\title{MotifHeatmap}
\usage{
MotifHeatmap(object, motifs, assay = NULL, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{motifs}{List of motifs to plot}

\item{assay}{Which assay to use. Default is the active assay.}

\item{...}{Additional arguments}
}
\description{
Plot motif enrichment scores for groups of cells
}
