% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Motifs}
\alias{Motifs}
\alias{Motifs<-}
\alias{Motifs.ChromatinAssay5}
\alias{Motifs.Seurat}
\alias{Motifs<-.ChromatinAssay5}
\alias{Motifs<-.Seurat}
\title{Get or set a motif information}
\usage{
Motifs(object, ...)

Motifs(object, ...) <- value

\method{Motifs}{ChromatinAssay5}(object, ...)

\method{Motifs}{Seurat}(object, assay = NULL, ...)

\method{Motifs}{ChromatinAssay5}(object, ...) <- value

\method{Motifs}{Seurat}(object, assay = NULL, ...) <- value
}
\arguments{
\item{object}{A Seurat or ChromatinAssay5 object}

\item{...}{Arguments passed to other methods}

\item{value}{A \link[=Motif-class]{Motif} object}

\item{assay}{Name of assay to use}
}
\description{
Get or set the \link[=Motif-class]{Motif} object for a Seurat object or
\link[=ChromatinAssay5-class]{ChromatinAssay5}.
}
\examples{
Motifs(atac_small[["peaks"]])
Motifs(atac_small)
motifs <- Motifs(atac_small)
Motifs(atac_small[["peaks"]]) <- motifs
motifs <- Motifs(atac_small)
Motifs(atac_small) <- motifs
}
\concept{assay}
\concept{motifs}
