% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{MultiRegionCutMatrix}
\alias{MultiRegionCutMatrix}
\title{Generate cut matrix for many regions}
\usage{
MultiRegionCutMatrix(object, regions, assay = NULL, cells = NULL,
  verbose = FALSE)
}
\arguments{
\item{object}{A Seurat object}

\item{regions}{A set of GRanges}

\item{assay}{Name of the assay to use}

\item{cells}{Vector of cells to include}

\item{verbose}{Display messages}
}
\description{
Run CutMatrix on multiple regions and add them together.
Assumes regions are pre-aligned.
}
