% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iranges-methods.R
\name{nearest-methods}
\alias{nearest-methods}
\alias{precede}
\alias{precede,ANY,GRangesAssay-method}
\alias{precede,GRangesAssay,ANY-method}
\alias{precede,GRangesAssay,GRangesAssay-method}
\alias{precede,ANY,Seurat-method}
\alias{precede,Seurat,ANY-method}
\alias{precede,Seurat,Seurat-method}
\alias{follow,ANY,GRangesAssay-method}
\alias{follow}
\alias{follow,GRangesAssay,ANY-method}
\alias{follow,GRangesAssay,GRangesAssay-method}
\alias{follow,ANY,Seurat-method}
\alias{follow,Seurat,ANY-method}
\alias{follow,Seurat,Seurat-method}
\alias{nearest,ANY,GRangesAssay-method}
\alias{nearest}
\alias{nearest,GRangesAssay,ANY-method}
\alias{nearest,GRangesAssay,GRangesAssay-method}
\alias{nearest,ANY,Seurat-method}
\alias{nearest,Seurat,ANY-method}
\alias{nearest,Seurat,Seurat-method}
\alias{distance,ANY,GRangesAssay-method}
\alias{distance}
\alias{distance,GRangesAssay,ANY-method}
\alias{distance,GRangesAssay,GRangesAssay-method}
\alias{distance,ANY,Seurat-method}
\alias{distance,Seurat,ANY-method}
\alias{distance,Seurat,Seurat-method}
\alias{distanceToNearest,ANY,GRangesAssay-method}
\alias{distanceToNearest}
\alias{distanceToNearest,GRangesAssay,ANY-method}
\alias{distanceToNearest,GRangesAssay,GRangesAssay-method}
\alias{distanceToNearest,ANY,Seurat-method}
\alias{distanceToNearest,Seurat,ANY-method}
\alias{distanceToNearest,Seurat,Seurat-method}
\title{Find the nearest range neighbors for GRangesAssay objects}
\usage{
\S4method{precede}{ANY,GRangesAssay}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{precede}{GRangesAssay,ANY}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{precede}{GRangesAssay,GRangesAssay}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{precede}{ANY,Seurat}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{precede}{Seurat,ANY}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{precede}{Seurat,Seurat}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{follow}{ANY,GRangesAssay}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{follow}{GRangesAssay,ANY}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{follow}{GRangesAssay,GRangesAssay}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{follow}{ANY,Seurat}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{follow}{Seurat,ANY}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{follow}{Seurat,Seurat}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{nearest}{ANY,GRangesAssay}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{nearest}{GRangesAssay,ANY}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{nearest}{GRangesAssay,GRangesAssay}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{nearest}{ANY,Seurat}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{nearest}{Seurat,ANY}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{nearest}{Seurat,Seurat}(x, subject, select = c("arbitrary", "all"), ignore.strand = FALSE)

\S4method{distance}{ANY,GRangesAssay}(x, y, ignore.strand = FALSE, ...)

\S4method{distance}{GRangesAssay,ANY}(x, y, ignore.strand = FALSE, ...)

\S4method{distance}{GRangesAssay,GRangesAssay}(x, y, ignore.strand = FALSE, ...)

\S4method{distance}{ANY,Seurat}(x, y, ignore.strand = FALSE, ...)

\S4method{distance}{Seurat,ANY}(x, y, ignore.strand = FALSE, ...)

\S4method{distance}{Seurat,Seurat}(x, y, ignore.strand = FALSE, ...)

\S4method{distanceToNearest}{ANY,GRangesAssay}(x, subject, ignore.strand = FALSE, ...)

\S4method{distanceToNearest}{GRangesAssay,ANY}(x, subject, ignore.strand = FALSE, ...)

\S4method{distanceToNearest}{GRangesAssay,GRangesAssay}(x, subject, ignore.strand = FALSE, ...)

\S4method{distanceToNearest}{ANY,Seurat}(x, subject, ignore.strand = FALSE, ...)

\S4method{distanceToNearest}{Seurat,ANY}(x, subject, ignore.strand = FALSE, ...)

\S4method{distanceToNearest}{Seurat,Seurat}(x, subject, ignore.strand = FALSE, ...)
}
\arguments{
\item{x}{A query \link[=GRangesAssay-class]{GRangesAssay} object}

\item{subject}{The subject \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} or
\link[=GRangesAssay-class]{GRangesAssay} object. If missing, \code{x} is used as the
subject.}

\item{select}{Logic for handling ties. See \code{\link[GenomicRanges:nearest-methods]{GenomicRanges::nearest-methods()}}}

\item{ignore.strand}{Logical argument controlling whether strand information
should be ignored.}

\item{y}{For the \code{distance} method, a
\code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} object or a \link[=GRangesAssay-class]{GRangesAssay}
object}

\item{...}{Additional arguments for methods}
}
\description{
The \verb{precede, follow, nearest, distance, distanceToNearest} methods
are available for \link[=GRangesAssay-class]{GRangesAssay} objects.
}
\section{Functions}{
\itemize{
\item \code{precede(x = GRangesAssay, subject = ANY)}: method for GRangesAssay, ANY

\item \code{precede(x = GRangesAssay, subject = GRangesAssay)}: method for GRangesAssay, GRangesAssay

\item \code{precede(x = ANY, subject = Seurat)}: method for ANY, Seurat

\item \code{precede(x = Seurat, subject = ANY)}: method for Seurat, ANY

\item \code{precede(x = Seurat, subject = Seurat)}: method for Seurat, Seurat

\item \code{follow(x = ANY, subject = GRangesAssay)}: method for ANY, GRangesAssay

\item \code{follow(x = GRangesAssay, subject = ANY)}: method for GRangesAssay, ANY

\item \code{follow(x = GRangesAssay, subject = GRangesAssay)}: method for GRangesAssay, GRangesAssay

\item \code{follow(x = ANY, subject = Seurat)}: method for ANY, Seurat

\item \code{follow(x = Seurat, subject = ANY)}: method for Seurat, ANY

\item \code{follow(x = Seurat, subject = Seurat)}: method for Seurat, Seurat

\item \code{nearest(x = ANY, subject = GRangesAssay)}: method for ANY, GRangesAssay

\item \code{nearest(x = GRangesAssay, subject = ANY)}: method for GRangesAssay, ANY

\item \code{nearest(x = GRangesAssay, subject = GRangesAssay)}: method for GRangesAssay, GRangesAssay

\item \code{nearest(x = ANY, subject = Seurat)}: method for ANY, Seurat

\item \code{nearest(x = Seurat, subject = ANY)}: method for Seurat, ANY

\item \code{nearest(x = Seurat, subject = Seurat)}: method for Seurat, Seurat

\item \code{distance(x = ANY, y = GRangesAssay)}: method for ANY, GRangesAssay

\item \code{distance(x = GRangesAssay, y = ANY)}: method for GRangesAssay, ANY

\item \code{distance(x = GRangesAssay, y = GRangesAssay)}: method for GRangesAssay, GRangesAssay

\item \code{distance(x = ANY, y = Seurat)}: method for ANY, Seurat

\item \code{distance(x = Seurat, y = ANY)}: method for Seurat, ANY

\item \code{distance(x = Seurat, y = Seurat)}: method for Seurat, Seurat

\item \code{distanceToNearest(x = ANY, subject = GRangesAssay)}: method for ANY, GRangesAssay

\item \code{distanceToNearest(x = GRangesAssay, subject = ANY)}: method for GRangesAssay, ANY

\item \code{distanceToNearest(x = GRangesAssay, subject = GRangesAssay)}: method for GRangesAssay, GRangesAssay

\item \code{distanceToNearest(x = ANY, subject = Seurat)}: method for ANY, Seurat

\item \code{distanceToNearest(x = Seurat, subject = ANY)}: method for Seurat, ANY

\item \code{distanceToNearest(x = Seurat, subject = Seurat)}: method for Seurat, Seurat

}}
\seealso{
\itemize{
\item \code{\link[IRanges:nearest]{IRanges::nearest()}}
\item \code{\link[GenomicRanges:nearest]{GenomicRanges::nearest()}}
\item \link[=GRangesAssay-class]{GRangesAssay}
}
}
\concept{nearest}
