% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Period}
\alias{Period}
\title{Calculate periodogram score per cell}
\usage{
Period(object, fragments = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{fragments}{Path to an indexed fragment file}
}
\description{
Calculate periodogram score per cell
}
