% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PeriodPlot}
\alias{PeriodPlot}
\title{Plot fragment length periodicity}
\usage{
PeriodPlot(object, assay = NULL, fragment.path = NULL,
  region = "chr1-1-2000000", group.by = NULL, cells = NULL,
  log.scale = FALSE, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Which assay to use. Default is the active assay.}

\item{fragment.path}{Path to an index fragment file. If NULL, will look for a path stored for the
requested assay using the \code{SetFragments} function}

\item{region}{Genomic range to use. Default is fist two megabases of chromosome 1. Can be a GRanges object, a string, or a vector
of strings.}

\item{group.by}{Name of one or more metadata columns to group (color) the cells by. Default is the current cell identities}

\item{cells}{Which cells to plot. Default all cells}

\item{log.scale}{Display Y-axis on log scale. Default is FALSE.}

\item{...}{Additional arguments passed to \code{\link{GetReadsInRegion}}}
}
\value{
Returns a ggplot2 object
}
\description{
Plot fragment length periodicity
}
