% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PileupPlot}
\alias{PileupPlot}
\title{Plot coverage pileup centered on a given genomic feature}
\usage{
PileupPlot(object, annotation, assay = NULL, feature = "TSS",
  fragment.path = NULL, group.by = NULL, cells = NULL, cols = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Which assay to use. Default is the active assay.}

\item{feature}{Which genomic feature to center on. Options are: 'TSS', 'TTS'}

\item{fragment.path}{Path to an index fragment file. If NULL, will look for a path stored for the
requested assay using the \code{SetFragments} function}

\item{group.by}{Name of one or more metadata columns to group (color) the cells by. Default is the current cell identities}

\item{cells}{Which cells to plot. Default all cells}

\item{cols}{Vector of colors, where each color corresponds to an identity class. By default, use the ggplot2 colors.}
}
\description{
Plot coverage pileup centered on a given genomic feature
}
