% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprinting.R
\name{PlotFootprint}
\alias{PlotFootprint}
\title{Plot footprinting results}
\usage{
PlotFootprint(object, features, assay = NULL, group.by = NULL, idents = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{A vector of features to plot}

\item{assay}{Name of assay to use}

\item{group.by}{A grouping variable}

\item{idents}{Set of identities to include in the plot}
}
\description{
Plot footprinting results
}
