% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Read10X_helper}
\alias{Read10X_helper}
\title{This function was modified from Seurat R package
citation: Butler et al., Nature Biotechnology 2018}
\usage{
Read10X_helper(data.dir = NULL, barcode.file = NULL,
  gene.file = NULL, matrix.file = NULL)
}
\arguments{
\item{data.dir}{Directory containing the matrix.mtx, genes.tsv, and barcodes.tsv
files provided by 10X. A vector or named vector can be given in order to load
several data directories. If a named vector is given, the cell barcode names
will be prefixed with the name.}

\item{barcode.file}{Name of the barcode file, usually is raw txt or zipped}

\item{gene.file}{Name of the gene file, usually is raw txt or zipped}

\item{matrix.file}{Name of the matrix file, usually is raw txt or zipped}
}
\value{
Returns a sparse matrix with rows and columns labeled
}
\description{
Load in data from 10X
}
\details{
Enables easy loading of sparse data matrices provided by 10X genomics.
}
