% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Read10XRAW}
\alias{Read10XRAW}
\title{Load in data from 10X}
\usage{
Read10XRAW(data.dir, version = 2)
}
\arguments{
\item{data.dir}{folder of filtered cell matrix}

\item{version}{chemistry version}
}
\value{
list of objects for genomes, each object comprises of a list of
sparse matrices for measurements
}
\description{
Enables easy loading of sparse data matrices provided by 10X genomics from
chemistry version 2 and chemistry version 3
}
\examples{
dir <- path2your10Xdir
list.data <- Read10XRAW(dir, version = 2)

}
