% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadColSumSpMt}
\alias{ReadColSumSpMt}
\title{ReadColSumSpMt}
\usage{
ReadColSumSpMt(filePath, groupName)
}
\arguments{
\item{filePath}{A string (HDF5 path)}

\item{groupName}{A string (HDF5 dataset)}
}
\description{
Read cols sums
}
