% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ReadCounts}
\alias{ReadCounts}
\title{Read counts}
\usage{
ReadCounts(dir)
}
\arguments{
\item{dir}{Path to directory containing the matrix.mtx,
barcodes.tsv, and features.tsv files}
}
\value{
Returns a sparse matrix
}
\description{
Read count matrix from mtx file
}
\concept{utilities}
