% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ReadDelim}
\alias{ReadDelim}
\title{Read expresison matrix from CSV or TSV}
\usage{
ReadDelim(mat.path, sep = ",", header = TRUE, toSparse = TRUE)
}
\arguments{
\item{mat.path}{Path to expression matrix (can be zipped)}

\item{sep}{Data has a header. Default is TRUE}
}
\description{
Read expresison matrix from CSV or TSV
}
