% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadDoubleVector}
\alias{ReadDoubleVector}
\title{ReadDoubleVector}
\usage{
ReadDoubleVector(filePath, groupName, datasetName)
}
\arguments{
\item{filePath}{A string (HDF5 path)}

\item{groupName}{A string (HDF5 dataset)}
}
\description{
This function is used to read a double vector from hdf5 file
}
