% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadGeneExpH5}
\alias{ReadGeneExpH5}
\title{ReadH5VectorRange}
\usage{
ReadGeneExpH5(filePath, groupName, g_idx, col_idx, g_exp)
}
\arguments{
\item{filePath}{A string (HDF5 path)}

\item{groupName}{A string (HDF5 dataset)}
}
\description{
This function is used to read a vector\link{i:j} from hdf5 file
}
