% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadIntegerVector}
\alias{ReadIntegerVector}
\title{ReadIntegerVector}
\usage{
ReadIntegerVector(filePath, groupName, datasetName)
}
\arguments{
\item{filePath}{A string (HDF5 path)}

\item{groupName}{A string (HDF5 dataset)}
}
\description{
This function is used to read a integer vector from hdf5 file
}
