% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ReadMatrix}
\alias{ReadMatrix}
\title{Read expresison matrix from CSV or TSV}
\usage{
ReadMatrix(mat.path, sep = ",", header = TRUE)
}
\arguments{
\item{mat.path}{Path to expression matrix (can be zipped)}

\item{sep}{Separator. Default is <code>,</code>}

\item{sep}{Data has a header. Default is TRUE}
}
\description{
Read expresison matrix from CSV or TSV
}
