% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mito.R
\name{ReadMQuad}
\alias{ReadMQuad}
\title{Read MQuad output}
\usage{
ReadMQuad(dir, cb, verbose = TRUE)
}
\arguments{
\item{dir}{Path to directory containing MQuad output files}

\item{cb}{Path to file containing cell barcodes used to
run MQuad}

\item{verbose}{Display messages}
}
\value{
Returns a list containing two sparse matrices of
alternate allele depth (AD) and total depth (DP) counts.

The AD matrix contains the counts to the alternate
allele for the informative mt variant in each cell.

The DP matrix contains the total number of counts
for the informative mt variant in each cell.
}
\description{
Read output files from MQuad
(\url{https://github.com/single-cell-genetics/MQuad}).
}
\concept{mito}
