% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadRootDataset}
\alias{ReadRootDataset}
\title{ReadRootDataset}
\usage{
ReadRootDataset(filePath, datasetName)
}
\arguments{
\item{filePath}{A fiel path}

\item{datasetName}{A dataset name}
}
\description{
Read a string vector from root group
}
