% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadRowSumSpMt}
\alias{ReadRowSumSpMt}
\title{ReadRowSumSpMt}
\usage{
ReadRowSumSpMt(filePath, groupName)
}
\arguments{
\item{filePath}{A string (HDF5 path)}

\item{groupName}{A string (HDF5 dataset)}
}
\description{
Read rows sums
}
