% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ReadSpMt}
\alias{ReadSpMt}
\title{Read sparse matrix from HDF5 file}
\usage{
ReadSpMt(h5.path, group.name = "bioturing")
}
\arguments{
\item{h5.path}{Path to HDF5 file}

\item{group.name}{Group name in dataset. Default is "bioturing"}
}
\description{
Read sparse matrix from HDF5 file
}
\examples{
spMat <- ReadSpMt(h5.path = path2hdf5, group.name = "bioturing")
spMat

}
