% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadSpMtAsS4}
\alias{ReadSpMtAsS4}
\title{ReadSpMtAsS4}
\usage{
ReadSpMtAsS4(filePath, groupName)
}
\arguments{
\item{filePath}{A string (HDF5 path)}

\item{groupName}{A string (HDF5 dataset)}
}
\description{
This function is used to read a sparse matrix from HDF5 file
}
