% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadSpMtV1}
\alias{ReadSpMtV1}
\title{ReadSpMtV1}
\usage{
ReadSpMtV1(filePath, groupName)
}
\arguments{
\item{filePath}{A string (HDF5 path)}

\item{groupName}{A string (HDF5 dataset)}
}
\description{
This function is used to read a sparse matrix from HDF5 file
}
