% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Cells}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{SeuratObject}{\code{\link[SeuratObject:Cells]{Cells()}}}
}}

