% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{RegionAggr}
\alias{RegionAggr}
\alias{RegionAggr<-}
\alias{RegionAggr.ChromatinAssay5}
\alias{RegionAggr.Seurat}
\alias{RegionAggr<-.ChromatinAssay5}
\alias{RegionAggr<-.Seurat}
\title{Region Aggregation}
\usage{
RegionAggr(object, ...)

RegionAggr(object, ...) <- value

\method{RegionAggr}{ChromatinAssay5}(object, features = NULL, ...)

\method{RegionAggr}{Seurat}(object, assay = NULL, ...)

\method{RegionAggr}{ChromatinAssay5}(object, ...) <- value

\method{RegionAggr}{Seurat}(object, assay = NULL, ...) <- value
}
\arguments{
\item{object}{A Seurat or ChromatinAssay5 object}

\item{...}{Arguments passed to other methods}

\item{value}{A \link[=RegionAggregation-class]{RegionAggregation} object or list of
\link[=RegionAggregation-class]{RegionAggregation} objects}

\item{features}{Optional character vector of region aggregation names to
return}

\item{assay}{Name of assay to use}
}
\value{
Returns a list of \link[=RegionAggregation-class]{RegionAggregation} objects
}
\description{
Get the region aggregation information from a \link[=ChromatinAssay5-class]{ChromatinAssay5} object
}
\examples{
RegionAggr(atac_small[["peaks"]])
RegionAggr(atac_small)
ra <- RegionAggr(atac_small)
RegionAggr(atac_small[["peaks"]]) <- ra
ra <- RegionAggr(atac_small)
RegionAggr(atac_small) <- ra
}
\concept{assay}
\concept{footprinting}
\concept{links}
