% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/objects.R
\docType{class}
\name{RegionAggregation-class}
\alias{RegionAggregation-class}
\alias{subset.RegionAggregation}
\title{RegionAggregation class}
\usage{
\method{subset}{RegionAggregation}(x, cells = NULL, ...)
}
\arguments{
\item{x}{A \link[=RegionAggregation-class]{RegionAggregation} object}

\item{cells}{Vector of cells to retain}

\item{...}{Arguments passed to other methods}
}
\description{
The RegionAggregation class enables storage of counts centered on a group of
genomic regions, aggregated across regions for each cell. The main data
object stored is a cell-by-position matrix.
}
\section{Slots}{

\describe{
\item{\code{matrix}}{A cell-by-position matrix}

\item{\code{regions}}{A \code{\link[GenomicRanges:granges]{GenomicRanges::granges()}} object containing the
regions aggregated across.}

\item{\code{upstream}}{Integer denoting number of bases upstream of the centered
position that are stored in the matrix}

\item{\code{downstream}}{Integer denoting number of bases downstream of the centered
position that are stored in the matrix}

\item{\code{name}}{A name for the set of regions aggregated}

\item{\code{expected}}{A vector containing expected number of Tn5 insertions per
position}

\item{\code{cells}}{A vector of cell barcodes present in the region aggregation
matrix. Each entry in the cells vector corresponds to a row in the region
aggregation matrix.}
}}

\concept{footprinting}
