% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{RegionAggregation}
\alias{RegionAggregation}
\alias{RegionAggregation.ChromatinAssay5}
\alias{RegionAggregation.Seurat}
\title{Region Aggregation}
\usage{
RegionAggregation(object, ...)

\method{RegionAggregation}{ChromatinAssay5}(object, ...)

\method{RegionAggregation}{Seurat}(object, assay = NULL, ...)
}
\arguments{
\item{object}{A Seurat or ChromatinAssay5 object}

\item{...}{Arguments passed to other methods}

\item{assay}{Name of assay to use}
}
\value{
Returns a list of \code{RegionAggregation} objects
}
\description{
Get the region aggregation information from a ChromatinAssay5 object
}
\examples{
RegionAggregation(atac_small[["peaks"]])
RegionAggregation(atac_small)
}
\concept{assay}
\concept{footprinting}
