% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{RegionHeatmap}
\alias{RegionHeatmap}
\title{Region heatmap}
\usage{
RegionHeatmap(
  object,
  key,
  window = 200,
  normalize = TRUE,
  order = "sum",
  assay = NULL,
  idents = NULL
)
}
\arguments{
\item{object}{A Seurat object}

\item{key}{Name of key to pull data from. Stores the results from
\code{\link{RegionMatrix}}}

\item{window}{Smoothing window to apply}

\item{normalize}{Normalize by number of cells in each group}

\item{order}{Define the order for plotting regions. If "sum," order by the 
total number of fragments in the region across all included identities}

\item{assay}{Name of assay to use}

\item{idents}{Cell identities to include. Note that cells cannot be
regrouped, this will require re-running \code{RegionMatrix} to generate a 
new set of matrices}
}
\value{
Returns a ggplot2 object
}
\description{
Plot fragment counts within a set of regions.
}
\seealso{
RegionMatrix
}
\concept{heatmap}
\concept{visualization}
