% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{RegionPileup}
\alias{RegionPileup}
\title{Plot pileup of Tn5 integration sites}
\usage{
RegionPileup(object, regions, assay = NULL, upstream = 200,
  downstream = 200, group.by = NULL, min.cells = 100, ymax = NULL,
  idents = NULL, verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{regions}{A set of GRanges to use}

\item{assay}{Name of the assay to use}

\item{upstream}{Number of bases to extend upstream of the region midpoint}

\item{downstream}{Number of bases to extend downstream of the region midpoint}

\item{group.by}{A set of identities to group the cells by. Can by anything in the metadata.
Default is to use the active identities.}

\item{min.cells}{Minimum number of cells in the group for the pileup to be displayed for that group.}

\item{ymax}{Maximum value for the y-axis. If NULL (default), will be set automatically.}

\item{idents}{Which identities to include in the plot. If NULL (default), include everything with more than
\code{min.cells} cells.}

\item{verbose}{Display messages}
}
\value{
Returns a \code{\link[ggplot2]{ggplot2}} object
}
\description{
Plots a pileup of integration sites centered on a set of genomic positions.
Each genomic region will be aligned on the region midpoint, and extended upstream
and downstream from the midpoint.
}
\examples{
\dontrun{

}
}
