% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{RenameCells.ChromatinAssay}
\alias{RenameCells.ChromatinAssay}
\alias{RenameCells.Fragment}
\title{Rename cells in a ChromatinAssay}
\usage{
\method{RenameCells}{ChromatinAssay}(object, new.names = NULL, ...)

\method{RenameCells}{Fragment}(object, new.names, ...)
}
\description{
Method for renaming cells in a \code{\link{ChromatinAssay}} object
}
\seealso{
\code{\link[Seurat]{RenameCells}} in the \pkg{Seurat} package
}
