% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ReturnTripleFiles}
\alias{ReturnTripleFiles}
\title{Flexible check file exists in an 10X folder}
\usage{
ReturnTripleFiles(dir)
}
\arguments{
\item{dir}{to the a 10X folder}
}
\value{
a list of three file names: barcode, genes, and matrix
}
\description{
Flexible check file exists in an 10X folder
}
