% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{RunFragtk}
\alias{RunFragtk}
\title{Run fragtk matrix}
\usage{
RunFragtk(
  fragments,
  features,
  cells,
  group = FALSE,
  pic = TRUE,
  fragtk.path = NULL,
  outdir = tempdir(),
  cleanup = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fragments}{A list of Fragment objects or fragment file paths}

\item{features}{A GRanges object containing a set of genomic intervals to
quantify. These genomic ranges will be passed to the `--bed` argument in
`fragtk matrix`}

\item{cells}{List of cells to include}

\item{group}{Group genomic ranges according to a grouping variable. If NULL,
no grouping variable is used. If a character string is provided, this should
match a column in the provided GRanges object supplied in the `features`
parameter.}

\item{pic}{Use paired insertion counting}

\item{fragtk.path}{Path to fragtk executable. If NULL, try to find fragtk
automatically.}

\item{outdir}{Path for output directory}

\item{cleanup}{Remove output files created by fragtk}

\item{verbose}{Display messages}
}
\value{
Returns a CsparseMatrix
}
\description{
Wrapper function to run \code{fragtk matrix} and return the output as a sparse
matrix in R.
}
\details{
See \url{https://crates.io/crates/fragtk} for fragtk documentation.
}
