% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimension_reduction.R
\name{RunMotifTSNE}
\alias{RunMotifTSNE}
\alias{RunMotifTSNE.Motif}
\alias{RunMotifTSNE.Assay}
\alias{RunMotifTSNE.Seurat}
\title{RunMotifTSNE}
\usage{
RunMotifTSNE(object, ...)

\method{RunMotifTSNE}{Motif}(object, verbose = TRUE, graph.name = "nn",
  ...)

\method{RunMotifTSNE}{Assay}(object, verbose = TRUE, ...)

\method{RunMotifTSNE}{Seurat}(object, assay = NULL, verbose = TRUE,
  ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{verbose}{Display messages}

\item{graph.name}{Name of the neighbor graph to use ('nn' or 'snn')}

\item{assay}{Name of assay to use}
}
\value{
Returns a Seurat object
}
\description{
Run tSNE on a Mofif object. This will project the motifs into two tSNE dimensions based on
a neighbor graph stored in the Motif object (commonly computed from the Jaccard similarity between motifs).
}
