% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ScCutMatrix}
\alias{ScCutMatrix}
\title{Generate a matrix of integration sites in sliding windows}
\usage{
ScCutMatrix(
  object,
  region,
  tabix.file = NULL,
  assay = NULL,
  cells = NULL,
  window_size = 100,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{region}{A GRanges object containing the region of interest}

\item{tabix.file}{A TabixFile object. If NULL, the file specified in
\code{fragment.path} will be opened and closed after the function completes.
If iterating over many regions, providing an open TabixFile is much faster
as it avoids opening and closing the connection each time.}

\item{assay}{Name of the assay to use}

\item{cells}{Which cells to include in the matrix. If NULL (default), use all
cells in the object}

\item{window_size}{The sliding window size. Default: 100.}

\item{verbose}{Default: TRUE}
}
\value{
A matrix.
}
\description{
Generates a cell-by-position matrix of Tn5 integration sites in 
sliding windows.
}
