% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ScCutTilePlot}
\alias{ScCutTilePlot}
\title{Tile plot for integration sites in sliding windows for cells}
\usage{
ScCutTilePlot(mat, gloc_lim, cells_group = NULL)
}
\arguments{
\item{mat}{A matrix returned by \code{\link{ScCutMatrix}}}

\item{gloc_lim}{Genomic location to focus on.}

\item{cells_group}{A named character vector to group cell names.}
}
\value{
ggplot
}
\description{
Tile plot for integration sites in sliding windows for cells
}
