% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqinfo-methods.R
\name{seqinfo-methods}
\alias{seqinfo-methods}
\alias{seqinfo}
\alias{seqinfo,GRangesAssay-method}
\alias{seqlevels,GRangesAssay-method}
\alias{seqlevels}
\alias{seqnames,GRangesAssay-method}
\alias{seqnames}
\alias{seqlengths,GRangesAssay-method}
\alias{seqlengths}
\alias{genome,GRangesAssay-method}
\alias{genome}
\alias{isCircular,GRangesAssay-method}
\alias{isCircular}
\alias{seqinfo,Seurat-method}
\alias{seqlevels,Seurat-method}
\alias{seqnames,Seurat-method}
\alias{seqlengths,Seurat-method}
\alias{genome,Seurat-method}
\alias{isCircular,Seurat-method}
\title{Access sequence information for GRangesAssay objects}
\usage{
\S4method{seqinfo}{GRangesAssay}(x)

\S4method{seqlevels}{GRangesAssay}(x)

\S4method{seqnames}{GRangesAssay}(x)

\S4method{seqlengths}{GRangesAssay}(x)

\S4method{genome}{GRangesAssay}(x)

\S4method{isCircular}{GRangesAssay}(x)

\S4method{seqinfo}{Seurat}(x)

\S4method{seqlevels}{Seurat}(x)

\S4method{seqnames}{Seurat}(x)

\S4method{seqlengths}{Seurat}(x)

\S4method{genome}{Seurat}(x)

\S4method{isCircular}{Seurat}(x)
}
\arguments{
\item{x}{A \link[=GRangesAssay-class]{GRangesAssay} object}
}
\description{
Methods for accessing the \code{\link[Seqinfo:Seqinfo]{Seqinfo::Seqinfo()}} object information stored in a
\link[=GRangesAssay-class]{GRangesAssay} object. To change the sequence information, set the
ranges with \verb{granges<-}.
}
\section{Functions}{
\itemize{
\item \code{seqlevels(GRangesAssay)}: get method for GRangesAssay objects

\item \code{seqnames(GRangesAssay)}: get method for GRangesAssay objects

\item \code{seqlengths(GRangesAssay)}: get method for GRangesAssay objects

\item \code{genome(GRangesAssay)}: get method for GRangesAssay objects

\item \code{isCircular(GRangesAssay)}: get method for GRangesAssay objects

\item \code{seqinfo(Seurat)}: get method for Seurat objects

\item \code{seqlevels(Seurat)}: get method for Seurat objects

\item \code{seqnames(Seurat)}: get method for Seurat objects

\item \code{seqlengths(Seurat)}: get method for Seurat objects

\item \code{genome(Seurat)}: get method for Seurat objects

\item \code{isCircular(Seurat)}: get method for Seurat objects

}}
\seealso{
\itemize{
\item{\link[Seqinfo:seqinfo]{seqinfo} in the \pkg{Seqinfo} package.}
\item{\link[=GRangesAssay-class]{GRangesAssay}}
}
}
\concept{seqinfo}
