% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SetAnnotations}
\alias{SetAnnotations}
\title{Add genomic annotation information to a Seurat object}
\usage{
SetAnnotations(object, annotation, key)
}
\arguments{
\item{object}{A Seurat object}

\item{annotation}{A genomic feature annotation}

\item{key}{Key used to store the annotation information, eg "gene", "TE", "promoter", etc.}
}
\description{
Add genomic annotation information to a Seurat object
}
