% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signac.R
\docType{class}
\name{Signac-class}
\alias{Signac-class}
\alias{signac}
\title{Signac class}
\description{
All functions will operate on this object
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{The original expression matrix}

\item{\code{filtered.data}}{The filtered expression matrix}

\item{\code{corrected.data}}{Data after batch effect removal}

\item{\code{scaled.data}}{Data after normalization}

\item{\code{log.data}}{Data after log}

\item{\code{cell.names}}{Id of cells}

\item{\code{batch}}{Batches for data come from multiple runs/sources}

\item{\code{metadata}}{Labels and annotations}

\item{\code{vdj}}{TCR information}

\item{\code{ssn}}{Shared nearest neighbor}

\item{\code{dimred}}{A named list of dimensionality reduction result}

\item{\code{assay}}{A named list of multimodal analysis}

\item{\code{trajectory}}{Pseudotime result}
}}

