% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimension_reduction.R
\name{SparsePCA}
\alias{SparsePCA}
\alias{SparsePCA.default}
\alias{SparsePCA.StdAssay}
\alias{SparsePCA.Assay}
\alias{SparsePCA.Seurat}
\title{Compute principal components}
\usage{
SparsePCA(object, ...)

\method{SparsePCA}{default}(
  object,
  assay = NULL,
  n = 50,
  weight.by.var = TRUE,
  reduction.key = "PCA_",
  irlba.work = n * 3,
  tol = 1e-05,
  verbose = TRUE,
  ...
)

\method{SparsePCA}{StdAssay}(
  object,
  assay = NULL,
  features = NULL,
  n = 50,
  weight.by.var = TRUE,
  irlba.work = n * 3,
  tol = 1e-05,
  reduction.key = "PCA_",
  verbose = TRUE,
  ...
)

\method{SparsePCA}{Assay}(
  object,
  assay = NULL,
  features = NULL,
  n = 50,
  weight.by.var = TRUE,
  irlba.work = n * 3,
  tol = 1e-05,
  reduction.key = "PCA_",
  verbose = TRUE,
  ...
)

\method{SparsePCA}{Seurat}(
  object,
  assay = NULL,
  features = NULL,
  n = 50,
  weight.by.var = TRUE,
  irlba.work = n * 3,
  tol = 1e-05,
  reduction.key = "PCA_",
  reduction.name = "pca",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{assay}{Which assay to use. If NULL, use the default assay}

\item{n}{Number of singular values to compute}

\item{weight.by.var}{Weight PCs by variance explained}

\item{reduction.key}{Key for dimension reduction object}

\item{irlba.work}{work parameter for \code{\link[irlba]{irlba}}.
Working subspace dimension, larger values can speed convergence at the
cost of more memory use.}

\item{tol}{Tolerance (tol) parameter for \code{\link[irlba]{irlba}}. Larger
values speed up convergence due to greater amount of allowed error.}

\item{verbose}{Print messages}

\item{features}{Which features to use. If NULL, use variable features}

\item{reduction.name}{Name for stored dimension reduction object.
Default 'pca'}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Compute principal components using \code{\link[irlba]{irlba}}. This function
is similar to the \code{\link{RunPCA}} function but uses implicit
scaling and centering of the data matrix to compute PCA faster and using less
memory. The scaled and centered data matrix is not stored, and the
\code{\link{ScaleData}} function does not need to be run before using this
function.
}
\examples{
x <- matrix(data = rnorm(100), ncol = 10)
SparsePCA(x)
\dontrun{
SparsePCA(atac_small[['peaks']])
}
\dontrun{
SparsePCA(atac_small[['peaks']])
}
\dontrun{
SparsePCA(atac_small)
}
}
\concept{dimension_reduction}
