% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SparsifiedRanks}
\alias{SparsifiedRanks}
\title{Replace non-zero entries in a sparse entries with non-zero ranks}
\usage{
SparsifiedRanks(X)
}
\description{
This method creates a rank matrix for a sparse matrix X using the following approach:
1. Use non-zero enries in a column to calculate the ranks
2. Add (z-1)/2 to the ranks (only non-zero entries are changed). z is the number of zeros
in the column
Since all the entries are shifted by the same constant (the zeros
are already shifted), the covariance matrix of this shifted matrix is
the same as the rank matrix of the entire matrix (where the zeros would
all also have a rank = (z+1)/2) where z is the number of zeros
}
\details{
This rank matrix can then be used to calculate pearson correlation
(pearson correlation )
}
