% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{StartHttpServer}
\alias{StartHttpServer}
\title{StartHttpServer}
\usage{
StartHttpServer(server.host = "127.0.0.1", server.port = 9009)
}
\arguments{
\item{server.host}{Server host}

\item{server.port}{Server port}
}
\value{
Returns HTTP SERVER instance
}
\description{
Create HTTP server
}
