% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{StopHttpServer}
\alias{StopHttpServer}
\title{StopHttpServer}
\usage{
StopHttpServer()
}
\arguments{
\item{server.host}{Server host}

\item{server.port}{Server port}
}
\value{
Returns HTTP SERVER instance
}
\description{
Stop HTTP server
}
