% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{subset.RegionAggregation}
\alias{subset.RegionAggregation}
\title{Subset a single RegionAggregation object}
\usage{
\method{subset}{RegionAggregation}(x, cells = NULL, ...)
}
\description{
Subset a single RegionAggregation object
}
\concept{RegionAggregation}
