% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SumMatrixByGroup}
\alias{SumMatrixByGroup}
\title{Sum integration sites per base per group}
\usage{
SumMatrixByGroup(mat, groups, group.scale.factors, normalize = TRUE,
  scale.factor = NULL)
}
\arguments{
\item{mat}{A cut matrix. See \code{\link{CutMatrix}}}

\item{groups}{A vector of group identities, with the name
of each element in the vector set to the cell name.}

\item{group.scale.factors}{Scaling factor for each group. Should
be computed using the number of cells in the group and the average number of counts
in the group.}

\item{normalize}{Perform sequencing depth and cell count normalization (default is TRUE)}

\item{scale.factor}{Scaling factor to use. If NULL (default), will use the median normalization
factor for all the groups.}
}
\description{
Perform colSums on a cut matrix with cells in the rows
and position in the columns, for each group of cells
separately.
}
