% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TabixOutputToDataFrame}
\alias{TabixOutputToDataFrame}
\title{TabixOutputToDataFrame}
\usage{
TabixOutputToDataFrame(reads, record.ident = TRUE)
}
\arguments{
\item{reads}{List of character vectors (the output of \code{\link{scanTabix}})}

\item{record.ident}{Add a column recording which region the reads overlapped with (default TRUE)}
}
\value{
Returns a data.frame
}
\description{
Create a single dataframe from list of character vectors
}
