% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{TestEnrichment}
\alias{TestEnrichment}
\title{TestEnrichment}
\usage{
TestEnrichment(motif.matrix, feature.matrix)
}
\arguments{
\item{motif.matrix}{A feature x motif matrix}

\item{feature.matrix}{A feature x cell matrix}
}
\value{
Returns a vector of motif counts
}
\description{
Compute motif counts per cell for a given set of features
}
