% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{theme_browser}
\alias{theme_browser}
\title{Genome browser theme}
\usage{
theme_browser(..., legend = TRUE)
}
\arguments{
\item{...}{Additional arguments}

\item{legend}{Display plot legend}
}
\description{
Theme applied to panels in the \code{\link{CoveragePlot}} function.
}
\examples{
PeakPlot(atac_small, region = "chr1-710000-715000") + theme_browser()
}
\concept{visualization}
