% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{TopLinks}
\alias{TopLinks}
\title{Find top links}
\usage{
TopLinks(links, cutoff = 0.25)
}
\arguments{
\item{links}{A sparse matrix containing genes in the rows and enhancers in
the columns.
Values correspond to the link score.}

\item{cutoff}{Percentile cutoff for retaining links. 0.50 (default) retains
the top 50\% of links.}
}
\value{
Returns a sparse matrix
}
\description{
Filter a set of promoter-enhancer links and retain the top X\%
based on correlation score. Links below the percentile cutoff
will be set to zero. Note that the cutoff is defined across
all genes, so some genes may have no links after filtering.
The cutoff value is determined considering non-zero link
scores only (eg, top 50\% among non-zero scores).
}
