% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{UpdateChromatinAssay}
\alias{UpdateChromatinAssay}
\title{Update Chromatin Assay}
\usage{
UpdateChromatinAssay(object)
}
\description{
Update an old ChromatinAssay object to the current GRangesAssay object class
}
