% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{UpdateChromatinObject}
\alias{UpdateChromatinObject}
\title{Update Chromatin Assay}
\usage{
UpdateChromatinObject(
  object,
  chromatin.assay,
  expression.assay = NULL,
  features = NULL
)
}
\description{
Update an old ChromatinAssay object to the current GRangesAssay object class
}
