% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{UpdatePath}
\alias{UpdatePath}
\title{Update the file path for a Fragment object}
\usage{
UpdatePath(object, new.path, new.index.path = NULL, verbose = TRUE)
}
\arguments{
\item{object}{A \link[=Fragment2-class]{Fragment2} object.}

\item{new.path}{Path to the fragment file.}

\item{new.index.path}{Path to the fragment file index. If NULL, the index is
assumed to be in the same directory as the fragment file.}

\item{verbose}{Display messages.}
}
\description{
Change the path to a fragment file store in a \link[=Fragment2-class]{Fragment2}
object. Path must be to the same file that was used to create the fragment
object. An MD5 hash will be computed using the new path and compared to the
hash stored in the Fragment object to verify that the files are the same.
}
\concept{fragments}
