% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venice.R
\name{VeniceAllMarkers}
\alias{VeniceAllMarkers}
\title{VeniceAllMarkers}
\usage{
VeniceAllMarkers(
  object,
  assay = "RNA",
  slot = "data",
  pvalue = 0.05,
  node = NULL,
  verbose = TRUE,
  only.pos = FALSE,
  logfc.threshold = 0.25,
  nperm = 0
)
}
\arguments{
\item{object}{An Seurat v3 object which contain at least one identification of clusters}

\item{assay}{name of the assay to be used}

\item{slot}{name of the slot to be used}

\item{pvalue}{cut off p.value}

\item{node}{input a tree of clusters instead of a cluster name}

\item{verbose}{display the progress bar or not}

\item{only.pos}{filter the positive markers only}
}
\description{
This function helps to find marker genes for all clusters in a Seurat v3 object.
We adapted some utilities functions in Seurat v3 package (Butler et al., Nature Biotechnology 2018) to handle the annotation of an
object, extract the counts data from an object, and build the cluster tree.
}
