% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venice.R
\name{VeniceFindMarkers}
\alias{VeniceFindMarkers}
\title{This function helps to find marker genes for two clusters in a Seurat v3 object.
We adapted some utilities functions in Seurat v3 package (Butler et al., Nature Biotechnology 2018) to handle the annotation of an
object, extract the counts data from an object, and build the cluster tree.
Some parameters are derived from the FindMarkers function of Seurat package v3
https://github.com/satijalab/seurat/blob/master/R/differential_expression.R}
\usage{
VeniceFindMarkers(
  object,
  ident.1 = NULL,
  ident.2 = NULL,
  group.by = NULL,
  subset.ident = NULL,
  assay = NULL,
  slot = "data",
  reduction = NULL,
  features = NULL,
  logfc.threshold = 0.25,
  verbose = TRUE,
  pvalue = 0.05,
  only.pos = FALSE,
  nperm = 0
)
}
\arguments{
\item{ident.1}{Identity class to define markers for; pass an object of class
\code{phylo} or 'clustertree' to find markers for a node in a cluster tree;
passing 'clustertree' requires \code{\link{BuildClusterTree}} to have been run}

\item{ident.2}{A second identity class for comparison; if \code{NULL},
use all other cells for comparison; if an object of class \code{phylo} or
'clustertree' is passed to \code{ident.1}, must pass a node to find markers for}

\item{group.by}{Regroup cells into a different identity class prior to performing differential expression (see example)}

\item{subset.ident}{Subset a particular identity class prior to regrouping. Only relevant if group.by is set (see example)}

\item{assay}{Assay to use in differential expression testing}

\item{slot}{Slot to pull data from; note that if \code{test.use} is "negbinom", "poisson", or "DESeq2",
\code{slot} will be set to "counts"}
}
\description{
This function helps to find marker genes for two clusters in a Seurat v3 object.
We adapted some utilities functions in Seurat v3 package (Butler et al., Nature Biotechnology 2018) to handle the annotation of an
object, extract the counts data from an object, and build the cluster tree.
Some parameters are derived from the FindMarkers function of Seurat package v3
https://github.com/satijalab/seurat/blob/master/R/differential_expression.R
}
