% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VeniceMarker}
\alias{VeniceMarker}
\title{VeniceMarker}
\usage{
VeniceMarker(
  S4_mtx,
  cluster,
  threshold = 0L,
  perm = 0L,
  correct = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{S4_mtx}{A sparse matrix}

\item{cluster}{A numeric vector}
}
\description{
Find gene marker for a cluster in sparse matrix
}
