% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VeniceMarkerH5}
\alias{VeniceMarkerH5}
\title{VeniceMarkerH5}
\usage{
VeniceMarkerH5(hdf5Path, cluster, threshold = 0L)
}
\arguments{
\item{hdf5Path}{A string path}

\item{cluster}{A numeric vector}
}
\description{
Find gene marker for a cluster in H5 file
}
