% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VeniceMarkerTransposedH5}
\alias{VeniceMarkerTransposedH5}
\title{VeniceMarkerTransposedH5}
\usage{
VeniceMarkerTransposedH5(
  hdf5Path,
  group_name,
  cluster,
  threshold = 0L,
  threshold_pct = 0,
  perm = 0L,
  correct = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{hdf5Path}{A string path}

\item{cluster}{A numeric vector}
}
\description{
Find gene marker for a cluster in tranposed H5 file
}
