% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{WeightAccessibility}
\alias{WeightAccessibility}
\title{Weight accessibility score based on distance from Tn5 integration site}
\usage{
WeightAccessibility(bases_away, internal, rate = 2, inflection = 20)
}
\arguments{
\item{bases_away}{Number of bases from the Tn5 integration site}

\item{internal}{Whether the postion is within a fragment (TRUE) or flanking a fragment (FALSE)}

\item{rate}{Rate of score decay. Lower value gives steeper slope in sigmoid function.}

\item{inflection}{Inflection point for sigmoid weight function. This base will have weight
equal to 0.5}
}
\description{
Weight accessibility score based on distance from Tn5 integration site
}
