% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WriteRootDataset}
\alias{WriteRootDataset}
\title{WriteRootDataset}
\usage{
WriteRootDataset(filePath, datasetName, datasetVal)
}
\arguments{
\item{filePath}{A fiel path}

\item{datasetName}{A dataset name}

\item{datasetVal}{A string vector}
}
\description{
Write a string vector to root group
}
