% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WriteSpMtAsS4}
\alias{WriteSpMtAsS4}
\title{WriteSpMtAsS4}
\usage{
WriteSpMtAsS4(filePath, groupName, mat)
}
\arguments{
\item{filePath}{A string (HDF5 path)}

\item{groupName}{A string (HDF5 dataset)}

\item{mat}{A sparse matrix}
}
\description{
This function is used to write a sparse S4 matrix
}
