% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WriteSpMtAsSpMatFromS4}
\alias{WriteSpMtAsSpMatFromS4}
\title{WriteSpMtAsSpMatFromS4}
\usage{
WriteSpMtAsSpMatFromS4(filePath, groupName, mat)
}
\arguments{
\item{filePath}{A string (HDF5 path)}

\item{groupName}{A string (HDF5 dataset)}

\item{mat}{A sparse matrix}
}
\description{
This function is used to write a sparse ARMA matrix from S4
}
