% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WriteSpMtV2}
\alias{WriteSpMtV2}
\title{WriteSpMtV2}
\usage{
WriteSpMtV2(filePath, groupName, mat)
}
\arguments{
\item{filePath}{A string (HDF5 path)}

\item{groupName}{A string (HDF5 dataset)}

\item{mat}{A sparse matrix}
}
\description{
This function is used to write a sparse matrix (as arma)
}
