% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{AddMotifs}
\alias{AddMotifs}
\alias{AddMotifs.default}
\alias{AddMotifs.ChromatinAssay}
\alias{AddMotifs.Seurat}
\title{Add DNA sequence motif information}
\usage{
AddMotifs(object, ...)

\method{AddMotifs}{default}(object, genome, pfm, verbose = TRUE, ...)

\method{AddMotifs}{ChromatinAssay}(object, genome, pfm, verbose = TRUE, ...)

\method{AddMotifs}{Seurat}(object, genome, pfm, assay = NULL, verbose = TRUE, ...)
}
\arguments{
\item{object}{A Seurat object or ChromatinAssay object}

\item{...}{Additional arguments passed to other methods}

\item{genome}{A \code{BSgenome} object}

\item{pfm}{A \code{PFMatrixList} object}

\item{verbose}{Display messages}

\item{assay}{Name of assay to use. If NULL, use the default assay}
}
\description{
Construct a \code{\link{Motif}} object containing DNA sequence motif
information and add it to an existing Seurat object or ChromatinAssay.
If running on a Seurat object, \code{AddMotifs} will also run
\code{\link{RegionStats}} to compute the GC content of each peak and store
the results in the feature metadata.
}
\concept{motifs}
