% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Annotation}
\alias{Annotation}
\alias{Annotation.ChromatinAssay}
\title{Annotation}
\usage{
Annotation(object, ...)

\method{Annotation}{ChromatinAssay}(object)
}
\arguments{
\item{object}{A ChromatinAssay object}

\item{...}{Arguments passed to other methods}

\item{object}{A ChromatinAssay object}
}
\value{
Returns a Seurat object
}
\description{
Get annotation name from ChromatinAssay
}
