% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Annotation}
\alias{Annotation}
\alias{Annotation.ChromatinAssay}
\alias{Annotation.Seurat}
\title{Annotation}
\usage{
Annotation(object, ...)

\method{Annotation}{ChromatinAssay}(object, ...)

\method{Annotation}{Seurat}(object, ...)
}
\arguments{
\item{object}{A Seurat object or ChromatinAssay object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a \code{\link[GenomicRanges]{GRanges}} object
if the annotation data is present, otherwise returns NULL
}
\description{
Get the annotation from a ChromatinAssay
}
