% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{AnnotationPlot}
\alias{AnnotationPlot}
\title{Plot gene annotations}
\usage{
AnnotationPlot(object, region)
}
\arguments{
\item{object}{A \code{\link[Seurat]{Seurat}} object}

\item{region}{A genomic region to plot}
}
\value{
Returns a \code{\link[ggplot2]{ggplot}} object
}
\description{
Display gene annotations in a given region of the genome.
}
\examples{
\donttest{
AnnotationPlot(object = atac_small, region = c("chr1-29554-39554"))
}
}
\concept{visualization}
