% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{as.ChromatinAssay}
\alias{as.ChromatinAssay}
\alias{as.ChromatinAssay.Assay}
\title{Convert objects to a ChromatinAssay}
\usage{
as.ChromatinAssay(x, ...)

\method{as.ChromatinAssay}{Assay}(x, ranges = NULL, genome = NULL,
  annotation = NULL, motifs = NULL, fragments = NULL, bias = NULL,
  sep = c("-", "-"), ...)
}
\arguments{
\item{x}{An object to convert to class \code{\link{ChromatinAssay}}}

\item{...}{Arguments passed to other methods}

\item{ranges}{A GRanges object}

\item{genome}{Name of genome used}

\item{annotation}{Genomic annotation}

\item{motifs}{A Motif object}

\item{fragments}{Path to fragments file}

\item{sep}{Charaters used to separate the chromosome, start, and end coordinates
in the row names of the data matrix}
}
\description{
Convert objects to a ChromatinAssay
}
