% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{as.ChromatinAssay5}
\alias{as.ChromatinAssay5}
\alias{as.ChromatinAssay5.Assay5}
\title{Convert objects to a ChromatinAssay}
\usage{
as.ChromatinAssay5(x, ...)

\method{as.ChromatinAssay5}{Assay5}(
  x,
  annotation = NULL,
  fragments = NULL,
  bias = NULL,
  motifs = NULL,
  links = NULL,
  positionEnrichment = NULL,
  ...
)
}
\arguments{
\item{x}{An object to convert to class \code{\link{ChromatinAssay5}}}

\item{...}{Arguments passed to other methods}

\item{annotation}{Genomic annotation. It must have the following columns:
\itemize{
  \item{tx_id or transcript_id: Transcript ID}
  \item{gene_name: Gene name}
  \item{gene_id: Gene ID}
  \item{gene_biotype: Gene biotype (e.g. "protein_coding", "lincRNA")}
  \item{type: Annotation type (e.g. "exon", "gap")}
}}

\item{fragments}{A list of \code{\link{Fragment}} objects}

\item{bias}{Tn5 integration bias matrix}

\item{motifs}{A \code{\link{Motif}} object}

\item{links}{Genomic links TODO}

\item{positionEnrichment}{A named list of position enrichment matrices.}
}
\description{
Convert objects to a ChromatinAssay
}
\concept{assay}
