% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{as.GRangesAssay}
\alias{as.GRangesAssay}
\alias{as.GRangesAssay.ChromatinAssay}
\alias{as.GRangesAssay.Assay5}
\alias{as.GRangesAssay.ChromatinAssay5}
\title{Convert objects to a \linkS4class{GRangesAssay} object}
\usage{
as.GRangesAssay(x, ...)

\method{as.GRangesAssay}{ChromatinAssay}(x, ...)

\method{as.GRangesAssay}{Assay5}(
  x,
  annotation = NULL,
  fragments = NULL,
  bias = NULL,
  region.aggregation = NULL,
  ranges = NULL,
  motifs = NULL,
  links = NULL,
  sep = c("-", "-"),
  ...
)

\method{as.GRangesAssay}{ChromatinAssay5}(x, ranges = NULL, sep = c("-", "-"), ...)
}
\arguments{
\item{x}{An object to convert to class \linkS4class{GRangesAssay}}

\item{...}{Arguments passed to other methods}

\item{ranges}{A GRanges object}

\item{sep}{Characters used to separate the chromosome, start, and end
coordinates in the row names of the data matrix}
}
\description{
Convert objects to a \linkS4class{GRangesAssay} object
}
\concept{assay}
