% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{as.GRangesAssay}
\alias{as.GRangesAssay}
\alias{as.GRangesAssay.ChromatinAssay5}
\title{Convert objects to a GRangesAssay}
\usage{
as.GRangesAssay(x, ...)

\method{as.GRangesAssay}{ChromatinAssay5}(
  x,
  ranges = NULL,
  motifs = NULL,
  links = NULL,
  sep = c("-", "-"),
  ...
)
}
\arguments{
\item{x}{An object to convert to class \code{\link{GRangesAssay}}}

\item{...}{Arguments passed to other methods}

\item{ranges}{A GRanges object}

\item{motifs}{A \code{\link{Motif}} object}

\item{links}{Genomic links TODO}

\item{sep}{Characters used to separate the chromosome, start, and end
coordinates in the row names of the data matrix}
}
\description{
Convert objects to a GRangesAssay
}
\concept{assay}
