% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atac_small}
\alias{atac_small}
\title{A small example scATAC-seq dataset}
\format{
A Seurat object with the following assays
\describe{
  \item{peaks}{A peak x cell dataset}
  \item{bins}{A 5 kb genome bin x cell dataset}
  \item{RNA}{A gene x cell dataset}
}
}
\source{
\url{https://support.10xgenomics.com/single-cell-atac/datasets/1.1.0/atac_v1_pbmc_10k}
}
\usage{
atac_small
}
\description{
A subsetted version of 10x Genomics 10k human (hg19) PBMC scATAC-seq dataset
}
\concept{data}
\keyword{datasets}
