% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{ATACqc}
\alias{ATACqc}
\alias{ATACqc.default}
\alias{ATACqc.Fragment2}
\alias{ATACqc.ChromatinAssay5}
\alias{ATACqc.Seurat}
\title{Compute scATAC-seq QC metrics}
\usage{
ATACqc(object, ...)

\method{ATACqc}{default}(
  object,
  annotations,
  fragtk.path = NULL,
  outdir = tempdir(),
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{ATACqc}{Fragment2}(
  object,
  annotations,
  fragtk.path = NULL,
  outdir = tempdir(),
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{ATACqc}{ChromatinAssay5}(
  object,
  annotations = NULL,
  fragtk.path = NULL,
  outdir = tempdir(),
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{ATACqc}{Seurat}(
  object,
  assay = NULL,
  annotations = NULL,
  fragtk.path = NULL,
  outdir = tempdir(),
  cleanup = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \link[SeuratObject:Seurat]{SeuratObject::Seurat} object, \link[=ChromatinAssay5-class]{ChromatinAssay5} object
, or path to a fragment file.}

\item{...}{Arguments passed to other methods}

\item{annotations}{\code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} object containing
gene annotations. If \code{NULL}, attempt to extract this from the
\link[=ChromatinAssay5-class]{ChromatinAssay5} object if provided.}

\item{fragtk.path}{Path to \code{fragtk} executable. If NULL, try to find \code{fragtk}
automatically.}

\item{outdir}{Path for output directory}

\item{cleanup}{Remove output files created by fragtk}

\item{verbose}{Display messages}

\item{assay}{Name of assay to use. If NULL, use the default assay.}
}
\value{
When run on a \link[SeuratObject:Seurat]{SeuratObject::Seurat} object, returns the input
object with per-cell QC metrics added to the cell metadata. When run on a
\link[=ChromatinAssay5-class]{ChromatinAssay5}, \link[=Fragment2-class]{Fragment2}, or path to a fragment file,
returns a \code{data.frame} of per-cell QC metrics with cell barcodes as row
names. Column names correspond to the metrics produced by \verb{fragtk qc}.
}
\description{
Wrapper function to run \verb{fragtk qc} and add the metadata to the Seurat
object.
}
\concept{qc}
