% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{ATACqc}
\alias{ATACqc}
\alias{ATACqc.ChromatinAssay}
\alias{ATACqc.Seurat}
\title{Compute scATAC-seq QC metrics}
\usage{
ATACqc(object, ...)

\method{ATACqc}{ChromatinAssay}(
  object,
  fragtk.path = NULL,
  outdir = tempdir(),
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{ATACqc}{Seurat}(
  object,
  assay = NULL,
  fragtk.path = NULL,
  outdir = tempdir(),
  cleanup = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object or}

\item{...}{Arguments passed to other methods}

\item{fragtk.path}{Path to fragtk executable. If NULL, try to find fragtk
automatically.}

\item{outdir}{Path for output directory}

\item{cleanup}{Remove output files created by fragtk}

\item{verbose}{Display messages}

\item{assay}{Name of assay to use. If NULL, use the default assay.}
}
\description{
Wrapper function to run \code{fragtk qc} and add the metadata to the Seurat
object.
}
\concept{qc}
