% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{ATACqc}
\alias{ATACqc}
\alias{ATACqc.default}
\alias{ATACqc.ChromatinAssay}
\alias{ATACqc.Seurat}
\title{Compute scATAC-seq QC metrics}
\usage{
ATACqc(object, ...)

\method{ATACqc}{default}(
  object,
  annotations,
  fragtk.path = NULL,
  outdir = tempdir(),
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{ATACqc}{ChromatinAssay}(
  object,
  annotations = NULL,
  fragtk.path = NULL,
  outdir = tempdir(),
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{ATACqc}{Seurat}(
  object,
  assay = NULL,
  annotations = NULL,
  fragtk.path = NULL,
  outdir = tempdir(),
  cleanup = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object, ChromatinAssay object, or path to a fragment
file.}

\item{...}{Arguments passed to other methods}

\item{annotations}{\code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges()}} object containing
gene annotations. If NULL, attempt to extract this from the ChromatinAssay
object if provided.}

\item{fragtk.path}{Path to fragtk executable. If NULL, try to find fragtk
automatically.}

\item{outdir}{Path for output directory}

\item{cleanup}{Remove output files created by fragtk}

\item{verbose}{Display messages}

\item{assay}{Name of assay to use. If NULL, use the default assay.}
}
\description{
Wrapper function to run \verb{fragtk qc} and add the metadata to the Seurat
object.
}
\concept{qc}
