% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Bias}
\alias{Bias}
\alias{Bias.ChromatinAssay5}
\alias{Bias.Seurat}
\title{Bias}
\usage{
Bias(object, ...)

\method{Bias}{ChromatinAssay5}(object, ...)

\method{Bias}{Seurat}(object, assay = NULL, ...)
}
\arguments{
\item{object}{A Seurat, GRangesAssay, or ChromatinAssay5 object}

\item{...}{Arguments passed to other methods}

\item{assay}{Name of assay to use}
}
\value{
Returns a \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges()}} object
if the annotation data is present, otherwise returns NULL
}
\description{
Get the bias information from a ChromatinAssay5 object
}
\examples{
\donttest{
Bias(atac_small[["peaks"]])
}
\donttest{
Bias(atac_small)
}
}
\concept{assay}
