% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Bias}
\alias{Bias}
\alias{Bias<-}
\alias{Bias.ChromatinAssay5}
\alias{Bias.Seurat}
\alias{Bias<-.ChromatinAssay5}
\alias{Bias<-.Seurat}
\title{Bias}
\usage{
Bias(object, ...)

Bias(object, ...) <- value

\method{Bias}{ChromatinAssay5}(object, ...)

\method{Bias}{Seurat}(object, assay = NULL, ...)

\method{Bias}{ChromatinAssay5}(object, ...) <- value

\method{Bias}{Seurat}(object, assay = NULL, ...) <- value
}
\arguments{
\item{object}{A Seurat, GRangesAssay, or ChromatinAssay5 object}

\item{...}{Arguments passed to other methods}

\item{value}{A value to set. Can be \code{NULL}, to remove the current bias
information, or a numeric vector object.}

\item{assay}{Name of assay to use}
}
\value{
Returns a numeric vector or \code{NULL}
}
\description{
Get the bias information from a \linkS4class{ChromatinAssay5} object
}
\examples{
\donttest{
Bias(atac_small[["peaks"]])
}
\donttest{
Bias(atac_small)
}
bases <- c("A", "C", "G", "T")
hexamers <- apply(expand.grid(rep(list(bases), 6)), 1, paste0, collapse = "")
b <- 1:length(hexamers)
names(b) <- hexamers

# assign bias in a Seurat object
Bias(atac_small) <- b

# assign bias in a ChromatinAssay5 or GRangesAssay
Bias(atac_small[["peaks"]]) <- b
}
\concept{assay}
