% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blacklist_hg19}
\alias{blacklist_hg19}
\title{Genomic blacklist regions for Human hg19 (0-based)}
\format{
A GRanges object
}
\source{
\url{https://github.com/Boyle-Lab/Blacklist}

\doi{10.1038/s41598-019-45839-z}
}
\usage{
blacklist_hg19
}
\description{
Genomic blacklist regions for Human hg19 (0-based)
}
\concept{data}
\keyword{datasets}
