% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blacklist_hg38}
\alias{blacklist_hg38}
\title{Genomic blacklist regions for Human GRCh38}
\format{
A GRanges object
}
\source{
\url{https://github.com/Boyle-Lab/Blacklist}

\url{https://doi.org/10.1038/s41598-019-45839-z}
}
\usage{
blacklist_hg38
}
\description{
Genomic blacklist regions for Human GRCh38
}
\concept{data}
\keyword{datasets}
