% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blacklist_mm10}
\alias{blacklist_mm10}
\title{Genomic blacklist regions for Mouse mm10}
\format{A GRanges object}
\source{
https://sites.google.com/site/anshulkundaje/projects/blacklists
}
\usage{
blacklist_mm10
}
\description{
Genomic blacklist regions for Mouse mm10
}
\keyword{datasets}
