% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalculateCoverages}
\alias{CalculateCoverages}
\title{CalculateCoverages}
\usage{
CalculateCoverages(reads, window = 100)
}
\arguments{
\item{reads}{Dataframe containing reads for a region, generated by \code{GetReadsInRegion}}

\item{window}{Smoothing window to use}
}
\description{
Calculate normalized read coverage per base per cell group
}
