% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/peaks.R
\name{CallPeaks}
\alias{CallPeaks}
\alias{CallPeaks.Seurat}
\alias{CallPeaks.ChromatinAssay5}
\alias{CallPeaks.Fragment2}
\alias{CallPeaks.default}
\title{Call peaks}
\usage{
CallPeaks(object, ...)

\method{CallPeaks}{Seurat}(
  object,
  assay = NULL,
  group.by = NULL,
  idents = NULL,
  cells = NULL,
  macs3.path = NULL,
  mode = "callpeak",
  broad = FALSE,
  barcodes = NULL,
  genome = "hs",
  gsize = NULL,
  outdir = tempdir(),
  combine.peaks = TRUE,
  additional.args = NULL,
  name = Project(object),
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{CallPeaks}{ChromatinAssay5}(
  object,
  macs3.path = NULL,
  combine.peaks = TRUE,
  mode = "callpeak",
  outdir = tempdir(),
  broad = FALSE,
  barcodes = NULL,
  cells = NULL,
  genome = "hs",
  gsize = NULL,
  additional.args = NULL,
  name = "macs3",
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{CallPeaks}{Fragment2}(
  object,
  macs3.path = NULL,
  mode = "callpeak",
  outdir = tempdir(),
  broad = FALSE,
  barcodes = NULL,
  cells = NULL,
  genome = "hs",
  gsize = NULL,
  additional.args = NULL,
  name = "macs3",
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{CallPeaks}{default}(
  object,
  macs3.path = NULL,
  mode = "callpeak",
  outdir = tempdir(),
  broad = FALSE,
  barcodes = NULL,
  genome = "hs",
  gsize = NULL,
  additional.args = NULL,
  name = "macs3",
  cleanup = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \link[SeuratObject:Seurat-class]{SeuratObject::Seurat} object, \linkS4class{ChromatinAssay5}
object, \linkS4class{Fragment2} object, or the path to fragment file/s.}

\item{...}{Arguments passed to other methods.}

\item{assay}{Name of assay to use.}

\item{group.by}{Grouping variable to use. If set, peaks will be called
independently on each group of cells. If multiple fragments are present in
the object,  peaks will be called separately for each fragment file and for
each group of cells. Note that to call peaks using a subset of cells, we
write the cell barcodes into \code{.txt} file/s that are stored in the temp
directory (\code{\link[base:tempfile]{base::tempdir()}}) by default and removed upon completion (if
\code{cleanup=TRUE}). If \code{NULL}, peaks are called using all cells.}

\item{idents}{List of identities to include if grouping cells (only valid if
also setting the \code{group.by} parameter). If \code{NULL}, peaks will be called
for all cell identities.}

\item{cells}{Vector of cell barcodes to call peaks on.}

\item{macs3.path}{Path to MACS3 program. If \code{NULL}, try to find MACS3
automatically.}

\item{mode}{MACS function to call, choose between \code{callpeak} or \code{hmmratac}.
Default is \code{callpeak}.}

\item{broad}{Call broad peaks (\code{--broad} parameter for MACS).}

\item{barcodes}{Path to cell barcodes (\code{--barcodes} parameter for MACS).}

\item{genome}{MACS3 built-in effective genome size. Default is \code{hs} (Human,
GRCh38). Genome sizes for \code{mm} (Mice, GRCm38), \code{ce} (C. elegans, WBcel235),
and \code{dm} (Drosophila M., dm6) are also available.}

\item{gsize}{Manually set effective genome size parameter. If specified,
overrides MACS3 built-in genome sizes.}

\item{outdir}{Path for output files.}

\item{combine.peaks}{Controls whether peak calls from different groups of
cells are combined using \code{\link[GenomicRanges:inter-range-methods]{GenomicRanges::reduce()}} when calling peaks for
different groups of cells (\code{group.by} parameter). If \code{FALSE}, a list of
\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object will be returned. Note that metadata fields
such as the p-value, q-value, and fold-change information for each peak will
be lost if combining peaks.}

\item{additional.args}{Additional arguments passed to MACS. This should be a
single character string.}

\item{name}{Name for output MACS files. This will also be placed in the
\code{name} field in the GRanges output.}

\item{cleanup}{Remove MACS output files.}

\item{verbose}{Display messages.}
}
\value{
Returns a \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges()}} object
}
\description{
Call peaks using \code{MACS}. Fragment files linked to the specified assay will be
used to call peaks. If multiple fragment files are present, all will be used
in a single MACS invocation. Returns the \code{.narrowPeak} MACS output as a
\code{GRanges} object.
}
\details{
See \url{https://macs3-project.github.io/MACS/} for MACS documentation.

If you call peaks using MACS2 please cite:
\doi{10.1186/gb-2008-9-9-r137}
}
\concept{quantification}
