% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{CallPeaks}
\alias{CallPeaks}
\title{Call peaks}
\usage{
CallPeaks(
  object,
  assay = NULL,
  macs2.path = NULL,
  outpath = tempdir(),
  effective.genome.size = 2.7e+09,
  extsize = 200
)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Name of assay to use}

\item{macs2.path}{Path to MACS2 program. If NULL, try to find MACS2
automatically.}

\item{outpath}{Path for output files. If NULL, use temp directory and remove
files after peak calling.}

\item{effective.genome.size}{Effective genome size parameter for MACS2
(\code{-g}). Default is the human effective genome size (2.7e9).}

\item{extsize}{\code{extsize} parameter for MACS2.}
}
\value{
Returns a \code{\link[GenomicRanges]{GRanges}} object
}
\description{
Call peaks using MACS2
}
