% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/peaks.R
\name{CallPeaks}
\alias{CallPeaks}
\alias{CallPeaks.Seurat}
\alias{CallPeaks.ChromatinAssay}
\alias{CallPeaks.Fragment}
\alias{CallPeaks.default}
\title{Call peaks}
\usage{
CallPeaks(object, ...)

\method{CallPeaks}{Seurat}(
  object,
  assay = NULL,
  macs2.path = NULL,
  outdir = tempdir(),
  effective.genome.size = 2.7e+09,
  extsize = 200,
  shift = -extsize/2,
  additional.args = NULL,
  name = Project(object),
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{CallPeaks}{ChromatinAssay}(
  object,
  macs2.path = NULL,
  outdir = tempdir(),
  effective.genome.size = 2.7e+09,
  extsize = 200,
  shift = -extsize/2,
  additional.args = NULL,
  name = "macs2",
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{CallPeaks}{Fragment}(
  object,
  macs2.path = NULL,
  outdir = tempdir(),
  effective.genome.size = 2.7e+09,
  extsize = 200,
  shift = -extsize/2,
  additional.args = NULL,
  name = "macs2",
  cleanup = TRUE,
  verbose = TRUE,
  ...
)

\method{CallPeaks}{default}(
  object,
  macs2.path = NULL,
  outdir = tempdir(),
  effective.genome.size = 2.7e+09,
  extsize = 200,
  shift = -extsize/2,
  additional.args = NULL,
  name = "macs2",
  cleanup = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object, ChromatinAssay object, Fragment object, or the
path to fragment file/s.}

\item{...}{Arguments passed to other methods}

\item{assay}{Name of assay to use}

\item{macs2.path}{Path to MACS program. If NULL, try to find MACS
automatically.}

\item{outdir}{Path for output files}

\item{effective.genome.size}{Effective genome size parameter for MACS
(\code{-g}). Default is the human effective genome size (2.7e9).}

\item{extsize}{\code{extsize} parameter for MACS.}

\item{shift}{\code{shift} parameter for MACS.}

\item{additional.args}{Additional arguments passed to MACS. This should be a
single character string}

\item{name}{Name for output MACS files. This will also be placed in the
\code{name} field in the GRanges output.}

\item{cleanup}{Remove MACS output files}

\item{verbose}{Display messages}
}
\value{
Returns a \code{\link[GenomicRanges]{GRanges}} object
}
\description{
Call peaks using MACS. Fragment files linked to the specified assay will be
used to call peaks. If multiple fragment files are present, all will be used
in a single MACS invocation. Returns the \code{.narrowPeak} MACS output as a
\code{GRanges} object.
}
\details{
See \url{https://macs3-project.github.io/MACS/} for MACS documentation.

If you call peaks using MACS2 please cite:
\url{https://doi.org/10.1186/gb-2008-9-9-r137}
}
