% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Cells<-}
\alias{Cells<-}
\title{Set and get cell barcode information for a Fragment object}
\usage{
Cells(x, ...) <- value
}
\arguments{
\item{x}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{value}{A character vector of cell barcodes}
}
\description{
Set and get cell barcode information for a Fragment object
}
