% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{Cells.Fragment}
\alias{Cells.Fragment}
\alias{Cells<-.Fragment}
\title{Set and get cell barcode information for a \code{\link{Fragment}} object}
\usage{
\method{Cells}{Fragment}(x)

\method{Cells}{Fragment}(object, ...) <- value
}
\description{
Set and get cell barcode information for a \code{\link{Fragment}} object
}
