% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{ChromatinAssay-class}
\alias{ChromatinAssay-class}
\alias{ChromatinAssay}
\title{The ChromatinAssay class}
\description{
The ChramatinAssay object is an extended \code{\link[Seurat]{Assay}}
for the storage and analysis of chromatin-based single-cell data.
}
\section{Slots}{

\describe{
\item{\code{ranges}}{A \code{\link[GenomicRanges]{GRanges}} object describing the
genomic location of features in the object}

\item{\code{motifs}}{A \code{\link{Motif}} object}

\item{\code{fragments}}{A character vector containing the path/s to tabix-indexed fragment file/s
for the cells in the assay}

\item{\code{genome}}{Name of the genome used}

\item{\code{annotation}}{An object containing genomic annotations}

\item{\code{bias}}{A matrix containing Tn5 integration bias information (frequency of Tn5 integration at
different kmers)}

\item{\code{positionEnrichment}}{A named list of matrices containing positional enrichment scores for Tn5 integration
(for example, enrichment at the TSS)}
}}

