% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ClosestFeature}
\alias{ClosestFeature}
\title{ClosestFeature}
\usage{
ClosestFeature(regions, annotation, ...)
}
\arguments{
\item{regions}{A set of genomic regions to query}

\item{...}{Additional arguments passed to \code{\link{StringToGRanges}}}

\item{annotations}{Annotation information. Can be a GRanges object or an EnsDb object}
}
\value{
Returns a dataframe with the name of each region, the closest feature in the annotation,
and the distance to the feature.
}
\description{
Find the closest feature to a given set of genomic regions
}
