% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/motifs.R
\name{ClusterMotifs}
\alias{ClusterMotifs}
\alias{ClusterMotifs.Motif}
\alias{ClusterMotifs.Assay}
\alias{ClusterMotifs.Seurat}
\title{ClusterMotifs}
\usage{
ClusterMotifs(object, ...)

\method{ClusterMotifs}{Motif}(object, assay = NULL, verbose = TRUE,
  ...)

\method{ClusterMotifs}{Assay}(object, verbose = TRUE, ...)

\method{ClusterMotifs}{Seurat}(object, assay = NULL, verbose = TRUE,
  ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Additional arguments passed to \code{\link[Seurat]{FindNeighbors}}
and \code{\link[Seurat]{FindClusters}}}

\item{assay}{Which assay to use}

\item{verbose}{Display messages}
}
\value{
Returns a Seurat object
}
\description{
Cluster motifs by co-occurrence in genomic regions.
Computes the Jaccard similarity between motifs based on the proportion of
regions that they co-occur in. This is normalized for the overall frequency of the motifs.
Motifs are then clustered, and cluster identities stored in the Motif object meta data.
}
