% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CombineTracks}
\alias{CombineTracks}
\title{Combine genome region plots}
\usage{
CombineTracks(plotlist)
}
\arguments{
\item{plotlist}{A list of plots to combine. Must be from the same genomic
region.}
}
\value{
Returns a patchworked ggplot2 object
}
\description{
This can be used to combine coverage plots, peak region plots, gene
annotation plots, and linked element plots. The different tracks are stacked
on top of each other and the x-axis combined.
}
\concept{visualization}
