% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ConstructBinMatrix}
\alias{ConstructBinMatrix}
\title{ConstructBinMatrix}
\usage{
ConstructBinMatrix(fragments, genome, cells = NULL, binsize = 5000,
  chunk = 10000, verbose = TRUE)
}
\arguments{
\item{fragments}{Path to tabix-indexed fragments file}

\item{genome}{A GRanges object for the genome}

\item{cells}{Vector of cells to include. If NULL, include all cells found
in the fragments file}

\item{binsize}{Size of the genome bins to use}

\item{verbose}{Display messages}
}
\description{
Construct a bin x cell matrix from a fragments file
}
