% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{CountFragments}
\alias{CountFragments}
\title{Count fragments}
\usage{
CountFragments(fragments, cells = NULL)
}
\arguments{
\item{fragments}{Path to a fragment file}

\item{cells}{Cells to include. If NULL, include all cells}
}
\value{
Returns a data.frame
}
\description{
Count total fragments per cell barcode present in a fragment file.
}
\examples{
fpath <- system.file("extdata", "fragments.tsv.gz", package="Signac")
counts <- CountFragments(fragments = fpath)
}
\concept{fragments}
