% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CountsInRegion}
\alias{CountsInRegion}
\title{Counts in region}
\usage{
CountsInRegion(object, regions, assay = NULL, layer = NULL, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{regions}{A GRanges object}

\item{assay}{Name of a chromatin assay in the object to use. If NULL, use the
default assay}

\item{layer}{Name of layer to use. If NULL, use the default layer}

\item{...}{Additional arguments passed to \code{\link[IRanges]{findOverlaps}}}
}
\value{
Returns a numeric vector
}
\description{
Count reads per cell overlapping a given set of regions
}
\examples{
\donttest{
CountsInRegion(
  object = atac_small,
  assay = 'bins',
  regions = blacklist_hg19
)
}
}
\concept{utilities}
