% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CountsInRegion}
\alias{CountsInRegion}
\title{CountsInRegion}
\usage{
CountsInRegion(object, assay, regions, sep = c("-", "-"), ...)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Name of assay in the object to use}

\item{regions}{A GRanges object}

\item{sep}{Separator to use when extracting genomic coordinates from the Seurat object}

\item{...}{Additional arguments passed to \code{\link[IRanges]{findOverlaps}}}
}
\value{
Returns a numeric vector
}
\description{
Count reads per cell overlapping a given set of regions
}
\examples{
\donttest{
CountsInRegion(
  object = atac_small,
  assay = 'bins',
  regions = blacklist_hg19
)
}
}
