% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iranges-methods.R
\name{coverage,GRangesAssay-method}
\alias{coverage,GRangesAssay-method}
\alias{coverage}
\alias{coverage,Seurat-method}
\title{Coverage of a GRangesAssay object}
\usage{
\S4method{coverage}{GRangesAssay}(
  x,
  shift = 0L,
  width = NULL,
  weight = 1L,
  method = c("auto", "sort", "hash")
)

\S4method{coverage}{Seurat}(
  x,
  shift = 0L,
  width = NULL,
  weight = 1L,
  method = c("auto", "sort", "hash")
)
}
\arguments{
\item{x}{A \code{\link{GRangesAssay}} object}

\item{shift}{How much each range should be shifted before coverage is
computed. See \code{\link[IRanges]{coverage}} in the \pkg{IRanges} package.}

\item{width}{Specifies the length of the returned coverage vectors.
See \code{\link[IRanges]{coverage}} in the \pkg{IRanges} package.}

\item{weight}{Assigns weight to each range in \code{x}.
See \code{\link[IRanges]{coverage}} in the \pkg{IRanges} package.}

\item{method}{See \code{\link[IRanges]{coverage}} in the \pkg{IRanges}
package}
}
\description{
This is the \code{coverage} method for \code{\link{GRangesAssay}} objects.
}
\section{Functions}{
\itemize{
\item \code{coverage(GRangesAssay)}: method for GRangesAssay objects

\item \code{coverage(Seurat)}: method for Seurat objects

}}
\seealso{
\itemize{
  \item{\link[IRanges]{coverage-methods} in the \pkg{IRanges} package.}
  \item{\link[GenomicRanges]{coverage-methods} in the \pkg{GenomicRanges}
  package}
  \item{\link{GRangesAssay-class}}
 }
}
\concept{coverage}
