% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CoverageBrowser}
\alias{CoverageBrowser}
\title{Genome browser}
\usage{
CoverageBrowser(object, region, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{region}{A set of genomic coordinates}

\item{...}{Parameters passed to \code{\link{CoveragePlot}}}
}
\description{
Interactive version of the \code{\link{CoveragePlot}} function. Allows
altering the genome position interactively.
}
