% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CoverageBrowser}
\alias{CoverageBrowser}
\title{Genome browser}
\usage{
CoverageBrowser(object, region, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{region}{A set of genomic coordinates}

\item{...}{Parameters passed to \code{\link{CoveragePlot}}}
}
\value{
Returns a ggplot object
}
\description{
Interactive version of the \code{\link{CoveragePlot}} function. Allows
altering the genome position interactively. Upon ending the browser session,
the current view will be returned as a \code{\link[ggplot2]{ggplot}} object.
}
\concept{visualization}
