% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CoveragePlot}
\alias{CoveragePlot}
\title{CoveragePlot}
\usage{
CoveragePlot(object, region, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{region}{A set of genomic coordinates to show}

\item{...}{Arguments passed to SingleCoveragePlot}
}
\description{
Plot coverage within given region for groups of cells
}
