% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SingleCoveragePlot}
\alias{SingleCoveragePlot}
\alias{CoveragePlot}
\title{CoveragePlot}
\usage{
SingleCoveragePlot(object, region, annotation = NULL, assay = NULL,
  fragment.path = NULL, group.by = NULL, window = 100,
  downsample = 0.1, extend.upstream = 0, extend.downstream = 0,
  cells = NULL, idents = NULL, sep = c("-", "-"))

CoveragePlot(object, region, annotation = NULL, assay = NULL,
  fragment.path = NULL, group.by = NULL, window = 100,
  downsample = 0.1, extend.upstream = 0, extend.downstream = 0,
  cells = NULL, idents = NULL, sep = c("-", "-"), ...)
}
\arguments{
\item{object}{A Seurat object}

\item{region}{A set of genomic coordinates to show. Can be a GRanges object, a string, or a vector of strings describing the genomic
coordinates to plot.}

\item{annotation}{An Ensembl based annotation package}

\item{assay}{Name of the  assay to plot}

\item{fragment.path}{Path to an index fragment file. If NULL, will look for a path stored for the
requested assay using the \code{\link{SetFragments}} function}

\item{group.by}{Name of one or more metadata columns to group (color) the cells by. Default is the current cell identities}

\item{window}{Smoothing window size}

\item{downsample}{Fraction of positions to retain in the plot. Default is 0.1 (retain 10 percent, ie every 10th position)}

\item{extend.upstream}{Number of bases to extend the region upstream (Default 0)}

\item{extend.downstream}{Number of bases to extend the region downstream (Default 0)}

\item{cells}{Which cells to plot. Default all cells}

\item{idents}{Which identities to include in the plot. Default is all identities.}

\item{...}{Additional arguments passed to \code{\link[patchwork]{wrap_plots}}}
}
\description{
Plot coverage within given regions for groups of cells
}
\details{
Thanks to Andrew Hill for providing an early version of this function
 \url{http://andrewjohnhill.com/blog/2019/04/12/streamlining-scatac-seq-visualization-and-analysis/}
}
