% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CoveragePlot}
\alias{CoveragePlot}
\title{Plot coverage within given region for groups of cells}
\usage{
CoveragePlot(object, coords, annotation = NULL, assay = NULL,
  fragment.path = NULL, group.by = NULL, cells = NULL, cols = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{coords}{A set of genomic coordinates to show}

\item{annotation}{A set of genomic feature annotations. Must be the name of a set of features
stored using the \code{SetAnnotations} function}

\item{assay}{Which assay to plot}

\item{fragment.path}{Path to an index fragment file. If NULL, will look for a path stored for the
requested assay using the \code{SetFragments} function}

\item{group.by}{Name of one or more metadata columns to group (color) the cells by. Default is the current cell identities}

\item{cells}{Which cells to plot. Default all cells}

\item{cols}{Vector of colors, where each color corresponds to an identity class. By default, use the ggplot2 colors.}
}
\description{
Plot coverage within given region for groups of cells
}
