% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CreateMotifActivityMatrix}
\alias{CreateMotifActivityMatrix}
\title{CreateMotifActivityMatrix}
\usage{
CreateMotifActivityMatrix(object, assay = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Which assay to use. This must contain a Motif object.}
}
\value{
Returns a matrix
}
\description{
Create a matrix of the normalized motif accessibility per cell
}
