% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{CreateMotifMatrix}
\alias{CreateMotifMatrix}
\title{CreateMotifMatrix}
\usage{
CreateMotifMatrix(features, genome, pwm)
}
\arguments{
\item{features}{A set of genomic features. Must be specified as chr:start-end}

\item{genome}{A BSgenome}

\item{pwm}{A list of position weight matrices}

\item{min.score}{Minimum score to record a motif match for the feature}
}
\value{
Returns a sparse matrix
}
\description{
Create a motif x feature matrix from a set of genomic ranges,
the genome, and a set of position weight matrices
}
