% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateMotifObject}
\alias{CreateMotifObject}
\title{CreateMotifObject}
\usage{
CreateMotifObject(data = NULL, pwm = NULL, meta.data = NULL)
}
\arguments{
\item{data}{A motif x region matrix}

\item{pwm}{A list of position weight matrices matching the motif names in \code{data}.
Can be of class PFMatrixList}

\item{meta.data}{A data.frame containing metadata}
}
\description{
Create an object of class \code{Motif}
}
