% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateMotifObject}
\alias{CreateMotifObject}
\title{CreateMotifObject}
\usage{
CreateMotifObject(data, pwm = NULL, meta.data = NULL)
}
\arguments{
\item{data}{A}
}
\description{
Create an object of class \code{Motif}
}
