% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateMotifObject}
\alias{CreateMotifObject}
\title{CreateMotifObject}
\usage{
CreateMotifObject(data = NULL, pwm = NULL, neighbors = NULL,
  reductions = NULL, meta.data = NULL)
}
\arguments{
\item{data}{A motif x region matrix}

\item{pwm}{A list of position weight matrices matching the motif names in \code{data}.
Can be of class PFMatrixList}

\item{neighbors}{Neighbor data}

\item{reductions}{Dimension reduction data}

\item{meta.data}{A data.frame containing metadata}
}
\description{
Create an object of class \code{Motif}
}
\examples{
motif.matrix <- matrix(data = sample(c(0,1), size = 100, replace = TRUE), ncol = 5)
motif <- CreateMotifObject(data = motif.matrix)
}
