% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateRegionAggregationObject}
\alias{CreateRegionAggregationObject}
\title{Create a RegionAggregation object}
\usage{
CreateRegionAggregationObject(
  matrix = NULL,
  regions = NULL,
  upstream = NULL,
  downstream = NULL,
  name = NULL,
  expected = NULL,
  cells = NULL
)
}
\arguments{
\item{matrix}{A cell-by-position matrix}

\item{regions, }{A \code{\link[GenomicRanges:genomic-range-squeezers]{GenomicRanges::granges()}} object containing the
regions aggregated across.}

\item{upstream}{Integer denoting number of bases upstream of the
centered position that are stored in the matrix}

\item{downstream}{Integer denoting number of bases downstream of the
centered position that are stored in the matrix}

\item{name}{A name for the set of regions aggregated}

\item{expected}{A vector containing expected number of Tn5 insertions per position}

\item{cells}{A named vector of cells where each element is the cell barcode
and the name of each element is the corresponding cell barcode as stored
in the ChromatinAssay5 object}
}
\value{
Returns a \code{\link[=RegionAggregation]{RegionAggregation()}} object
}
\description{
Create a \code{\link[=RegionAggregation-class]{RegionAggregation-class()}} object to store the insertions over group of regions
}
\concept{regionaggregation}
