% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{CreateSignacObject}
\alias{CreateSignacObject}
\title{Initialize and setup the Signac object}
\usage{
CreateSignacObject(raw.data, type = "mtx", project = "UntitleProject",
  ...)
}
\arguments{
\item{raw.data}{Raw input data}

\item{type}{Type of the input file "mtx", "h5", "csv", "tsv", "sparse",
"matrix"}

\item{project}{Project name (string)}

\item{...}{parameters for h5 file, tsv file,}
}
\value{
Returns a Signac object with the raw data stored in object@raw.data.
object@data, object@meta.data, object@ident, also initialized.
}
\description{
Initializes the Signac object and some optional filtering
}
\dontrun{
pbmc_small <- CreateSignacObject(raw.data = path2pbmc, type = "mtx")
pbmc_small

}
