% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DepthCor}
\alias{DepthCor}
\title{Sequencing depth correlation}
\usage{
DepthCor(object, reduction = "lsi", assay = "peaks", n = 10, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{reduction}{Name of a dimension reduction stored in the
input object}

\item{assay}{Name of assay to use for sequencing depth}

\item{n}{Number of components to use. If NULL, use all components.}

\item{...}{Additional arguments passed to \code{\link[base]{cor}}}
}
\value{
Returns a ggplot object
}
\description{
Compute the correlation between total counts and each reduced
dimension component.
}
\examples{
DepthCor(object = atac_small)
}
