% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{EnrichedTerms}
\alias{EnrichedTerms}
\title{Find enriched ontology terms}
\usage{
EnrichedTerms(
  object,
  terms,
  group.by = NULL,
  assay = NULL,
  var.features = TRUE,
  scoreType = "std",
  top.n = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{terms}{Ontology term list. The name of each element in the list should
be an ontology term, and the list elements a vector of feature names present
in the assay.}

\item{group.by}{Name of grouping variable to use. If \code{NULL}, use the active
cell identities.}

\item{assay}{Name of assay to use. If \code{NULL}, use the default assay.}

\item{var.features}{Subset to only variable features for ontology term
enrichment.}

\item{scoreType}{\code{scoreType} parameter for \code{\link[fgsea:fgseaSimple]{fgsea::fgseaSimple()}}.
Options are "std", "pos", "neg" (two-tailed or one-tailed tests).}

\item{top.n}{Number of top enriched terms to retain for each set of cells. If
NULL, retain all terms.}

\item{verbose}{Display messages.}

\item{...}{Additional arguments passed to \code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}}}
}
\value{
Returns a named list of dataframes. Each element of the list contains
a dataframe with the term enrichment results for an identity class.
}
\description{
Run ontology term enrichment testing on differential features for each of the
identity classes. Runs \code{\link[fgsea:fgsea]{fgsea::fgsea()}} on differential testing
results from each identity class.
}
\concept{utilities}
