% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwig.R
\name{ExportGroupBW}
\alias{ExportGroupBW}
\title{Export bigwig files for groups of cells}
\usage{
ExportGroupBW(
  object,
  assay = NULL,
  group.by = NULL,
  idents = NULL,
  normMethod = "RC",
  tileSize = 100,
  minCells = 5,
  cutoff = NULL,
  chromosome = NULL,
  outdir = getwd(),
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Name of assay to use}

\item{group.by}{The metadata variable used to group the cells}

\item{idents}{Identities to include (defined by group.by parameter)}

\item{normMethod}{Normalization method for the bigwig files. Deafult 'RC'.
'RC' will divide the number of fragments in a tile by the total number of
fragments in the group. A scaling factor of 10^4 will be applied. 
'ncells' will divide the number of fragments in a tile by the number of cell
in the group. 'none' will apply no normalization method. 
A vector of values for each cell can also be passed as a metadata column
name. A scaling factor of 10^4 will be applied}

\item{tileSize}{The size of the tiles in the bigwig file}

\item{minCells}{The minimum of cells in a group to be exported}

\item{cutoff}{The maximum number of fragments in a given genomic tile}

\item{chromosome}{A chromosomes vector to be exported}

\item{outdir}{Directory to write output files (splitted bed and bigwigs)}

\item{verbose}{Display messages}
}
\description{
Export bigwig files for groups of cells
}
\examples{
\dontrun{
ExportGroupBW(object, assay = "peaks")
}
}
