% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Extend}
\alias{Extend}
\title{Extend}
\usage{
Extend(x, upstream = 0, downstream = 0, from.midpoint = FALSE)
}
\arguments{
\item{x}{A range}

\item{upstream}{Length to extend upstream}

\item{downstream}{Length to extend downstream}

\item{from.midpoint}{Extend from the midpoint of the region, rather than the
start and end coordinates (default FALSE).}
}
\description{
Resize GenomicRanges upstream and or downstream.
From \url{https://support.bioconductor.org/p/78652/}
}
\examples{
Extend(x = blacklist_hg19, upstream = 100, downstream = 100)
}
