% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastGetSubSparseMat}
\alias{FastGetSubSparseMat}
\title{FastGetSubSparseMat}
\usage{
FastGetSubSparseMat(mat, rrvec, ccvec)
}
\arguments{
\item{mat}{A sparse matrix}

\item{rrvec}{A row vector}

\item{ccvec}{A col vector}
}
\description{
Get subview of a sparse matrix
}
