% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FeatureMatrix}
\alias{FeatureMatrix}
\title{FeatureMatrix}
\usage{
FeatureMatrix(fragments, features, cells = NULL, chunk = 50,
  sep = c("-", "-"), verbose = TRUE)
}
\arguments{
\item{fragments}{Path to tabix-indexed fragments file}

\item{cells}{Vector of cells to include. If NULL, include all cells found
in the fragments file}

\item{chunk}{Number of chunks to use when processing the fragments file. Fewer chunks may enable faster processing,
but will use more memory.}

\item{sep}{Vector of separators to use for genomic string. First element is used to separate chromosome
and coordinates, second separator is used to separate start and end coordinates.}

\item{verbose}{Display messages}

\item{genome}{A GRanges object for the genome}

\item{binsize}{Size of the genome bins to use}
}
\description{
Construct a feature x cell matrix from a genomic fragments file
}
