% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FilterFragments}
\alias{FilterFragments}
\title{FilterFragments}
\usage{
FilterFragments(fragment.path, cells, output.path, chunk, verbose)
}
\arguments{
\item{fragment.path}{Path to a tabix-indexed fragments file}

\item{cells}{A vector of cells to retain}

\item{output.path}{Name and path for output tabix file. A tabix index file will also be created in the same location, with
the .tbi file extension.}

\item{chunk}{Number of chunks to use when processing the fragments. Fewer chunks may be faster, but will use more memory.}

\item{verbose}{Display messages}
}
\description{
Remove cells from a fragments file that are not present in a given list of cells.
}
