% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichments.R
\name{FindMotifs}
\alias{FindMotifs}
\title{FindMotifs}
\usage{
FindMotifs(object, cells = NULL, features = NULL, assay = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{cells}{A vector of cells to test}

\item{assay}{Which assay to use. Default is the active assay}
}
\value{
Returns a data.frame
}
\description{
Find motifs enriched in a given set of peaks
}
