% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{FindMotifs}
\alias{FindMotifs}
\title{FindMotifs}
\usage{
FindMotifs(object, features, assay = NULL, background = NULL,
  verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{A vector of features to test for enrichments over background}

\item{assay}{Which assay to use. Default is the active assay}

\item{background}{Vector of features to use as the background set. If NULL, use all features in the assay.}

\item{verbose}{Display messages}
}
\value{
Returns a data frame
}
\description{
Find motifs overrepresented in a given set of genomic features. Computes the number of features
containing the motif (observed) and compares this to the total number of features containing the
motif (background) using the hypergeometric test.
}
