% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{FitMeanVar}
\alias{FitMeanVar}
\alias{FitMeanVar.Seurat}
\alias{FitMeanVar.Assay5}
\alias{FitMeanVar.default}
\alias{FitMeanVar.data.frame}
\title{Find variable features fitting LOESS}
\usage{
FitMeanVar(object, ...)

\method{FitMeanVar}{Seurat}(
  object,
  assay = NULL,
  layer = NULL,
  nfeatures = 20000,
  loess.span = 0.1,
  min.cutoff = 10,
  weight.mean = 0.5,
  bins = 1000,
  sample_per_bin = 50,
  key = "dsLoess",
  verbose = FALSE,
  ...
)

\method{FitMeanVar}{Assay5}(
  object,
  layer = NULL,
  loess.span = 0.1,
  nfeatures = 20000,
  min.cutoff = 10,
  weight.mean = 0,
  bins = 1000,
  sample_per_bin = 50,
  key = "dsLoess",
  verbose = FALSE,
  ...
)

\method{FitMeanVar}{default}(
  object,
  nfeatures = 20000,
  loess.span = 0.1,
  min.cutoff = 10,
  weight.mean = 0,
  bins = 1000,
  sample_per_bin = 50,
  random.seed = 1234,
  verbose = FALSE,
  ...
)

\method{FitMeanVar}{data.frame}(
  object,
  loess.span = 0.1,
  weight.mean = 0,
  bins = 1000,
  sample_per_bin = 50,
  random.seed = 1234,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{assay}{Name of assay to use. If NULL, use the default assay.}

\item{layer}{Name of layer to use. If NULL, use the default layer(s).}

\item{nfeatures}{Number of features to selected as top variable features.}

\item{loess.span}{\code{span} parameter passed to the \code{\link[stats]{loess}} function}

\item{min.cutoff}{Minimum number of counts for a feature to be eligible for variable feature selection.}

\item{weight.mean}{How much to weight the ranking of features according to their mean.
Setting \code{weight.mean=0} will rank features according to their residual variance only.}

\item{bins}{Number of bins to use when downsampling features across the range of mean count values.}

\item{sample_per_bin}{Number of features to select per mean count bin in feature downsampling step.}

\item{key}{Key to use when storing the highly variable feature information in the assay.}

\item{verbose}{Display messages.}

\item{random.seed}{Random seed to set for sampling.}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Find highly variable features by fitting a locally polynomial regression
model to the log(mean) and log(variance) of downsampled features.
}
\details{
This function is similar to the \code{\link[Seurat]{FindVariableFeatures}}
function (with \code{selection.method="vst"}), but downsamples the features
evenly across the range of mean values. This speeds up fitting of the loess 
curve when the number of features is large.

The function also provides the ability to combine ranking of features
according to their mean count and their residual variance, using a weighted
rank sum with weights set by the \code{weight.mean} parameter. This can help
to avoid selecting features with high residual variance but very low mean.
}
\examples{
\dontrun{
FitMeanVar(atac_small)
}
}
\concept{preprocessing}
