% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/footprinting.R
\name{Footprint}
\alias{Footprint}
\alias{Footprint.ChromatinAssay}
\alias{Footprint.Seurat}
\title{Transcription factor footprinting analysis}
\usage{
Footprint(object, ...)

\method{Footprint}{ChromatinAssay}(
  object,
  genome,
  motif.name = NULL,
  regions = NULL,
  group.by = NULL,
  idents = NULL,
  assay = NULL,
  upstream = 250,
  downstream = 250,
  verbose = TRUE,
  ...
)

\method{Footprint}{Seurat}(
  object,
  genome,
  regions = NULL,
  motif.name = NULL,
  group.by = NULL,
  idents = NULL,
  assay = NULL,
  upstream = 250,
  downstream = 250,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat or ChromatinAssay object}

\item{...}{Arguments passed to other methods}

\item{genome}{A \code{\link[BSgenome]{BSgenome}} object}

\item{motif.name}{Name of a motif stored in the assay to footprint. If not
supplied, must supply a set of regions.}

\item{regions}{A set of genomic ranges containing the motif instances}

\item{group.by}{Grouping variable for the cells}

\item{idents}{Which identities to include}

\item{assay}{Name of assay to use}

\item{upstream}{Number of bases to extend upstream}

\item{downstream}{Number of bases to extend downstream}

\item{verbose}{Display messages}
}
\value{
Returns a \code{\link[Seurat]{Seurat}} object
}
\description{
Compute the normalized observed/expected Tn5 insertion frequency
for each position surrounding a set of motif instances.
}
