% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprinting.R
\name{Footprint}
\alias{Footprint}
\title{Transcription factor footprinting analysis}
\usage{
Footprint(object, regions, genome, group.by = NULL, idents = NULL,
  assay = NULL, upstream = 250, downstream = 250, verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{regions}{A set of GRanges containing the motif instances}

\item{genome}{A BSgenome object}

\item{group.by}{Grouping variable for the cells}

\item{idents}{Which identities to include}

\item{assay}{Name of assay to use}

\item{upstream}{Number of bases to extend upstream}

\item{downstream}{Number of bases to extend downstream}

\item{verbose}{Display messages}
}
\description{
Compute the normalized observed/expected Tn5 insertion frequency
for each position surrounding a set of motif instances.
}
