% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Fragment-class}
\alias{Fragment-class}
\alias{Fragment}
\title{The Fragment class}
\description{
The Fragment class
}
\section{Slots}{

\describe{
\item{\code{path}}{Path to the fragment file on disk.
See \url{https://support.10xgenomics.com/single-cell-atac/software/
pipelines/latest/output/fragments}}

\item{\code{hash}}{A vector of two md5sums: first element is the md5sum of the
fragment file, the second element is the md5sum of the index.}

\item{\code{cells}}{A character vector containing the cell barcodes present in the
fragment file}

\item{\code{prefix}}{A string to append to the beginning of the cell barcodes
in the fragment file for them to match the barcodes in the Seurat object.}

\item{\code{suffix}}{A string to append to the end of the cell barcodes
in the fragment file for them to match the barcodes in the Seurat object.}
}}

