% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Fragment-class}
\alias{Fragment-class}
\alias{Fragment}
\title{The Fragment class}
\description{
The Fragment class is designed to hold information needed for working with
fragment files.
}
\section{Slots}{

\describe{
\item{\code{path}}{Path to the fragment file on disk.
See \url{https://support.10xgenomics.com/single-cell-atac/software/pipelines/latest/output/fragments}}

\item{\code{hash}}{A vector of two md5sums: first element is the md5sum of the
fragment file, the second element is the md5sum of the index.}

\item{\code{cells}}{A named vector of cells where each element is the cell barcode
as it appears in the fragment file, and the name of each element is the
corresponding cell barcode as stored in the ChromatinAssay object.}
}}

\concept{fragments}
