% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Fragments}
\alias{Fragments}
\alias{Fragments<-}
\alias{Fragments.ChromatinAssay5}
\alias{Fragments.Seurat}
\alias{Fragments<-.ChromatinAssay5}
\alias{Fragments<-.Seurat}
\title{Get the Fragment objects}
\usage{
Fragments(object, ...)

Fragments(object, ...) <- value

\method{Fragments}{ChromatinAssay5}(object, ...)

\method{Fragments}{Seurat}(object, assay = NULL, ...)

\method{Fragments}{ChromatinAssay5}(object, ...) <- value

\method{Fragments}{Seurat}(object, assay = NULL, ...) <- value
}
\arguments{
\item{object}{A Seurat, GRangesAssay, or ChromatinAssay5 object}

\item{...}{Arguments passed to other methods}

\item{value}{A \link[=Fragment2-class]{Fragment2} object or list of Fragment objects}

\item{assay}{Name of assay to use}
}
\value{
Returns a list of \link[=Fragment2-class]{Fragment2} objects. If there are
no \link[=Fragment2-class]{Fragment2} objects present, returns an empty list.
}
\description{
Get the Fragment objects
}
\examples{
Fragments(atac_small[["peaks"]])
Fragments(atac_small)
fpath <- system.file("extdata", "fragments.tsv.gz", package = "Signac")
fragments <- CreateFragmentObject(
  path = fpath,
  cells = colnames(atac_small),
  validate.fragments = FALSE
)
# Seurat object method
Fragments(atac_small) <- fragments

# Remove fragment objects
Fragments(atac_small) <- NULL

# Assay method
Fragments(atac_small[["peaks"]]) <- fragments
}
\concept{assay}
\concept{fragments}
