% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{FRiP}
\alias{FRiP}
\title{Calculate approximate fraction of reads in peaks per cell}
\usage{
FRiP(object, peak.assay, bin.assay, verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{peak.assay}{Name of the assay containing a peak x cell matrix}

\item{bin.assay}{Name of the assay containing a bin x cell matrix}

\item{verbose}{Display messages}
}
\description{
Calculate approximate fraction of reads in peaks per cell
}
