% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GenomeBinMatrix}
\alias{GenomeBinMatrix}
\title{GenomeBinMatrix}
\usage{
GenomeBinMatrix(fragments, genome, cells = NULL, binsize = 5000,
  chunk = 50, sep = c("-", "-"), verbose = TRUE)
}
\arguments{
\item{fragments}{Path to tabix-indexed fragments file}

\item{genome}{A GRanges object for the genome}

\item{cells}{Vector of cells to include. If NULL, include all cells found
in the fragments file}

\item{binsize}{Size of the genome bins to use}

\item{chunk}{Number of chunks to use when processing the fragments file. Fewer chunks may enable faster processing,
but will use more memory.}

\item{sep}{Vector of separators to use for genomic string. First element is used to separate chromosome
and coordinates, second separator is used to separate start and end coordinates.}

\item{verbose}{Display messages}
}
\description{
Construct a bin x cell matrix from a fragments file.
}
\details{
This function bins the genome and calls \code{\link{FeatureMatrix}} to
construct a bin x cell matrix.
}
