% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetCellsInBin}
\alias{GetCellsInBin}
\title{GetCellsInBin}
\usage{
GetCellsInBin(fragments, region, bins, cells = NULL, ...)
}
\arguments{
\item{fragments}{Tabix-indexed fragments file}

\item{region}{A string giving the region to extract from the fragments file}

\item{bins}{A GRanges object containing the full set of genome bins}

\item{cells}{Vector of cells to include in output. If NULL, include all cells}

\item{...}{Additional arguments passed to GRangesToString}
}
\description{
Extract cell names containing reads mapped within a given genomic region
}
