% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetCellsInRegion}
\alias{GetCellsInRegion}
\title{GetCellsInRegion}
\usage{
GetCellsInRegion(tabix, region, sep = c("-", "-"), cells = NULL)
}
\arguments{
\item{tabix}{Tabix object}

\item{region}{A string giving the region to extract from the fragments file}

\item{cells}{Vector of cells to include in output. If NULL, include all cells}
}
\description{
Extract cell names containing reads mapped within a given genomic region
}
