% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetFragmentData}
\alias{GetFragmentData}
\title{Get Fragment object data}
\usage{
GetFragmentData(object, slot = "file.path")
}
\arguments{
\item{object}{A \linkS4class{Fragment2} object}

\item{slot}{Information to pull from object
(file.path, index.path, hash, cells, seqlevels)}
}
\description{
Extract data from a \linkS4class{Fragment2} object
}
\concept{assay}
