% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{GetFragmentData}
\alias{GetFragmentData}
\title{Get Fragment object data}
\usage{
GetFragmentData(object, slot = "path")
}
\arguments{
\item{object}{A \code{\link{Fragment}} object}

\item{slot}{Information to pull from object (path, hash, cells, prefix, suffix)}
}
\description{
Extract data from a \code{\link{Fragment-class}} object
}
\concept{assay}
