% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{GetLinkedGenes}
\alias{GetLinkedGenes}
\title{Get genes linked to peaks}
\usage{
GetLinkedGenes(object, features, assay = NULL, min.abs.score = 0)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{A list of peaks to find linked genes for}

\item{assay}{Name of assay to use. If NULL, use the default assay}

\item{min.abs.score}{Minimum absolute value of the link score for a link to
be returned}
}
\description{
Find genes linked to a given set of peaks
}
\seealso{
GetLinkedPeaks
}
\concept{links}
