% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{GetLinkedPeaks}
\alias{GetLinkedPeaks}
\title{Get peaks linked to genes}
\usage{
GetLinkedPeaks(object, features, assay = NULL, min.abs.score = 0)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{A list of genes to find linked peaks for}

\item{assay}{Name of assay to use. If NULL, use the default assay}

\item{min.abs.score}{Minimum absolute value of the link score for a link to
be returned}
}
\description{
Find peaks linked to a given set of genes
}
\seealso{
GetLinkedGenes
}
\concept{links}
