% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{GetMotifData}
\alias{GetMotifData}
\alias{GetMotifData.Assay}
\alias{GetMotifData.Seurat}
\title{GetMotifData}
\usage{
GetMotifData(object, ...)

\method{GetMotifData}{Assay}(object, ...)

\method{GetMotifData}{Seurat}(object, assay = NULL, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{assay}{Which assay to use. Default is the current active assay}
}
\value{
Returns a Seurat object
}
\description{
Get motif matrix for given assay
}
