% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetReadsInRegion}
\alias{GetReadsInRegion}
\title{GetReadsInRegion}
\usage{
GetReadsInRegion(object, region, assay = NULL, fragment.path = NULL,
  group.by = NULL, cells = NULL, verbose = TRUE, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{region}{A genomic region, specified as a string in the format
'chr:start-end'. Can be a vector of regions.}

\item{assay}{Name of assay to use}

\item{fragment.path}{Path to indexed fragment file}

\item{group.by}{Cell grouping information to add}

\item{cells}{Cells to include. Default is all cells present in the object.}

\item{verbose}{Display messages}

\item{...}{Additional arguments passed to \code{\link{StringToGRanges}}}
}
\value{
Returns a data frame
}
\description{
Extract reads for each cell within a given genomic region or set of regions
}
