% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges-methods.R
\name{granges-methods}
\alias{granges-methods}
\alias{granges}
\alias{granges,ChromatinAssay-method}
\alias{granges,Seurat-method}
\title{Access genomic ranges for ChromatinAssay objects}
\usage{
\S4method{granges}{ChromatinAssay}(x, use.names = TRUE, use.mcols = FALSE, ...)

\S4method{granges}{Seurat}(x, use.names = TRUE, use.mcols = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{ChromatinAssay}} object}

\item{use.names}{Whether the names on the genomic ranges should be
propagated to the returned object.}

\item{use.mcols}{Not supported for \code{\link{ChromatinAssay}} objects}

\item{...}{Additional arguments}
}
\value{
Returns a \code{\link[GenomicRanges]{GRanges}} object
}
\description{
Methods for accessing \code{\link[GenomicRanges]{GRanges}} object
information stored in a \code{\link{ChromatinAssay}} object.
}
\section{Functions}{
\itemize{
\item \code{granges,Seurat-method}: method for Seurat objects
}}

\examples{
granges(atac_small)
}
\seealso{
\itemize{
  \item{\link[GenomicRanges]{granges} in the \pkg{GenomicRanges} package.}
  \item{\link{ChromatinAssay-class}}
 }
}
\concept{granges}
