% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ChromatinAssay5-class}
\alias{ChromatinAssay5-class}
\alias{GRangesAssay-class}
\title{ChromatinAssay5 and GRangesAssay object classes}
\description{
The \code{GRangesAssay} and \code{ChromatinAssay5} classes are extended
\code{\link[SeuratObject:Assay5]{SeuratObject::Assay5()}} classes for the storage and analysis of
single-cell chromatin data. The \code{GRangesAssay} class requires that
features in the assay are genomic ranges.
}
\section{Slots}{

\describe{
\item{\code{fragments}}{A list of \code{\link[=Fragment]{Fragment()}} objects. Each cell can be linked to
at most one fragment file, so a cell can appear in only one of these
objects. A fragment object holding no cell information is assumed to contain
every cell in the assay, and so cannot be stored alongside any other
fragment object.}

\item{\code{annotation}}{A  \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} object containing
genomic annotations. This should be a GRanges object with the following
columns:
\itemize{
\item{tx_id: Transcript ID}
\item{gene_name: Gene name}
\item{gene_id: Gene ID}
\item{gene_biotype: Gene biotype (e.g. "protein_coding", "lincRNA")}
\item{type: Annotation type (e.g. "exon", "gap")}
}}

\item{\code{bias}}{A vector containing Tn5 integration bias information
(frequency of Tn5 integration at different kmers). This must be a named
numeric vector where the name of the element corresponds to the DNA sequence
and the value represents the bias value. All DNA hexamers must be present in
the vector.}

\item{\code{region.aggregation}}{A list of \link[=RegionAggregation-class]{RegionAggregation} objects}

\item{\code{motifs}}{A \link[=Motif-class]{Motif} object}

\item{\code{links}}{A list of \code{\link[InteractionSet:GInteractions]{InteractionSet::GInteractions()}} objects
describing linked genomic positions, such as co-accessible sites, eQTLs,
Hi-C contact, or enhancer-gene regulatory relationships.}

\item{\code{ranges}}{A \code{\link[GenomicRanges:GRanges]{GenomicRanges::GRanges()}} object describing the
genomic location of features in the object}
}}

\seealso{
\code{\link[SeuratObject:Assay5]{SeuratObject::Assay5()}}
}
\concept{assay}
