% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{GRangesAssay-class}
\alias{GRangesAssay-class}
\alias{GRangesAssay}
\title{The GRangesAssay class}
\description{
The GRangesAssay object is an extended \code{\link{ChromatinAssay5}}
for the storage and analysis of single-cell chromatin data with associated
genomic ranges.
}
\section{Slots}{

\describe{
\item{\code{ranges}}{A \code{\link[GenomicRanges]{GRanges}} object describing the
genomic location of features in the object}

\item{\code{motifs}}{A \code{\link{Motif}} object}

\item{\code{links}}{A \code{\link[GenomicRanges]{GRanges}} object describing linked
genomic positions, such as co-accessible sites or enhancer-gene regulatory
relationships. This should be a \code{GRanges} object, where the start and
end coordinates are the two linked genomic positions, and must contain a
"score" metadata column.}
}}

\concept{assay}
