% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ChromatinAssay5-class}
\alias{ChromatinAssay5-class}
\alias{ChromatinAssay5}
\alias{GRangesAssay-class}
\alias{GRangesAssay}
\title{ChromatinAssay5 and GRangesAssay object classes}
\description{
The \code{GRangesAssay} and \code{ChromatinAssay5} classes are extended
\code{\link[SeuratObject]{Assay5}} classes for the storage and analysis of
single-cell chromatin data. The \code{GRangesAssay} class requires that
features in the assay are genomic ranges.
}
\section{Slots}{

\describe{
\item{\code{fragments}}{A list of \code{\link{Fragment}} objects.}

\item{\code{annotation}}{A  \code{\link[GenomicRanges]{GRanges}} object containing
genomic annotations. This should be a GRanges object with the following 
columns:
\itemize{
  \item{tx_id: Transcript ID}
  \item{gene_name: Gene name}
  \item{gene_id: Gene ID}
  \item{gene_biotype: Gene biotype (e.g. "protein_coding", "lincRNA")}
  \item{type: Annotation type (e.g. "exon", "gap")}
}}

\item{\code{bias}}{A vector containing Tn5 integration bias information
(frequency of Tn5 integration at different kmers)}

\item{\code{cell.aggregation}}{A list of \code{\link{CellAggregation}} objects}

\item{\code{region.aggregation}}{A list of \code{\link{RegionAggregation}} objects}

\item{\code{motifs}}{A \code{\link{Motif}} object}

\item{\code{links}}{A list of \code{\link[InteractionSet]{GInteractions}} objects
describing linked genomic positions, such as co-accessible sites, eQTLs, 
Hi-C contact, or enhancer-gene regulatory relationships.}

\item{\code{ranges}}{A \code{\link[GenomicRanges]{GRanges}} object describing the
genomic location of features in the object}
}}

\seealso{
\code{\link[SeuratObject]{Assay5}}
}
\concept{assay}
