% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GRangesToString}
\alias{GRangesToString}
\title{GRanges to String}
\usage{
GRangesToString(grange, sep = c("-", "-"))
}
\arguments{
\item{grange}{A GRanges object}

\item{sep}{Vector of separators to use for genomic string. First element is
used to separate chromosome and coordinates, second separator is used to
separate start and end coordinates.}
}
\value{
Returns a character vector
}
\description{
Convert GRanges object to a vector of strings
}
\examples{
GRangesToString(grange = blacklist_hg19)
}
\concept{utilities}
