% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{InsertionBias}
\alias{InsertionBias}
\title{Compute Tn5 insertion bias}
\usage{
InsertionBias(
  object,
  genome,
  assay = NULL,
  region = "chr1-1-249250621",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{genome}{A BSgenome object}

\item{assay}{Name of assay to use. If NULL, use the default assay.}

\item{region}{Region to use when assessing bias. Default is human chromosome 1.}

\item{verbose}{Display messages}

\item{...}{Additional arguments passed to \code{\link{StringToGRanges}}}
}
\value{
Returns a Seurat object
}
\description{
Counts the Tn5 insertion frequency for each DNA hexamer.
}
