% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{LinkPeaks}
\alias{LinkPeaks}
\title{Link peaks to genes}
\usage{
LinkPeaks(
  object,
  peak.assay,
  expression.assay,
  peak.layer = "counts",
  expression.layer = "data",
  method = "pearson",
  key = "linkpeaks",
  gene.coords = NULL,
  distance = 5e+05,
  min.distance = NULL,
  min.cells = 10,
  genes.use = NULL,
  n_sample = 200,
  pvalue_cutoff = 0.05,
  score_cutoff = 0.05,
  gene.id = FALSE,
  verbose = TRUE,
  peak.slot = deprecated(),
  expression.slot = deprecated()
)
}
\arguments{
\item{object}{A Seurat object}

\item{peak.assay}{Name of assay containing peak information}

\item{expression.assay}{Name of assay containing gene expression information}

\item{peak.layer}{Name of layer to pull chromatin data from}

\item{expression.layer}{Name of layer to pull expression data from}

\item{method}{Correlation method to use. One of "pearson" or "spearman"}

\item{key}{Key to use when storing link information in the assay}

\item{gene.coords}{GRanges object containing coordinates of genes in the
expression assay. If NULL, extract from gene annotations stored in the assay.}

\item{distance}{Distance threshold for peaks to include in regression model}

\item{min.distance}{Minimum distance between peak and TSS to include in
regression model. If NULL (default), no minimum distance is used.}

\item{min.cells}{Minimum number of cells positive for the peak and gene
needed to include in the results.}

\item{genes.use}{Genes to test. If NULL, determine from expression assay.}

\item{n_sample}{Number of peaks to sample at random when computing the null
distribution.}

\item{pvalue_cutoff}{Minimum p-value required to retain a link. Links with a
p-value equal or greater than this value will be removed from the output.}

\item{score_cutoff}{Minimum absolute value correlation coefficient for a link
to be retained}

\item{gene.id}{Set to TRUE if genes in the expression assay are named
using gene IDs rather than gene names.}

\item{verbose}{Display messages}

\item{peak.slot}{Deprecated (use \code{peak.layer})}

\item{expression.slot}{Deprecated (used \code{expression.layer})}
}
\value{
Returns a Seurat object with results added to the \code{links} slot in the
assay, stored under the key specified in the function. The results are stored
as an \link[InteractionSet:GInteractions-class]{InteractionSet::GInteractions} object accessible via the \code{\link[=Links]{Links()}}
function. This contains the \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} for the pair of linked
regions (peak and gene), with \code{anchor1} corresponding to the peak region and
\code{anchor2} corresponding to the gene region linked to the peak. The following
metadata is also stored in the \code{GInteractions} object:
\itemize{
\item \code{anchor2.gene_id}: The gene ID for the linked gene
\item \code{anchor2.gene_name}: The name of the linked gene
\item \code{score}: the correlation coefficient between the accessibility of the
peak and expression of the gene
\item \code{zscore}: the z-score of the correlation coefficient, computed based on
the distribution of correlation coefficients from a set of background peaks
\item \code{pvalue}: the p-value associated with the z-score for the link
}
}
\description{
Find peaks that are correlated with the expression of nearby genes.
For each gene, this function computes the correlation coefficient between
the gene expression and accessibility of each peak within a given distance
from the gene TSS, and computes an expected correlation coefficient for each
peak given the GC content, accessibility, and length of the peak. The expected
coefficient values for the peak are then used to compute a z-score and p-value.
}
\details{
This function was inspired by the method originally described by SHARE-seq
(Sai Ma et al. 2020, Cell). Please consider citing the original SHARE-seq
work if using this function: \href{https://pubmed.ncbi.nlm.nih.gov/33098772/}{doi: 10.1016/j.cell.2020.09.056}
}
\concept{links}
