% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{LinkPeaks}
\alias{LinkPeaks}
\title{Link peaks to genes}
\usage{
LinkPeaks(object, peak.assay, expression.assay, gene.coords, sep = c("-",
  "-"), binary = TRUE, alpha = 0.95, distance = 2e+05,
  verbose = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{peak.assay}{Name of assay containing peak information}

\item{expression.assay}{Name of assay containing gene expression information}

\item{gene.coords}{GRanges object containing coordinates of genes in the expression assay}

\item{sep}{Separators to use to extract peak coordinates from peak assay row names}

\item{binary}{Binarize the peak counts (default TRUE)}

\item{alpha}{Alpha parameter for elastic net regularization. Controls balance between LASSO (alpha=1) and ridge (alpha=0) regression.}

\item{distance}{Distance threshold for peaks to include in regression model}

\item{verbose}{Display messages}
}
\value{
Returns a sparse matrix
}
\description{
Find peaks that are predictive of gene expression. Fits a regression model with elastic net regularization between peak
accessibility and gene expression, for all peaks within a given distance threshold (default 200 kb).
Returns a peak x gene sparse matrix, with each entry being the correlation coefficient between the
accessibility of the peak and expression of the gene.
}
