% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{LinkPeaks}
\alias{LinkPeaks}
\title{Link peaks to genes}
\usage{
LinkPeaks(
  object,
  peak.assay,
  expression.assay,
  expression.slot = "data",
  gene.coords = NULL,
  binary = FALSE,
  distance = 5e+05,
  min.cells = 10,
  expression.threshold = 0.1,
  method = "spearman",
  genes.use = NULL,
  n_sample = 100,
  pvalue_cutoff = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{peak.assay}{Name of assay containing peak information}

\item{expression.assay}{Name of assay containing gene expression information}

\item{expression.slot}{Name of slot to pull expression data from}

\item{gene.coords}{GRanges object containing coordinates of genes in the
expression assay. If NULL, extract from gene annotations stored in the assay.}

\item{binary}{Binarize the peak counts (default FALSE)}

\item{distance}{Distance threshold for peaks to include in regression model}

\item{min.cells}{Minimum number of cells positive for the peak and gene
needed to include in the results.}

\item{expression.threshold}{Minimum value for a gene to be classified as
expressed. Only used for filtering genes from regression.}

\item{method}{Which correlation coefficient to compute. Can be "spearman"
(default), "pearson", or "kendall".}

\item{genes.use}{Genes to test. If NULL, determine from expression assay.}

\item{n_sample}{Number of peaks to sample at random when computing the null
distribution.}

\item{pvalue_cutoff}{Minimum p-value required to retain a link. Links with a
p-value equal or greater than this value will be removed from the output.}

\item{verbose}{Display messages}
}
\value{
Returns a Seurat object with the \code{Links} information set, with
each recorded link being the correlation coefficient between the
accessibility of the peak and expression of the gene.
}
\description{
Find peaks that are correlated with gene expression. Fits a generalized
linear model with elastic net regularization between peak accessibility and
gene expression for all peaks within a given distance threshold from the gene
TSS.
}
\concept{links}
