% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Links}
\alias{Links}
\alias{Links<-}
\alias{Links.ChromatinAssay5}
\alias{Links.Seurat}
\alias{Links<-.ChromatinAssay5}
\alias{Links<-.Seurat}
\title{Get or set links information}
\usage{
Links(object, ...)

Links(object, ...) <- value

\method{Links}{ChromatinAssay5}(object, ...)

\method{Links}{Seurat}(object, assay = NULL, ...)

\method{Links}{ChromatinAssay5}(object, ...) <- value

\method{Links}{Seurat}(object, assay = NULL, ...) <- value
}
\arguments{
\item{object}{A Seurat or ChromatinAssay5 object}

\item{...}{Arguments passed to other methods}

\item{value}{A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges()}} object}

\item{assay}{Name of assay to use}
}
\description{
Get or set the genomic link information for a Seurat object or
\linkS4class{ChromatinAssay5}
}
\examples{
Links(atac_small[["peaks"]])
Links(atac_small)
links <- Links(atac_small)
Links(atac_small[["peaks"]]) <- links
links <- Links(atac_small)
Links(atac_small) <- links
}
\concept{assay}
\concept{links}
