% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{LoadGWASRegion}
\alias{LoadGWASRegion}
\title{Load GWAS summary statistics for a genomic region}
\usage{
LoadGWASRegion(gwas.file, region)
}
\arguments{
\item{gwas.file}{Path to GWAS summary statistics file}

\item{region}{Genomic region (GRanges object or string like "chr1:1000000-2000000")}
}
\value{
data.frame with standardized columns
}
\description{
Load GWAS summary statistics for a genomic region
}
