% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{LoadLDData}
\alias{LoadLDData}
\title{Load LD data from file or LDlink output}
\usage{
LoadLDData(ld.file)
}
\arguments{
\item{ld.file}{Path to LD file (LDlink format or pairwise)}
}
\value{
data.frame with chr, pos, r2 columns for position-based matching
}
\description{
Load LD data from file or LDlink output
}
