% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{LoadLDData}
\alias{LoadLDData}
\title{Load LD data for a lead SNP}
\usage{
LoadLDData(ld.file)
}
\arguments{
\item{ld.file}{Path to LD file}
}
\value{
data.frame with columns: chromosome, base_pair_location, r2
}
\description{
Load LD data for a lead SNP
}
\details{
Required columns: \code{chromosome}, \code{position}, \code{r2}

Column matching is case-insensitive. Values should be pairwise r-squared to a lead SNP.
}
