% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Motif-class}
\alias{Motif-class}
\title{The Motif class}
\description{
The Motif class is designed to store DNA sequence motif information,
including motif PWMs or PFMs, motif positions, and metadata.
}
\section{Slots}{

\describe{
\item{\code{data}}{A feature x motif matrix. Columns
correspond to motif IDs, rows correspond to features of interest (for
example, genomic regions). Entries in the matrix should be 1 if the feature
contains the motif and 0 otherwise.}

\item{\code{pwm}}{A named list of position weight matrices}

\item{\code{motif.names}}{A list containing the name of each motif}

\item{\code{positions}}{A \code{\link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList()}} object containing
exact positions of each motif in the genome.}

\item{\code{meta.data}}{A dataframe for storage of additional
information related to each motif. This could include the
names of proteins that bind the motif.}
}}

\concept{motifs}
