% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Motif-class}
\alias{Motif-class}
\alias{Motif}
\title{The Motif class}
\description{
The Motif class stores DNA motif information
}
\section{Slots}{

\describe{
\item{\code{data}}{A sparse, binary, motif x feature matrix. Rows correspond to motif IDs, columns correspond to genomic features (peaks or bins).
Entries in the matrix should be 1 if the genomic feature contains the motif, and 0 otherwise.}

\item{\code{pwm}}{A list of position weight matrices for each motif}

\item{\code{neighbors}}{A list of nearest neighbors graphs}

\item{\code{reductions}}{A list of \code{\link[Seurat]{DimReduc}} objects}

\item{\code{meta.data}}{A dataframe for storage of additional information related to each motif. This could include the names of proteins that bind the motif.}
}}

