% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{MotifPlot}
\alias{MotifPlot}
\title{MotifPlot}
\usage{
MotifPlot(object, motifs, assay = NULL, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{motifs}{A list of motifs to plot}

\item{assay}{Name of the assay to use}

\item{...}{Additional parameters passed to \code{\link{ggseqlogo}}}
}
\description{
Plot motifs
}
