% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{Motifs}
\alias{Motifs}
\alias{Motifs.ChromatinAssay}
\alias{Motifs.Seurat}
\title{Motifs
Retrieve a Motif object}
\usage{
Motifs(object, ...)

\method{Motifs}{ChromatinAssay}(object, ...)

\method{Motifs}{Seurat}(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}
}
\description{
Get the Motif object
}
