% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{MultiCoveragePlot}
\alias{MultiCoveragePlot}
\title{Plot multiple coverage plots}
\usage{
MultiCoveragePlot(
  object,
  region_list = NULL,
  region_names = NULL,
  extend.upstream = NULL,
  extend.downstream = NULL,
  region.highlight = NULL,
  assay = "peaks",
  assay.scale = "common",
  split.assays = FALSE,
  annotation = TRUE,
  peaks = TRUE,
  group.by = NULL,
  idents = NULL,
  split.by = NULL,
  cells = NULL,
  show.bulk = FALSE,
  ranges_list = NULL,
  ranges.title = "Ranges",
  max.downsample = 3000,
  downsample.rate = 0.1,
  scale.factor = NULL,
  ymax = NULL,
  window = 100,
  bigwig = NULL,
  bigwig.type = "coverage",
  bigwig.scale = "common",
  title_size = 6,
  subtitle_size = 5
)
}
\arguments{
\item{object}{A Seurat object}

\item{region_list}{List of genes or GRanges object to plot}

\item{region_names}{List of plot titles for each region. If NULL, default to
region_list as plot title}

\item{extend.upstream}{Number of bases to extend the region upstream. Default
is 0. Can be supplied with a vector the same length as \code{region_list}. If
supplied with a single integer, extend upstream the same number of bases for
every region in \code{region_list}}

\item{extend.downstream}{Number of bases to extend the region downstream.
Default is 0. Can be supplied with a vector the same length as \code{region_list}.
If supplied with a single integer, extend upstream the same number of bases
for every region in \code{region_list}}

\item{region.highlight}{List of regions to highlight for each plotted region,
should be the same length as \code{region_list}. Each item should be a GRanges
object containing the coordinates to highlight, or \code{NULL} if no highlight is
needed for that particular region. By default, regions will be highlighted in
grey. To change the color of the highlighting, include a metadata column in
the GRanges object named "color" containing the color to use for each region.}

\item{assay}{Name of the assay to plot. If a list of assays is provided,
data from each assay will be shown overlaid on each track. The first assay in
the list will define the assay used for gene annotations, links, and peaks
(if shown). The order of assays given defines the plotting order.}

\item{assay.scale}{Scaling to apply to data from different assays. Can be:
\itemize{
\item{common: plot all assays on a common scale (default)}
\item{separate: plot each assay on a separate scale ranging from zero to the
maximum value for that assay within the plotted region}
}}

\item{split.assays}{When plotting data from multiple assays, display each
assay as a separate track. If FALSE, data from different assays are overlaid
on a single track with transparancy applied.}

\item{annotation}{Display gene annotations. Set to TRUE or FALSE to control
whether genes models are displayed, or choose "transcript" to display all
transcript isoforms, or "gene" to display gene models only (same as setting
TRUE).}

\item{peaks}{Display peaks}

\item{group.by}{Name of one or more metadata columns to group (color) the
cells by. Default is the current cell identities}

\item{idents}{Which identities to include in the plot. Default is all
identities.}

\item{split.by}{A metadata variable to split the tracks by. For example,
grouping by "celltype" and splitting by "batch" will create separate tracks
for each combination of celltype and batch.}

\item{cells}{Which cells to plot. Default all cells}

\item{show.bulk}{Include coverage track for all cells combined (pseudo-bulk).
Note that this will plot the combined accessibility for all cells included in
the plot (rather than all cells in the object).}

\item{ranges_list}{List of additional genomic ranges to plot for each region,
should be the same length as \code{region_list}. Each item should be a GRanges
object or \code{NULL} if no additional ranges is needed for a region. Can also be
a single GRanges object to be plotted for every region.}

\item{ranges.title}{Y-axis title for ranges track. Only relevant if
\code{ranges} parameter is set.}

\item{max.downsample}{Minimum number of positions kept when downsampling.
Downsampling rate is adaptive to the window size, but this parameter will set
the minimum possible number of positions to include so that plots do not
become too sparse when the window size is small.}

\item{downsample.rate}{Fraction of positions to retain when downsampling.
Retaining more positions can give a higher-resolution plot but can make the
number of points large, resulting in larger file sizes when saving the plot
and a longer period of time needed to draw the plot.}

\item{scale.factor}{Scaling factor for track height. If NULL (default),
use the median group scaling factor determined by total number of fragments
sequences in each group.}

\item{ymax}{Maximum value for Y axis. Can be one of:
\itemize{
\item \code{NULL}: set to the highest value among all the tracks (default)
\item qXX: clip the maximum value to the XX quantile (for example, q95 will
set the maximum value to 95\% of the maximum value in the data). This can
help remove the effect of extreme values that may otherwise distort the
scale.
\item numeric: manually define a Y-axis limit
}}

\item{window}{Smoothing window size}

\item{bigwig}{List of bigWig file paths to plot data from. Files can be
remotely hosted. The name of each element in the list will determine the
y-axis label given to the track.}

\item{bigwig.type}{Type of track to use for bigWig files ("line" or
"coverage"). Should either be a single value, or a list of values giving
the type for each individual track in the provided list of bigwig files.}

\item{bigwig.scale}{Same as \code{assay.scale} parameter, except for bigWig
files when plotted with \code{bigwig.type="coverage"}}

\item{title_size}{Size of plot title (\code{region_list} or \code{region_names}).
Default is 6}

\item{subtitle_size}{Size of plot subtitle (coverage plot range).
Default is 5}
}
\value{
Returns a ggplot object
}
\description{
Plot multiple coverage plots
}
