% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PeakPlot}
\alias{PeakPlot}
\title{Plot peaks in a genomic region}
\usage{
PeakPlot(object, region, peaks = NULL)
}
\arguments{
\item{object}{A \code{\link[Seurat]{Seurat}} object}

\item{region}{A genomic region to plot}

\item{peaks}{A GRanges object containing peak coordinates. If NULL, use
coordinates stored in the Seurat object.}
}
\value{
Returns a \code{\link[ggplot2]{ggplot}} object
}
\description{
Display the genomic ranges in a \code{\link{ChromatinAssay}} object that fall
in a given genomic region
}
\examples{
PeakPlot(atac_small, region = "chr1-710000-715000")
}
\concept{visualization}
