% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{PearsonResidualVar}
\alias{PearsonResidualVar}
\alias{PearsonResidualVar.default}
\alias{PearsonResidualVar.Assay}
\alias{PearsonResidualVar.StdAssay}
\alias{PearsonResidualVar.Seurat}
\title{Compute analytic Pearson residual variance}
\usage{
PearsonResidualVar(object, ...)

\method{PearsonResidualVar}{default}(
  object,
  assay = NULL,
  nfeatures = 20000,
  min.counts = 100,
  ncell.batch = 100,
  theta = 10,
  verbose = TRUE,
  ...
)

\method{PearsonResidualVar}{Assay}(
  object,
  assay = NULL,
  nfeatures = 20000,
  theta = 10,
  min.counts = 100,
  weight.mean = 0,
  ncell.batch = 100,
  verbose = TRUE,
  ...
)

\method{PearsonResidualVar}{StdAssay}(
  object,
  assay = NULL,
  min.counts = 100,
  weight.mean = 0,
  theta = 10,
  ncell.batch = 100,
  verbose = TRUE,
  ...
)

\method{PearsonResidualVar}{Seurat}(
  object,
  assay = NULL,
  min.counts = 100,
  weight.mean = 0,
  theta = 10,
  ncell.batch = 100,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{assay}{Name of assay to use}

\item{nfeatures}{Number of top features to set as the variable features}

\item{min.counts}{Minimum number of counts for feature to be eligible for variable features}

\item{ncell.batch}{Number of cells to process in each batch. Higher number
increases speed but uses more memory.}

\item{theta}{Theta value for analytic Pearson residual calculation}

\item{verbose}{Display messages}

\item{weight.mean}{Weighting to apply to the feature mean relative to the
Pearson residual variance for ranking features. \code{weight.mean=0} will
rank features based on the Pearson residual variance only.}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Find the top features for a given assay based on analytic Pearson residual
variance. This function computes the Pearson residual variance for each
feature without constructing the entire dense matrix of Pearson residuals to
reduce the memory required.
}
\examples{
PearsonResidualVar(object = atac_small[["peaks"]]["counts"])
PearsonResidualVar(object = atac_small[["peaks"]])
PearsonResidualVar(object = atac_small[["peaks"]])
PearsonResidualVar(atac_small)
}
\references{
Lause, J., Berens, P. & Kobak, D. Analytic Pearson residuals for
normalization of single-cell RNA-seq UMI data. Genome Biol 22, 258 (2021).
\url{https://doi.org/10.1186/s13059-021-02451-7}
}
\concept{preprocessing}
