% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PeriodPlot}
\alias{PeriodPlot}
\title{Plot fragment length periodicity}
\usage{
PeriodPlot(object, assay = NULL, fragment.path = NULL,
  group.by = NULL, cells = NULL, cols = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Which assay to use. Default is the active assay.}

\item{fragment.path}{Path to an index fragment file. If NULL, will look for a path stored for the
requested assay using the \code{SetFragments} function}

\item{group.by}{Name of one or more metadata columns to group (color) the cells by. Default is the current cell identities}

\item{cells}{Which cells to plot. Default all cells}

\item{cols}{Vector of colors, where each color corresponds to an identity class. By default, use the ggplot2 colors.}
}
\description{
Plot fragment length periodicity
}
