% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PeriodPlot}
\alias{PeriodPlot}
\title{Plot fragment length periodicity}
\usage{
PeriodPlot(object, assay = NULL, fragment.path = NULL,
  region = "chr1:1-1000000", group.by = NULL, cells = NULL)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Which assay to use. Default is the active assay.}

\item{fragment.path}{Path to an index fragment file. If NULL, will look for a path stored for the
requested assay using the \code{SetFragments} function}

\item{region}{Genomic range to use. Default is fist megabase of chromosome 1.}

\item{group.by}{Name of one or more metadata columns to group (color) the cells by. Default is the current cell identities}

\item{cells}{Which cells to plot. Default all cells}
}
\value{
Returns a ggplot2 object
}
\description{
Plot fragment length periodicity
}
