% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{PlotLinks}
\alias{PlotLinks}
\title{Plot peak-gene links}
\usage{
PlotLinks(links)
}
\arguments{
\item{links}{A dataframe containing the peak and gene coordinates, and the link score}
}
\value{
Returns a ggplot2 object
}
\description{
Plot the connections between peaks and genes
}
