% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Read10XH5Content}
\alias{Read10XH5Content}
\title{Read10XH5}
\usage{
Read10XH5Content(filePath, use_names, unique_features)
}
\arguments{
\item{filePath}{A fiel path}

\item{use_names}{Use names flag}

\item{unique_features}{Unique features flag}
}
\description{
Get list triplet of SEXP
}
