% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{ReadPWM}
\alias{ReadPWM}
\title{Read PWM files into a PWMatrixList}
\usage{
ReadPWM(pwm_dir, short_names = TRUE)
}
\arguments{
\item{pwm_dir}{Path to directory containing \code{.pwm} files}

\item{short_names}{Use the first section of the motif ID (before the first
\code{.}) as the motif name. For example, \code{AHR.H14CORE.0.P.B} becomes \code{AHR}.
Default is \code{TRUE}.}
}
\value{
A \link[TFBSTools:XMatrixList-class]{TFBSTools::PWMatrixList}
}
\description{
Read position weight matrices from a directory of \code{.pwm} files
and return as a \link[TFBSTools:XMatrixList-class]{TFBSTools::PWMatrixList}. Each \code{.pwm} file
should contain a header line starting with \code{>} followed by the motif
ID, and subsequent lines containing a position x 4 nucleotide matrix
(columns: A, C, G, T).
}
\concept{motifs}
