% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{RegionAggNames}
\alias{RegionAggNames}
\alias{RegionAggNames.ChromatinAssay5}
\alias{RegionAggNames.Seurat}
\title{List stored RegionAggregation objects}
\usage{
RegionAggNames(object, ...)

\method{RegionAggNames}{ChromatinAssay5}(object)

\method{RegionAggNames}{Seurat}(object, assay = NULL, ...)
}
\arguments{
\item{object}{A Seurat or ChromatinAssay5 object}

\item{...}{Additional arguments passed to methods}

\item{assay}{Name of assay to use}
}
\value{
Character vector of stored result names

Character vector of stored result names
}
\description{
Generic function for listing stored RegionAggregation objects

Returns the names of all stored  \linkS4class{RegionAggregation} objects
}
\examples{
RegionAggr(atac_small)
}
\concept{assay}
\concept{footprinting}
