% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{RegionAggregation-class}
\alias{RegionAggregation-class}
\alias{RegionAggregation}
\title{RegionAggregation class}
\description{
The RegionAggregation class enables storage of counts centered on a group of
genomic regions, aggregated across regions for each cell. The main data
object stored is a cell-by-position matrix.
}
\section{Slots}{

\describe{
\item{\code{matrix}}{A cell-by-position matrix}

\item{\code{regions}}{A \code{\link[GenomicRanges]{granges}} object containing the
regions aggregated across.}

\item{\code{upstream}}{Integer denoting number of bases upstream of the centered
position that are stored in the matrix}

\item{\code{downstream}}{Integer denoting number of bases downstream of the centered
position that are stored in the matrix}

\item{\code{name}}{A name for the set of regions}
}}

\concept{footprinting}
