% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{RegionAggregation}
\alias{RegionAggregation}
\alias{RegionAggregation<-}
\alias{RegionAggregation.ChromatinAssay5}
\alias{RegionAggregation.Seurat}
\alias{RegionAggregation<-.ChromatinAssay5}
\alias{RegionAggregation<-.Seurat}
\title{Region Aggregation}
\usage{
RegionAggregation(object, ...)

RegionAggregation(object, ...) <- value

\method{RegionAggregation}{ChromatinAssay5}(object, ...)

\method{RegionAggregation}{Seurat}(object, assay = NULL, ...)

\method{RegionAggregation}{ChromatinAssay5}(object, ...) <- value

\method{RegionAggregation}{Seurat}(object, assay = NULL, ...) <- value
}
\arguments{
\item{object}{A Seurat or ChromatinAssay5 object}

\item{...}{Arguments passed to other methods}

\item{value}{A \linkS4class{RegionAggregation} object or list of \linkS4class{RegionAggregation} objects}

\item{assay}{Name of assay to use}
}
\value{
Returns a list of \linkS4class{RegionAggregation} objects
}
\description{
Get the region aggregation information from a \link{ChromatinAssay5} object
}
\examples{
RegionAggregation(atac_small[["peaks"]])
RegionAggregation(atac_small)
ra <- RegionAggregation(atac_small)
RegionAggregation(atac_small[["peaks"]]) <- ra
ra <- RegionAggregation(atac_small)
RegionAggregation(atac_small) <- ra
}
\concept{assay}
\concept{footprinting}
\concept{links}
