% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/heatmaps.R
\name{RegionMatrix}
\alias{RegionMatrix}
\alias{RegionMatrix.Seurat}
\alias{RegionMatrix.ChromatinAssay5}
\alias{RegionMatrix.default}
\title{Region enrichment analysis}
\usage{
RegionMatrix(object, ...)

\method{RegionMatrix}{Seurat}(
  object,
  regions,
  assay = NULL,
  group.by = NULL,
  idents = NULL,
  upstream = 3000,
  downstream = 3000,
  verbose = TRUE,
  ...
)

\method{RegionMatrix}{ChromatinAssay5}(
  object,
  regions,
  assay = NULL,
  group.by = NULL,
  idents = NULL,
  upstream = 3000,
  downstream = 3000,
  verbose = TRUE,
  ...
)

\method{RegionMatrix}{default}(
  object,
  regions,
  assay = NULL,
  group.by = NULL,
  idents = NULL,
  upstream = 3000,
  downstream = 3000,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat or \linkS4class{ChromatinAssay5} object}

\item{...}{Arguments passed to other methods}

\item{regions}{A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges()}} object containing the
set of genomic ranges to quantify}

\item{assay}{Name of assay to use. If NULL, use the default assay}

\item{group.by}{Grouping variable to use when aggregating data across cells.
If NULL, use the active cell identities}

\item{idents}{Cell identities to include. If NULL, include all identities}

\item{upstream}{Number of bases to extend regions upstream}

\item{downstream}{Number of bases to extend regions downstream}

\item{verbose}{Display messages}
}
\value{
Returns a list containing two elements:
\itemize{
\item \code{matrix}: a named list of region by position matrices, one for each group
of cells, with the name of each element corresponding to the group identity.
\item \code{parameters}: a list of function parameters "upstream", "downstream" and
"cells".
This is the input data format required by the \code{\link[=RegionHeatmap]{RegionHeatmap()}} and
\code{\link[=RegionPlot]{RegionPlot()}} functions, or you can extract the data matrix for other
plotting or downstream processing steps
}
}
\description{
Count fragments within a set of regions for different groups of
cells.
}
\concept{heatmap}
