% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{RunMotifTSNE}
\alias{RunMotifTSNE}
\title{RunMotifTSNE}
\usage{
RunMotifTSNE(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a Seurat object
}
\description{
Run tSNE on a Mofif object. This will project the motifs into two tSNE dimensions based on
a neighbor graph stored in the Motif object (commonly computed from the Jaccard similarity between motifs).
}
