% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R, R/generics.R
\name{RunMotifUMAP.Motif}
\alias{RunMotifUMAP.Motif}
\alias{RunMotifUMAP.Assay}
\alias{RunMotifUMAP.Seurat}
\alias{RunMotifUMAP}
\title{RunMotifUMAP}
\usage{
\method{RunMotifUMAP}{Motif}(object, assay = NULL, verbose = TRUE,
  graph.name = "nn", ...)

\method{RunMotifUMAP}{Assay}(object, assay = NULL, verbose = TRUE, ...)

\method{RunMotifUMAP}{Seurat}(object, assay = NULL, verbose = TRUE,
  ...)

RunMotifUMAP(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a Seurat object
}
\description{
Run UMAP on a Mofif object. This will project the motifs into two UMAP dimensions based on
a neighbor graph stored in the Motif object (commonly computed from the Jaccard similarity between motifs).
}
