% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimension_reduction.R
\name{RunMotifUMAP}
\alias{RunMotifUMAP}
\alias{RunMotifUMAP.Motif}
\alias{RunMotifUMAP.Assay}
\alias{RunMotifUMAP.Seurat}
\title{RunMotifUMAP}
\usage{
RunMotifUMAP(object, ...)

\method{RunMotifUMAP}{Motif}(object, verbose = TRUE, graph.name = "nn",
  ...)

\method{RunMotifUMAP}{Assay}(object, verbose = TRUE, ...)

\method{RunMotifUMAP}{Seurat}(object, assay = NULL, verbose = TRUE,
  ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{verbose}{Display messages}

\item{graph.name}{Name of the neighbor graph to use ('nn' or 'snn')}

\item{assay}{Name of assay to use}
}
\value{
Returns a Seurat object
}
\description{
Run UMAP on a Mofif object. This will project the motifs into two UMAP dimensions based on
a neighbor graph stored in the Motif object (commonly computed from the Jaccard similarity between motifs).
}
