% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimension_reduction.R
\name{RunSVD}
\alias{RunSVD}
\alias{RunSVD.default}
\alias{RunSVD.Assay5}
\alias{RunSVD.StdAssay}
\alias{RunSVD.Seurat}
\title{Run singular value decomposition}
\usage{
RunSVD(object, ...)

\method{RunSVD}{default}(
  object,
  assay = NULL,
  n = 50,
  scale.embeddings = !pca,
  pca = FALSE,
  reduction.key = ifelse(pca, "PCA_", "LSI_"),
  scale.max = NULL,
  verbose = TRUE,
  tol = 1e-05,
  ...
)

\method{RunSVD}{Assay5}(
  object,
  assay = NULL,
  layer = "data",
  features = NULL,
  pca = FALSE,
  n = 50,
  reduction.key = ifelse(pca, "PCA_", "LSI_"),
  scale.max = NULL,
  verbose = TRUE,
  ...
)

\method{RunSVD}{StdAssay}(
  object,
  assay = NULL,
  layer = "data",
  features = NULL,
  pca = FALSE,
  n = 50,
  reduction.key = ifelse(pca, "PCA_", "LSI_"),
  scale.max = NULL,
  verbose = TRUE,
  ...
)

\method{RunSVD}{Seurat}(
  object,
  assay = NULL,
  features = NULL,
  layer = "data",
  n = 50,
  pca = FALSE,
  reduction.key = ifelse(pca, "PCA_", "LSI_"),
  reduction.name = ifelse(pca, "pca", "lsi"),
  scale.max = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{assay}{Which assay to use. If NULL, use the default assay}

\item{n}{Number of singular values to compute}

\item{scale.embeddings}{Scale cell embeddings within each component to
mean 0 and SD 1 (default TRUE).}

\item{pca}{Run PCA. Setting this option to TRUE will perform implicit scaling
and centering of the input matrix to enable memory-efficient computation of
the principal components.}

\item{reduction.key}{Key for dimension reduction object}

\item{scale.max}{Clipping value for cell embeddings.
Default (NULL) is no clipping.}

\item{verbose}{Print messages}

\item{tol}{Tolerance (tol) parameter for \code{\link[RSpectra]{svds}}.
Larger values speed up convergence due to greater amount of allowed error.}

\item{layer}{Name of layer to use.}

\item{features}{Which features to use. If NULL, use variable features}

\item{reduction.name}{Name for stored dimension reduction object.}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Run partial singular value decomposition using \code{\link[RSpectra]{svds}}
}
\examples{
x <- matrix(data = rnorm(100), ncol = 10)
RunSVD(x)
RunSVD(atac_small[["peaks"]], features = rownames(atac_small))
RunSVD(atac_small[["peaks"]], features = rownames(atac_small))
RunSVD(atac_small, features = rownames(atac_small))
}
\concept{dimension_reduction}
