% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dimension_reduction.R
\name{RunSVD}
\alias{RunSVD}
\alias{RunSVD.default}
\alias{RunSVD.Assay}
\alias{RunSVD.Seurat}
\title{RunSVD}
\usage{
RunSVD(object, ...)

\method{RunSVD}{default}(object, assay = NULL, n = 50,
  reduction.key = "SVD_", standardize.embeddings = TRUE,
  scale.max = NULL, seed.use = NULL, verbose = TRUE, ...)

\method{RunSVD}{Assay}(object, assay = NULL, features = NULL, n = 50,
  reduction.key = "SVD_", scale.max = NULL, verbose = TRUE, ...)

\method{RunSVD}{Seurat}(object, assay = NULL, features = NULL,
  n = 50, reduction.key = "SVD_", reduction.name = "svd",
  scale.max = NULL, verbose = TRUE, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{assay}{Which assay to use. If NULL, use the default assay}

\item{n}{Number of singular values to compute}

\item{reduction.key}{Key for dimension reduction object}

\item{standardize.embeddings}{Scale and center cell embeddings}

\item{scale.max}{Clipping value for cell embeddings. Default (NULL) is no clipping.}

\item{seed.use}{Set a random seed. By default, no seed is set.}

\item{verbose}{Print messages}

\item{features}{Which features to use. If NULL, use variable features}

\item{reduction.name}{Name for stored dimension reduction object. Default 'lsi'}
}
\value{
Returns a Seurat object
}
\description{
Run singular value decomposition
}
\examples{
x <- matrix(data = rnorm(100), ncol = 10)
RunSVD(x)
}
