% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R, R/generics.R
\name{RunSVD.default}
\alias{RunSVD.default}
\alias{RunSVD.Assay}
\alias{RunSVD.Seurat}
\alias{RunSVD}
\title{RunSVD}
\usage{
\method{RunSVD}{default}(object, assay = NULL, n = 50,
  reduction.key = "LSI_", scale.max = NULL, seed.use = 42,
  verbose = TRUE, ...)

\method{RunSVD}{Assay}(object, assay = NULL, features = NULL, n = 50,
  reduction.key = "LSI_", scale.max = NULL, verbose = TRUE, ...)

\method{RunSVD}{Seurat}(object, assay = NULL, features = NULL,
  n = 50, reduction.key = "LSI_", reduction.name = "lsi",
  scale.max = NULL, verbose = TRUE, ...)

RunSVD(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Which assay to use. If NULL, use the default assay}

\item{n}{Number of singular values to compute}

\item{reduction.key}{Key for dimension reduction object}

\item{scale.max}{Clipping value for cell embeddings. Default (NULL) is no clipping.}

\item{seed.use}{Set a random seed. By default, sets the seed to 42. Setting
NULL will not set a seed.}

\item{verbose}{Print messages}

\item{...}{Arguments passed to other methods}

\item{features}{Which features to use. If NULL, use variable features}

\item{reduction.name}{Name for stored dimension reduction object. Default 'lsi'}
}
\value{
Returns a Seurat object
}
\description{
Run singular value decomposition
}
