% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeinfodb-methods.R
\name{seqinfo-methods}
\alias{seqinfo-methods}
\alias{seqinfo}
\alias{seqinfo,GRangesAssay-method}
\alias{seqlevels,GRangesAssay-method}
\alias{seqlevels}
\alias{seqnames,GRangesAssay-method}
\alias{seqnames}
\alias{seqlengths,GRangesAssay-method}
\alias{seqlengths}
\alias{genome,GRangesAssay-method}
\alias{genome}
\alias{isCircular,GRangesAssay-method}
\alias{isCircular}
\alias{seqinfo,Seurat-method}
\alias{seqlevels,Seurat-method}
\alias{seqnames,Seurat-method}
\alias{seqlengths,Seurat-method}
\alias{genome,Seurat-method}
\alias{isCircular,Seurat-method}
\title{Access and modify sequence information for GRangesAssay objects}
\usage{
\S4method{seqinfo}{GRangesAssay}(x)

\S4method{seqlevels}{GRangesAssay}(x)

\S4method{seqnames}{GRangesAssay}(x)

\S4method{seqlengths}{GRangesAssay}(x)

\S4method{genome}{GRangesAssay}(x)

\S4method{isCircular}{GRangesAssay}(x)

\S4method{seqinfo}{Seurat}(x)

\S4method{seqlevels}{Seurat}(x)

\S4method{seqnames}{Seurat}(x)

\S4method{seqlengths}{Seurat}(x)

\S4method{genome}{Seurat}(x)

\S4method{isCircular}{Seurat}(x)
}
\arguments{
\item{x}{A \code{\link{GRangesAssay}} object}
}
\description{
Methods for accessing and modifying
\code{\link[GenomeInfoDb]{Seqinfo}} object information stored in a
\code{\link{GRangesAssay}} object.
}
\section{Functions}{
\itemize{
\item \code{seqlevels(GRangesAssay)}: get method for GRangesAssay objects

\item \code{seqnames(GRangesAssay)}: get method for GRangesAssay objects

\item \code{seqlengths(GRangesAssay)}: get method for GRangesAssay objects

\item \code{genome(GRangesAssay)}: get method for GRangesAssay objects

\item \code{isCircular(GRangesAssay)}: get method for GRangesAssay objects

\item \code{seqinfo(Seurat)}: get method for Seurat objects

\item \code{seqlevels(Seurat)}: get method for Seurat objects

\item \code{seqnames(Seurat)}: get method for Seurat objects

\item \code{seqlengths(Seurat)}: get method for Seurat objects

\item \code{genome(Seurat)}: get method for Seurat objects

\item \code{isCircular(Seurat)}: get method for Seurat objects

}}
\seealso{
\itemize{
  \item{\link[GenomeInfoDb]{seqinfo} in the \pkg{GenomeInfoDb} package.}
  \item{\link{GRangesAssay-class}}
 }
}
\concept{seqinfo}
