% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeinfodb-methods.R
\name{seqinfo-methods}
\alias{seqinfo-methods}
\alias{seqinfo}
\alias{seqlevels,ChromatinAssay-method}
\alias{seqnames,ChromatinAssay-method}
\alias{seqlengths,ChromatinAssay-method}
\alias{genome,ChromatinAssay-method}
\alias{isCircular,ChromatinAssay-method}
\alias{seqinfo,Seurat-method}
\alias{seqlevels,Seurat-method}
\alias{seqnames,Seurat-method}
\alias{seqlengths,Seurat-method}
\alias{genome,Seurat-method}
\alias{isCircular,Seurat-method}
\title{Access and modify sequence information for ChromatinAssay objects}
\usage{
\S4method{seqinfo}{ChromatinAssay}(x)

\S4method{seqlevels}{ChromatinAssay}(x)

\S4method{seqnames}{ChromatinAssay}(x)

\S4method{seqlengths}{ChromatinAssay}(x)

\S4method{genome}{ChromatinAssay}(x)

\S4method{isCircular}{ChromatinAssay}(x)

\S4method{seqinfo}{Seurat}(x)

\S4method{seqlevels}{Seurat}(x)

\S4method{seqnames}{Seurat}(x)

\S4method{seqlengths}{Seurat}(x)

\S4method{genome}{Seurat}(x)

\S4method{isCircular}{Seurat}(x)
}
\description{
Generic functions for accessing and modifying
\code{\link[GenomeInfoDb]{Seqinfo}} information stored in a
\code{\link{ChromatinAssay}} object.
}
\section{Functions}{
\itemize{
\item \code{seqlevels,ChromatinAssay-method}: method for ChromatinAssay objects

\item \code{seqnames,ChromatinAssay-method}: method for ChromatinAssay objects

\item \code{seqlengths,ChromatinAssay-method}: method for ChromatinAssay objects

\item \code{genome,ChromatinAssay-method}: method for ChromatinAssay objects

\item \code{isCircular,ChromatinAssay-method}: method for ChromatinAssay objects

\item \code{seqinfo,Seurat-method}: method for Seurat objects objects

\item \code{seqlevels,Seurat-method}: method for Seurat objects

\item \code{seqnames,Seurat-method}: method for Seurat objects

\item \code{seqlengths,Seurat-method}: method for Seurat objects

\item \code{genome,Seurat-method}: method for Seurat objects

\item \code{isCircular,Seurat-method}: method for Seurat objects
}}

\seealso{
\itemize{
  \item{\link[GenomeInfoDb]{seqinfo} in the \pkg{GenomeInfoDb} package.}
  \item{\link{ChromatinAssay-class}}
 }
}
