% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeinfodb-methods.R
\name{seqinfo-methods}
\alias{seqinfo-methods}
\alias{seqinfo}
\alias{seqinfo,ChromatinAssay-method}
\alias{seqinfo<-,ChromatinAssay-method}
\alias{seqlevels,ChromatinAssay-method}
\alias{seqlevels}
\alias{seqlevels<-,ChromatinAssay-method}
\alias{seqnames,ChromatinAssay-method}
\alias{seqnames}
\alias{seqnames<-,ChromatinAssay-method}
\alias{seqlengths,ChromatinAssay-method}
\alias{seqlengths}
\alias{seqlengths<-,ChromatinAssay-method}
\alias{genome,ChromatinAssay-method}
\alias{genome}
\alias{genome<-,ChromatinAssay-method}
\alias{isCircular,ChromatinAssay-method}
\alias{isCircular}
\alias{isCircular<-,ChromatinAssay-method}
\alias{seqinfo,Seurat-method}
\alias{seqinfo<-,Seurat-method}
\alias{seqlevels,Seurat-method}
\alias{seqlevels<-,Seurat-method}
\alias{seqnames,Seurat-method}
\alias{seqnames<-,Seurat-method}
\alias{seqlengths,Seurat-method}
\alias{seqlengths<-,Seurat-method}
\alias{genome,Seurat-method}
\alias{genome<-,Seurat-method}
\alias{isCircular,Seurat-method}
\alias{isCircular<-,Seurat-method}
\title{Access and modify sequence information for ChromatinAssay objects}
\usage{
\S4method{seqinfo}{ChromatinAssay}(x)

\S4method{seqinfo}{ChromatinAssay}(x) <- value

\S4method{seqlevels}{ChromatinAssay}(x)

\S4method{seqlevels}{ChromatinAssay}(x) <- value

\S4method{seqnames}{ChromatinAssay}(x)

\S4method{seqnames}{ChromatinAssay}(x) <- value

\S4method{seqlengths}{ChromatinAssay}(x)

\S4method{seqlengths}{ChromatinAssay}(x) <- value

\S4method{genome}{ChromatinAssay}(x)

\S4method{genome}{ChromatinAssay}(x) <- value

\S4method{isCircular}{ChromatinAssay}(x)

\S4method{isCircular}{ChromatinAssay}(x) <- value

\S4method{seqinfo}{Seurat}(x)

\S4method{seqinfo}{Seurat}(x) <- value

\S4method{seqlevels}{Seurat}(x)

\S4method{seqlevels}{Seurat}(x) <- value

\S4method{seqnames}{Seurat}(x)

\S4method{seqnames}{Seurat}(x) <- value

\S4method{seqlengths}{Seurat}(x)

\S4method{seqlengths}{Seurat}(x) <- value

\S4method{genome}{Seurat}(x)

\S4method{genome}{Seurat}(x) <- value

\S4method{isCircular}{Seurat}(x)

\S4method{isCircular}{Seurat}(x) <- value
}
\arguments{
\item{x}{A \code{\link{ChromatinAssay}} object}

\item{value}{A \code{\link[GenomeInfoDb]{Seqinfo}} object or name of a UCSC
genome to store in the \code{\link{ChromatinAssay}}}
}
\description{
Methods for accessing and modifying
\code{\link[GenomeInfoDb]{Seqinfo}} object information stored in a
\code{\link{ChromatinAssay}} object.
}
\section{Functions}{
\itemize{
\item \code{seqinfo<-,ChromatinAssay-method}: set method for ChromatinAssay objects

\item \code{seqlevels,ChromatinAssay-method}: get method for ChromatinAssay objects

\item \code{seqlevels<-,ChromatinAssay-method}: set method for ChromatinAssay objects

\item \code{seqnames,ChromatinAssay-method}: get method for ChromatinAssay objects

\item \code{seqnames<-,ChromatinAssay-method}: set method for ChromatinAssay objects

\item \code{seqlengths,ChromatinAssay-method}: get method for ChromatinAssay objects

\item \code{seqlengths<-,ChromatinAssay-method}: set method for ChromatinAssay objects

\item \code{genome,ChromatinAssay-method}: get method for ChromatinAssay objects

\item \code{genome<-,ChromatinAssay-method}: set method for ChromatinAssay objects

\item \code{isCircular,ChromatinAssay-method}: get method for ChromatinAssay objects

\item \code{isCircular<-,ChromatinAssay-method}: set method for ChromatinAssay objects

\item \code{seqinfo,Seurat-method}: get method for Seurat objects

\item \code{seqinfo<-,Seurat-method}: set method for Seurat objects

\item \code{seqlevels,Seurat-method}: get method for Seurat objects

\item \code{seqlevels<-,Seurat-method}: set method for Seurat objects

\item \code{seqnames,Seurat-method}: get method for Seurat objects

\item \code{seqnames<-,Seurat-method}: set method for Seurat objects

\item \code{seqlengths,Seurat-method}: get method for Seurat objects

\item \code{seqlengths<-,Seurat-method}: set method for Seurat objects

\item \code{genome,Seurat-method}: get method for Seurat objects

\item \code{genome<-,Seurat-method}: set method for Seurat objects

\item \code{isCircular,Seurat-method}: get method for Seurat objects

\item \code{isCircular<-,Seurat-method}: set method for Seurat objects
}}

\seealso{
\itemize{
  \item{\link[GenomeInfoDb]{seqinfo} in the \pkg{GenomeInfoDb} package.}
  \item{\link{ChromatinAssay-class}}
 }
}
\concept{seqinfo}
