% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{SetMotifData}
\alias{SetMotifData}
\alias{SetMotifData.Assay}
\alias{SetMotifData.Seurat}
\title{SetMotifData}
\usage{
SetMotifData(object, ...)

\method{SetMotifData}{Assay}(object, data, ...)

\method{SetMotifData}{Seurat}(object, data, assay = NULL, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}

\item{data}{motif matrix to add. Should be matrix or sparse matrix class}

\item{assay}{Name of assay whose data should be set}
}
\value{
Returns a Seurat object
}
\description{
Set motif matrix for given assay
}
