% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signac-package.R, R/zzz.R
\docType{package}
\name{Signac-package}
\alias{Signac-package}
\title{Signac: Analysis of Single-Cell Chromatin Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A framework for the analysis and exploration of single-cell chromatin data. The 'Signac' package contains functions for quantifying single-cell chromatin data, computing per-cell quality control metrics, dimension reduction and normalization, visualization, and DNA sequence motif analysis. Reference: Stuart et al. (2021) \doi{10.1038/s41592-021-01282-5}.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A framework for the analysis and exploration of single-cell chromatin data. The 'Signac' package contains functions for quantifying single-cell chromatin data, computing per-cell quality control metrics, dimension reduction and normalization, visualization, and DNA sequence motif analysis. Reference: Stuart et al. (2021) \doi{10.1038/s41592-021-01282-5}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stuart-lab/signac}
  \item \url{https://stuartlab.org/signac}
  \item Report bugs at \url{https://github.com/stuart-lab/signac/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/stuart-lab/signac}
  \item \url{https://stuartlab.org/signac}
  \item Report bugs at \url{https://github.com/stuart-lab/signac/issues}
}

}
\author{
\strong{Maintainer}: Tim Stuart \email{stuartt@a-star.edu.sg} (\href{https://orcid.org/0000-0002-3044-0897}{ORCID})

Other contributors:
\itemize{
  \item Avi Srivastava \email{asrivastava@wistar.org} (\href{https://orcid.org/0000-0001-9798-2079}{ORCID}) [contributor]
  \item Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID}) [contributor]
  \item Rahul Satija \email{rsatija@nygenome.org} (\href{https://orcid.org/0000-0001-9448-8833}{ORCID}) [contributor]
  \item Saket Choudhary \email{saketc@iitb.ac.in} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID}) [contributor]
  \item Caleb Lareau \email{lareauc@mskcc.org} (\href{https://orcid.org/0000-0003-4179-4807}{ORCID}) [contributor]
  \item Chrysania Lim \email{Chrysania_Lim@a-star.edu.sg} [contributor]
  \item Javen Tan \email{javentyr@u.nus.edu} [contributor]
  \item Jonathan Bowles \email{jonathan_bowles@a-star.edu.sg} [contributor]
  \item Ting Wei Lim \email{Lim_Ting_Wei@a-star.edu.sg} [contributor]
  \item Jay Leung \email{leung_jia_yu@a-star.edu.sg} [contributor]
}

}
\concept{assay}
\keyword{internal}
