% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Signac-package}
\alias{Signac}
\alias{Signac-package}
\title{Signac: Analysis of Single-Cell Chromatin Data}
\description{
A framework for the analysis and exploration of single-cell chromatin data. The 'Signac' package contains functions for quantifying single-cell chromatin data, computing per-cell quality control metrics, dimension reduction and normalization, visualization, and DNA sequence motif analysis. Reference: Stuart et al. (2021) <doi:10.1038/s41592-021-01282-5>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/timoast/signac}
  \item \url{https://satijalab.org/signac}
  \item Report bugs at \url{https://github.com/timoast/signac/issues}
}

}
\author{
\strong{Maintainer}: Tim Stuart \email{tstuart@nygenome.org} (\href{https://orcid.org/0000-0002-3044-0897}{ORCID})

Authors:
\itemize{
  \item Avi Srivastava \email{asrivastava@nygenome.org} (\href{https://orcid.org/0000-0001-9798-2079}{ORCID})
}

Other contributors:
\itemize{
  \item Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID}) [contributor]
  \item Rahul Satija \email{rsatija@nygenome.org} (\href{https://orcid.org/0000-0001-9448-8833}{ORCID}) [contributor]
}

}
