% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{StringToGRanges}
\alias{StringToGRanges}
\title{StringToGRanges}
\usage{
StringToGRanges(regions, sep = "-")
}
\arguments{
\item{regions}{Vector of genomic region strings}

\item{sep}{Regular expression used to separate the genomic coordinates}
}
\value{
Returns a GRanges object
}
\description{
Convert a genomic coordinate string to a GRanges object
}
\examples{
regions <- c('chr1-1-10', 'chr2-12-3121')
StringToGRanges(regions = regions)

}
