% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{subset}
\alias{subset}
\alias{subset.Motif}
\alias{[.Motif}
\title{Return a subset of a Motif object}
\usage{
subset(object, ...)

\method{subset}{Motif}(x, features = NULL, motifs = NULL, ...)

\method{[}{Motif}(x, i, j, ...)
}
\arguments{
\item{object}{A Motif object}
}
\description{
Return a subset of a Motif object
}
\seealso{
\code{\link[base]{subset}}
}
