% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{subset}
\alias{subset}
\alias{subset.Motif}
\alias{[.Motif}
\title{Return a subset of a Motif object}
\usage{
subset(x, ...)

\method{subset}{Motif}(x, features = NULL, motifs = NULL, ...)

\method{[}{Motif}(x, i, j, ...)
}
\arguments{
\item{x}{A Motif object}

\item{...}{Arguments passed to other methods}

\item{features}{Which features to retain}

\item{motifs}{Which motifs to retain}

\item{i}{Which columns to retain}

\item{j}{Which rows to retain}
}
\description{
Return a subset of a Motif object
}
\seealso{
\code{\link[base]{subset}}
}
