% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{ValidateCells}
\alias{ValidateCells}
\title{Validate cells present in fragment file}
\usage{
ValidateCells(
  object,
  cells = NULL,
  chunksize = 5e+05,
  tolerance = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{\link{Fragment}} object}

\item{cells}{A character vector containing cell barcodes to search for.
If NULL, use the cells stored in the Fragment object.}

\item{chunksize}{Number of fragments to read at each iteration.}

\item{tolerance}{Fraction of input cells that can be unseen before returning
TRUE. For example, \code{tolerance = 0.01} will return TRUE when 99% of cells
have observed fragments in the file. This can be useful if there are cells
present that have much fewer total counts, and would require extensive
searching before a fragment from those cells are found.}

\item{verbose}{Display messages}
}
\description{
Search for a fragment from each cell that should exist in the fragment file.
Will iterate through chunks of the fragment file until at least one fragment
from each cell barcode requested is found.
}
