% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{ValidateFragments}
\alias{ValidateFragments}
\title{Validate Fragment object}
\usage{
ValidateFragments(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{A \code{\link{Fragment}} object}

\item{verbose}{Display messages}

\item{...}{Additional parameters passed to \code{\link{ValidateCells}}}
}
\description{
Verify that the cells listed in the object exist in the fragment file
and that the fragment file or index have not changed since creating the
fragment object.
}
\concept{fragments}
