% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{ValidateHash}
\alias{ValidateHash}
\title{Validate hashes for Fragment object}
\usage{
ValidateHash(object, verbose = TRUE)
}
\arguments{
\item{object}{A Fragment object}

\item{verbose}{Display messages}
}
\description{
Validate hashes for Fragment object
}
